Module rustc::dep_graph::label_strs

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Contains variant => str representations for constructing DepNode groups for tests.

Constants

AdtDefOfItem 	 [
Experimental
]

AdtDestructor 	 [
Experimental
]

AllCrateNums 	 [
Experimental
]

AllLocalTraitImpls 	 [
Experimental
]

AllTraitImplementations 	 [
Experimental
]

AssociatedItemDefIds 	 [
Experimental
]

AssociatedItems 	 [
Experimental
]

BorrowCheck 	 [
Experimental
]

BorrowCheckKrate 	 [
Experimental
]

ClosureKind 	 [
Experimental
]

CodegenUnit 	 [
Experimental
]

CoerceUnsizedInfo 	 [
Experimental
]

Coherence 	 [
Experimental
]

CoherenceCheckTrait 	 [
Experimental
]

CoherenceInherentImplOverlapCheck 	 [
Experimental
]

CollectAndPartitionTranslationItems 	 [
Experimental
]

CompileCodegenUnit 	 [
Experimental
]

ConstEval 	 [
Experimental
]

ConstIsRvaluePromotableToStatic 	 [
Experimental
]

ContainsExternIndicator 	 [
Experimental
]

CrateDisambiguator 	 [
Experimental
]

CrateHash 	 [
Experimental
]

CrateMetadata 	 [
Experimental
]

CrateName 	 [
Experimental
]

CrateVariances 	 [
Experimental
]

DefSpan 	 [
Experimental
]

DefinedLangItems 	 [
Experimental
]

DepKind 	 [
Experimental
]

DeriveRegistrarFn 	 [
Experimental
]

DescribeDef 	 [
Experimental
]

DtorckConstraint 	 [
Experimental
]

DylibDepFormats 	 [
Experimental
]

EraseRegionsTy 	 [
Experimental
]

ExportName 	 [
Experimental
]

ExportedSymbolIds 	 [
Experimental
]

ExportedSymbols 	 [
Experimental
]

ExternConstBody 	 [
Experimental
]

ExternCrate 	 [
Experimental
]

ExternModStmtCnum 	 [
Experimental
]

FnArgNames 	 [
Experimental
]

FnSignature 	 [
Experimental
]

Freevars 	 [
Experimental
]

FulfillObligation 	 [
Experimental
]

GenSignature 	 [
Experimental
]

GenericsOfItem 	 [
Experimental
]

GetLangItems 	 [
Experimental
]

GetPanicStrategy 	 [
Experimental
]

HasCloneClosures 	 [
Experimental
]

HasCopyClosures 	 [
Experimental
]

HasGlobalAllocator 	 [
Experimental
]

HasTypeckTables 	 [
Experimental
]

Hir 	 [
Experimental
]

HirBody 	 [
Experimental
]

ImplDefaultness 	 [
Experimental
]

ImplParent 	 [
Experimental
]

ImplPolarity 	 [
Experimental
]

ImplTraitRef 	 [
Experimental
]

ImplementationsOfTrait 	 [
Experimental
]

InScopeTraits 	 [
Experimental
]

InferredOutlivesOf 	 [
Experimental
]

InherentImpls 	 [
Experimental
]

InstanceSymbolName 	 [
Experimental
]

IsCompilerBuiltins 	 [
Experimental
]

IsConstFn 	 [
Experimental
]

IsCopy 	 [
Experimental
]

IsDefaultImpl 	 [
Experimental
]

IsDllimportForeignItem 	 [
Experimental
]

IsExportedSymbol 	 [
Experimental
]

IsForeignItem 	 [
Experimental
]

IsFreeze 	 [
Experimental
]

IsLateBound 	 [
Experimental
]

IsMirAvailable 	 [
Experimental
]

IsNoBuiltins 	 [
Experimental
]

IsPanicRuntime 	 [
Experimental
]

IsProfilerRuntime 	 [
Experimental
]

IsSanitizerRuntime 	 [
Experimental
]

IsSized 	 [
Experimental
]

IsStaticallyIncludedForeignItem 	 [
Experimental
]

IsTranslatedFunction 	 [
Experimental
]

ItemAttrs 	 [
Experimental
]

ItemBodyNestedBodies 	 [
Experimental
]

ItemChildren 	 [
Experimental
]

ItemVarianceConstraints 	 [
Experimental
]

ItemVariances 	 [
Experimental
]

Krate 	 [
Experimental
]

Layout 	 [
Experimental
]

LinkArgs 	 [
Experimental
]

LintLevels 	 [
Experimental
]

LookupDeprecationEntry 	 [
Experimental
]

LookupStability 	 [
Experimental
]

MaybeUnusedExternCrates 	 [
Experimental
]

MaybeUnusedTraitImport 	 [
Experimental
]

MirBorrowCheck 	 [
Experimental
]

MirConst 	 [
Experimental
]

MirConstQualif 	 [
Experimental
]

MirKeys 	 [
Experimental
]

MirOptimized 	 [
Experimental
]

MirShim 	 [
Experimental
]

MirValidated 	 [
Experimental
]

MissingExternCrateItem 	 [
Experimental
]

MissingLangItems 	 [
Experimental
]

ModuleExports 	 [
Experimental
]

NamedRegion 	 [
Experimental
]

NativeLibraries 	 [
Experimental
]

NativeLibraryKind 	 [
Experimental
]

NeedsDrop 	 [
Experimental
]

Null 	 [
Experimental
]

ObjectLifetimeDefaults 	 [
Experimental
]

ObjectSafety 	 [
Experimental
]

OriginalCrateName 	 [
Experimental
]

OutputFilenames 	 [
Experimental
]

ParamEnv 	 [
Experimental
]

PluginRegistrarFn 	 [
Experimental
]

PostorderCnums 	 [
Experimental
]

PredicatesOfItem 	 [
Experimental
]

PrivacyAccessLevels 	 [
Experimental
]

Reachability 	 [
Experimental
]

RegionScopeTree 	 [
Experimental
]

RvaluePromotableMap 	 [
Experimental
]

SizedConstraint 	 [
Experimental
]

SpecializationGraph 	 [
Experimental
]

Specializes 	 [
Experimental
]

StabilityIndex 	 [
Experimental
]

SuperPredicatesOfItem 	 [
Experimental
]

SymbolName 	 [
Experimental
]

TraitDefOfItem 	 [
Experimental
]

TraitImpls 	 [
Experimental
]

TraitOfItem 	 [
Experimental
]

TraitSelect 	 [
Experimental
]

TypeOfItem 	 [
Experimental
]

TypeParamPredicates 	 [
Experimental
]

TypeckBodiesKrate 	 [
Experimental
]

TypeckTables 	 [
Experimental
]

UnsafetyViolations 	 [
Experimental
]

UsedCrateSource 	 [
Experimental
]

Visibility 	 [
Experimental
]

VisibleParentMap 	 [
Experimental
]

VtableMethods 	 [
Experimental
]

WorkProduct 	 [
Experimental
]