Struct rustc::dep_graph::DepTrackingMap

pub struct DepTrackingMap<M: DepTrackingMapConfig> {
    // some fields omitted
}

Unstable (rustc_private)

A DepTrackingMap offers a subset of the Map API and ensures that we make calls to read and write as appropriate. We key the maps with a unique type for brevity.

Methods

impl<M: DepTrackingMapConfig> DepTrackingMap<M>

fn new(graph: DepGraph) -> DepTrackingMap<M>

Unstable (rustc_private)

fn get(&self, k: &M::Key) -> Option<&M::Value>

Unstable (rustc_private)

fn insert(&mut self, k: M::Key, v: M::Value) -> Option<M::Value>

Unstable (rustc_private)

fn contains_key(&self, k: &M::Key) -> bool

Unstable (rustc_private)