Struct rustc::dep_graph::SerializedDepGraph

pub struct SerializedDepGraph {
    pub nodes: IndexVec<SerializedDepNodeIndex, (DepNode, Fingerprint)>,
    pub edge_list_indices: IndexVec<SerializedDepNodeIndex, (u32, u32)>,
    pub edge_list_data: Vec<SerializedDepNodeIndex>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Data for use when recompiling the current crate.

Fields
nodes: IndexVec<SerializedDepNodeIndex, (DepNode, Fingerprint)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The set of all DepNodes in the graph

edge_list_indices: IndexVec<SerializedDepNodeIndex, (u32, u32)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

For each DepNode, stores the list of edges originating from that DepNode. Encoded as a [start, end) pair indexing into edge_list_data, which holds the actual DepNodeIndices of the target nodes.

edge_list_data: Vec<SerializedDepNodeIndex>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A flattened list of all edge targets in the graph. Edge sources are implicit in edge_list_indices.

Methods

impl SerializedDepGraph

pub fn new() -> SerializedDepGraph

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn edge_targets_from(
&self,
source: SerializedDepNodeIndex
) -> &[SerializedDepNodeIndex]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?