Struct rustc::hir::def::PathResolution

pub struct PathResolution { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The result of resolving a path before lowering to HIR. base_def is definition of resolved part of the path, unresolved_segments is the number of unresolved segments. module::Type::AssocX::AssocY::MethodOrAssocType ~~~~~~~~~~~ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ base_def unresolved_segments = 3

<T as Trait>::AssocX::AssocY::MethodOrAssocType
      ^~~~~~~~~~~~~~  ^~~~~~~~~~~~~~~~~~~~~~~~~
      base_def        unresolved_segments = 2

Methods

impl PathResolution

pub fn new(def: Def) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn with_unresolved_segments(def: Def, unresolved_segments: usize) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn base_def(&self) -> Def

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn unresolved_segments(&self) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn kind_name(&self) -> &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?