Module rustc::hir::def_id

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

CrateNum 	 [
Experimental
]

DefId 	 [
Experimental
]
A DefId identifies a particular definition, by combining a crate index and a def index.

DefIndex 	 [
Experimental
]
A DefIndex is an index into the hir-map for a crate, identifying a particular definition. It should really be considered an interned shorthand for a particular DefPath.

Enums

DefIndexAddressSpace 	 [
Experimental
]

Constants

BUILTIN_MACROS_CRATE 	 [
Experimental
]
Virtual crate for builtin macros

CRATE_DEF_INDEX 	 [
Experimental
]
The crate root is always assigned index 0 by the AST Map code, thanks to NodeCollector::new.

INVALID_CRATE 	 [
Experimental
]
A CrateNum value that indicates that something is wrong.

LOCAL_CRATE 	 [
Experimental
]
Item definitions in the currently-compiled crate would have the CrateNum LOCAL_CRATE in their DefId.