Struct rustc::hir::def_id::DefIndex

pub struct DefIndex(_);

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A DefIndex is an index into the hir-map for a crate, identifying a particular definition. It should really be considered an interned shorthand for a particular DefPath.

At the moment we are allocating the numerical values of DefIndexes into two ranges: the "low" range (starting at zero) and the "high" range (starting at DEF_INDEX_HI_START). This allows us to allocate the DefIndexes of all item-likes (Items, TraitItems, and ImplItems) into one of these ranges and consequently use a simple array for lookup tables keyed by DefIndex and known to be densely populated. This is especially important for the HIR map.

Since the DefIndex is mostly treated as an opaque ID, you probably don't have to care about these ranges.

Methods

impl DefIndex

pub fn new(x: usize) -> DefIndex

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn from_u32(x: u32) -> DefIndex

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn as_usize(&self) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn as_u32(&self) -> u32

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn address_space(&self) -> DefIndexAddressSpace

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn as_array_index(&self) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Converts this DefIndex into a zero-based array index. This index is the offset within the given "range" of the DefIndex, that is, if the DefIndex is part of the "high" range, the resulting index will be (DefIndex - DEF_INDEX_HI_START).

pub fn from_array_index(
i: usize,
address_space: DefIndexAddressSpace
) -> DefIndex

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?