Enum rustc::hir::Item_

pub enum Item_ {
    ItemExternCrate(Option<Name>),
    ItemUse(P<Path>, UseKind),
    ItemStatic(P<Ty>, Mutability, BodyId),
    ItemConst(P<Ty>, BodyId),
    ItemFn(P<FnDecl>, Unsafety, Constness, Abi, Generics, BodyId),
    ItemMod(Mod),
    ItemForeignMod(ForeignMod),
    ItemGlobalAsm(P<GlobalAsm>),
    ItemTy(P<Ty>, Generics),
    ItemEnum(EnumDef, Generics),
    ItemStruct(VariantData, Generics),
    ItemUnion(VariantData, Generics),
    ItemTrait(Unsafety, Generics, TyParamBounds, HirVec<TraitItemRef>),
    ItemDefaultImpl(Unsafety, TraitRef),
    ItemImpl(Unsafety, ImplPolarity, Defaultness, Generics, Option<TraitRef>, P<Ty>, HirVec<ImplItemRef>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
ItemExternCrate(Option<Name>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An extern crate item, with optional original crate name,

e.g. extern crate foo or extern crate foo_bar as foo

ItemUse(P<Path>, UseKind)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

use foo::bar::*; or use foo::bar::baz as quux;

or just

use foo::bar::baz; (with as baz implicitly on the right)

ItemStatic(P<Ty>, Mutability, BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A static item

ItemConst(P<Ty>, BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A const item

ItemFn(P<FnDecl>, Unsafety, Constness, Abi, Generics, BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A function declaration

ItemMod(Mod)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A module

ItemForeignMod(ForeignMod)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An external module

ItemGlobalAsm(P<GlobalAsm>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Module-level inline assembly (from global_asm!)

ItemTy(P<Ty>, Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A type alias, e.g. type Foo = Bar<u8>

ItemEnum(EnumDef, Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An enum definition, e.g. enum Foo<A, B> {C<A>, D<B>}

ItemStruct(VariantData, Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A struct definition, e.g. struct Foo<A> {x: A}

ItemUnion(VariantData, Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A union definition, e.g. union Foo<A, B> {x: A, y: B}

ItemTrait(Unsafety, Generics, TyParamBounds, HirVec<TraitItemRef>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Represents a Trait Declaration

ItemDefaultImpl(Unsafety, TraitRef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl Trait for .. {}

ItemImpl(Unsafety, ImplPolarity, Defaultness, Generics, Option<TraitRef>, P<Ty>, HirVec<ImplItemRef>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An implementation, eg impl<A> Trait for Foo { .. }

Methods

impl Item_

pub fn descriptive_variant(&self) -> &str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn adt_kind(&self) -> Option<AdtKind>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?