Module rustc::hir::intravisit

Unstable (rustc_private)

HIR walker. Each overridden visit method has full control over what happens with its node, it can do its own traversal of the node's children, call intravisit::walk_* to apply the default traversal algorithm, or prevent deeper traversal by doing nothing.

When visiting the HIR, the contents of nested items are NOT visited by default. This is different from the AST visitor, which does a deep walk. Hence this module is called intravisit; see the method visit_nested_item for more details.

Note: it is an important invariant that the default visitor walks the body of a function in "execution order" (more concretely, reverse post-order with respect to the CFG implied by the AST), meaning that if AST node A may execute before AST node B, then A is visited first. The borrow checker in particular relies on this property.

Structs

IdRange 	 [Unstable]

IdRangeComputingVisitor 	 [Unstable]

IdVisitor 	 [Unstable]

Enums

FnKind 	 [Unstable]

Traits

IdVisitingOperation 	 [Unstable]

Visitor 	 [Unstable]
Each method of the Visitor trait is a hook to be potentially overridden. Each method's default implementation recursively visits the substructure of the input via the corresponding walk method; e.g. the visit_mod method by default calls intravisit::walk_mod.

Functions

compute_id_range_for_fn_body 	 [Unstable]
Computes the id range for a single fn body, ignoring nested items.

walk_arm 	 [Unstable]

walk_assoc_type_binding 	 [Unstable]

walk_block 	 [Unstable]

walk_crate 	 [Unstable]
Walks the contents of a crate. See also Crate::visit_all_items.

walk_decl 	 [Unstable]

walk_enum_def 	 [Unstable]

walk_expr 	 [Unstable]

walk_fn 	 [Unstable]

walk_fn_decl 	 [Unstable]

walk_fn_decl_nopat 	 [Unstable]

walk_fn_kind 	 [Unstable]

walk_fn_ret_ty 	 [Unstable]

walk_foreign_item 	 [Unstable]

walk_generics 	 [Unstable]

walk_impl_item 	 [Unstable]

walk_item 	 [Unstable]

walk_lifetime 	 [Unstable]

walk_lifetime_def 	 [Unstable]

walk_local 	 [Unstable]

walk_macro_def 	 [Unstable]

walk_mod 	 [Unstable]

walk_opt_name 	 [Unstable]

walk_pat 	 [Unstable]

walk_path 	 [Unstable]

walk_path_list_item 	 [Unstable]

walk_path_parameters 	 [Unstable]

walk_path_segment 	 [Unstable]

walk_poly_trait_ref 	 [Unstable]

walk_stmt 	 [Unstable]

walk_struct_def 	 [Unstable]

walk_struct_field 	 [Unstable]

walk_trait_item 	 [Unstable]

walk_trait_ref 	 [Unstable]

walk_ty 	 [Unstable]

walk_ty_param_bound 	 [Unstable]

walk_variant 	 [Unstable]

walk_vis 	 [Unstable]