Function rustc::hir::intravisit::walk_fn

pub fn walk_fn<'v, V: Visitor<'v>>(
    visitor: &mut V, 
    function_kind: FnKind<'v>, 
    function_declaration: &'v FnDecl, 
    body_id: BodyId, 
    _span: Span, 
    id: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?