Struct rustc::hir::intravisit::IdRangeComputingVisitor

pub struct IdRangeComputingVisitor<'a, 'hir: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'hir> IdRangeComputingVisitor<'a, 'hir>

pub fn new(map: &'a Map<'hir>) -> IdRangeComputingVisitor<'a, 'hir>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn result(&self) -> IdRange

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?