Trait rustc::hir::intravisit::Visitor

pub trait Visitor<'v>: Sized {
    fn nested_visit_map<'this>(&'this mut self) -> NestedVisitorMap<'this, 'v>;

    fn visit_nested_item(&mut self, id: ItemId) { ... }

    fn visit_nested_trait_item(&mut self, id: TraitItemId) { ... }

    fn visit_nested_impl_item(&mut self, id: ImplItemId) { ... }

    fn visit_nested_body(&mut self, id: BodyId) { ... }

    fn visit_item(&mut self, i: &'v Item) { ... }

    fn visit_body(&mut self, b: &'v Body) { ... }

    fn as_deep_visitor<'s>(&'s mut self) -> DeepVisitor<'s, Self> { ... }

    fn visit_id(&mut self, _node_id: NodeId) { ... }

    fn visit_def_mention(&mut self, _def: Def) { ... }

    fn visit_name(&mut self, _span: Span, _name: Name) { ... }

    fn visit_mod(&mut self, m: &'v Mod, _s: Span, n: NodeId) { ... }

    fn visit_foreign_item(&mut self, i: &'v ForeignItem) { ... }

    fn visit_local(&mut self, l: &'v Local) { ... }

    fn visit_block(&mut self, b: &'v Block) { ... }

    fn visit_stmt(&mut self, s: &'v Stmt) { ... }

    fn visit_arm(&mut self, a: &'v Arm) { ... }

    fn visit_pat(&mut self, p: &'v Pat) { ... }

    fn visit_decl(&mut self, d: &'v Decl) { ... }

    fn visit_expr(&mut self, ex: &'v Expr) { ... }

    fn visit_ty(&mut self, t: &'v Ty) { ... }

    fn visit_generics(&mut self, g: &'v Generics) { ... }

    fn visit_where_predicate(&mut self, predicate: &'v WherePredicate) { ... }

    fn visit_fn_decl(&mut self, fd: &'v FnDecl) { ... }

    fn visit_fn(
        &mut self, 
        fk: FnKind<'v>, 
        fd: &'v FnDecl, 
        b: BodyId, 
        s: Span, 
        id: NodeId
    ) { ... }

    fn visit_trait_item(&mut self, ti: &'v TraitItem) { ... }

    fn visit_trait_item_ref(&mut self, ii: &'v TraitItemRef) { ... }

    fn visit_impl_item(&mut self, ii: &'v ImplItem) { ... }

    fn visit_impl_item_ref(&mut self, ii: &'v ImplItemRef) { ... }

    fn visit_trait_ref(&mut self, t: &'v TraitRef) { ... }

    fn visit_ty_param_bound(&mut self, bounds: &'v TyParamBound) { ... }

    fn visit_poly_trait_ref(
        &mut self, 
        t: &'v PolyTraitRef, 
        m: TraitBoundModifier
    ) { ... }

    fn visit_variant_data(
        &mut self, 
        s: &'v VariantData, 
        _: Name, 
        _: &'v Generics, 
        _parent_id: NodeId, 
        _: Span
    ) { ... }

    fn visit_struct_field(&mut self, s: &'v StructField) { ... }

    fn visit_enum_def(
        &mut self, 
        enum_definition: &'v EnumDef, 
        generics: &'v Generics, 
        item_id: NodeId, 
        _: Span
    ) { ... }

    fn visit_variant(
        &mut self, 
        v: &'v Variant, 
        g: &'v Generics, 
        item_id: NodeId
    ) { ... }

    fn visit_lifetime(&mut self, lifetime: &'v Lifetime) { ... }

    fn visit_lifetime_def(&mut self, lifetime: &'v LifetimeDef) { ... }

    fn visit_qpath(&mut self, qpath: &'v QPath, id: NodeId, span: Span) { ... }

    fn visit_path(&mut self, path: &'v Path, _id: NodeId) { ... }

    fn visit_path_segment(
        &mut self, 
        path_span: Span, 
        path_segment: &'v PathSegment
    ) { ... }

    fn visit_path_parameters(
        &mut self, 
        path_span: Span, 
        path_parameters: &'v PathParameters
    ) { ... }

    fn visit_assoc_type_binding(&mut self, type_binding: &'v TypeBinding) { ... }

    fn visit_attribute(&mut self, _attr: &'v Attribute) { ... }

    fn visit_macro_def(&mut self, macro_def: &'v MacroDef) { ... }

    fn visit_vis(&mut self, vis: &'v Visibility) { ... }

    fn visit_associated_item_kind(&mut self, kind: &'v AssociatedItemKind) { ... }

    fn visit_defaultness(&mut self, defaultness: &'v Defaultness) { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Each method of the Visitor trait is a hook to be potentially overridden. Each method's default implementation recursively visits the substructure of the input via the corresponding walk method; e.g. the visit_mod method by default calls intravisit::walk_mod.

Note that this visitor does NOT visit nested items by default (this is why the module is called intravisit, to distinguish it from the AST's visit module, which acts differently). If you simply want to visit all items in the crate in some order, you should call Crate::visit_all_items. Otherwise, see the comment on visit_nested_item for details on how to visit nested items.

If you want to ensure that your code handles every variant explicitly, you need to override each method. (And you also need to monitor future changes to Visitor in case a new method with a new default implementation gets introduced.)

Required Methods

fn nested_visit_map<'this>(&'this mut self) -> NestedVisitorMap<'this, 'v>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The default versions of the visit_nested_XXX routines invoke this method to get a map to use. By selecting an enum variant, you control which kinds of nested HIR are visited; see NestedVisitorMap for details. By "nested HIR", we are referring to bits of HIR that are not directly embedded within one another but rather indirectly, through a table in the crate. This is done to control dependencies during incremental compilation: the non-inline bits of HIR can be tracked and hashed separately.

If for some reason you want the nested behavior, but don't have a Map at your disposal: then you should override the visit_nested_XXX methods, and override this method to panic!(). This way, if a new visit_nested_XXX variant is added in the future, we will see the panic in your code and fix it appropriately.

Provided Methods

fn visit_nested_item(&mut self, id: ItemId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Invoked when a nested item is encountered. By default does nothing unless you override nested_visit_map to return Some(_), in which case it will walk the item. You probably don't want to override this method -- instead, override nested_visit_map or use the "shallow" or "deep" visit patterns described on itemlikevisit::ItemLikeVisitor. The only reason to override this method is if you want a nested pattern but cannot supply a Map; see nested_visit_map for advice.

fn visit_nested_trait_item(&mut self, id: TraitItemId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Like visit_nested_item(), but for trait items. See visit_nested_item() for advice on when to override this method.

fn visit_nested_impl_item(&mut self, id: ImplItemId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Like visit_nested_item(), but for impl items. See visit_nested_item() for advice on when to override this method.

fn visit_nested_body(&mut self, id: BodyId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Invoked to visit the body of a function, method or closure. Like visit_nested_item, does nothing by default unless you override nested_visit_map to return Some(_), in which case it will walk the body.

fn visit_item(&mut self, i: &'v Item)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Visit the top-level item and (optionally) nested items / impl items. See visit_nested_item for details.

fn visit_body(&mut self, b: &'v Body)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn as_deep_visitor<'s>(&'s mut self) -> DeepVisitor<'s, Self>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

When invoking visit_all_item_likes(), you need to supply an item-like visitor. This method converts a "intra-visit" visitor into an item-like visitor that walks the entire tree. If you use this, you probably don't want to process the contents of nested item-like things, since the outer loop will visit them as well.

fn visit_id(&mut self, _node_id: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_def_mention(&mut self, _def: Def)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_name(&mut self, _span: Span, _name: Name)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_mod(&mut self, m: &'v Mod, _s: Span, n: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_foreign_item(&mut self, i: &'v ForeignItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_local(&mut self, l: &'v Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_block(&mut self, b: &'v Block)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_stmt(&mut self, s: &'v Stmt)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_arm(&mut self, a: &'v Arm)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_pat(&mut self, p: &'v Pat)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_decl(&mut self, d: &'v Decl)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_expr(&mut self, ex: &'v Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty(&mut self, t: &'v Ty)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_generics(&mut self, g: &'v Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_where_predicate(&mut self, predicate: &'v WherePredicate)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_fn_decl(&mut self, fd: &'v FnDecl)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_fn(
&mut self,
fk: FnKind<'v>,
fd: &'v FnDecl,
b: BodyId,
s: Span,
id: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_trait_item(&mut self, ti: &'v TraitItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_trait_item_ref(&mut self, ii: &'v TraitItemRef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_impl_item(&mut self, ii: &'v ImplItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_impl_item_ref(&mut self, ii: &'v ImplItemRef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_trait_ref(&mut self, t: &'v TraitRef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty_param_bound(&mut self, bounds: &'v TyParamBound)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_poly_trait_ref(&mut self, t: &'v PolyTraitRef, m: TraitBoundModifier)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_variant_data(
&mut self,
s: &'v VariantData,
_: Name,
_: &'v Generics,
_parent_id: NodeId,
_: Span
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_struct_field(&mut self, s: &'v StructField)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_enum_def(
&mut self,
enum_definition: &'v EnumDef,
generics: &'v Generics,
item_id: NodeId,
_: Span
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_variant(&mut self, v: &'v Variant, g: &'v Generics, item_id: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_lifetime(&mut self, lifetime: &'v Lifetime)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_lifetime_def(&mut self, lifetime: &'v LifetimeDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_qpath(&mut self, qpath: &'v QPath, id: NodeId, span: Span)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path(&mut self, path: &'v Path, _id: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path_segment(&mut self, path_span: Span, path_segment: &'v PathSegment)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_path_parameters(
&mut self,
path_span: Span,
path_parameters: &'v PathParameters
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_assoc_type_binding(&mut self, type_binding: &'v TypeBinding)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_attribute(&mut self, _attr: &'v Attribute)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_macro_def(&mut self, macro_def: &'v MacroDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_vis(&mut self, vis: &'v Visibility)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_associated_item_kind(&mut self, kind: &'v AssociatedItemKind)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_defaultness(&mut self, defaultness: &'v Defaultness)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?