Struct rustc::hir::itemlikevisit::DeepVisitor

pub struct DeepVisitor<'v, V: 'v> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'v, 'hir, V> DeepVisitor<'v, V> where
V: Visitor<'hir> + 'v,

pub fn new(base: &'v mut V) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?