Module rustc::hir::map

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use self::Node::*;

pub use self::definitions::Definitions;

pub use self::definitions::DefKey;

pub use self::definitions::DefPath;

pub use self::definitions::DefPathData;

pub use self::definitions::DisambiguatedDefPathData;

pub use self::definitions::DefPathHash;

Modules

blocks 	 [
Experimental
]
This module provides a simplified abstraction for working with code blocks identified by their integer node-id. In particular, it captures a common set of attributes that all "function-like things" (represented by FnLike instances) share. For example, all FnLike instances have a type signature (be it explicit or inferred). And all FnLike instances have a body, i.e. the code that is run when the function-like thing it represents is invoked.

definitions 	 [
Experimental
]
For each definition, we track the following data. A definition here is defined somewhat circularly as "something with a def-id", but it generally corresponds to things like structs, enums, etc. There are also some rather random cases (like const initializer expressions) that are mostly just leftovers.

Structs

DefCollector 	 [
Experimental
]
Creates def ids for nodes in the AST.

Forest 	 [
Experimental
]
Stores a crate and any number of inlined items from other crates.

MacroInvocationData 	 [
Experimental
]

Map 	 [
Experimental
]
Represents a mapping from Node IDs to AST elements and their parent Node IDs

NodesMatchingSuffix 	 [
Experimental
]

Enums

Node 	 [
Experimental
]

Constants

ITEM_LIKE_SPACE 	 [
Experimental
]

REGULAR_SPACE 	 [
Experimental
]

Functions

map_crate 	 [
Experimental
]