Struct rustc::hir::map::Map

pub struct Map<'ast> {
    pub forest: &'ast Forest,
    pub dep_graph: DepGraph,
    // some fields omitted
}

Unstable (rustc_private)

Represents a mapping from Node IDs to AST elements and their parent Node IDs

Fields
forest: &'ast Forest
Unstable (rustc_private)

The backing storage for all the AST nodes.

dep_graph: DepGraph
Unstable (rustc_private)

Same as the dep_graph in forest, just available with one fewer deref. This is a gratuitious micro-optimization.

Methods

impl<'ast> Map<'ast>

fn num_local_def_ids(&self) -> usize

Unstable (rustc_private)

fn def_key(&self, def_id: DefId) -> DefKey

Unstable (rustc_private)

fn def_path_from_id(&self, id: NodeId) -> Option<DefPath>

Unstable (rustc_private)

fn def_path(&self, def_id: DefId) -> DefPath

Unstable (rustc_private)

fn def_index_for_def_key(&self, def_key: DefKey) -> Option<DefIndex>

Unstable (rustc_private)

fn local_def_id(&self, node: NodeId) -> DefId

Unstable (rustc_private)

fn opt_local_def_id(&self, node: NodeId) -> Option<DefId>

Unstable (rustc_private)

fn as_local_node_id(&self, def_id: DefId) -> Option<NodeId>

Unstable (rustc_private)

fn krate(&self) -> &'ast Crate

Unstable (rustc_private)

fn krate_attrs(&self) -> &'ast [Attribute]

Unstable (rustc_private)

Get the attributes on the krate. This is preferable to invoking krate.attrs because it registers a tighter dep-graph access.

fn get(&self, id: NodeId) -> Node<'ast>

Unstable (rustc_private)

Retrieve the Node corresponding to id, panicking if it cannot be found.

fn get_if_local(&self, id: DefId) -> Option<Node<'ast>>

Unstable (rustc_private)

fn find(&self, id: NodeId) -> Option<Node<'ast>>

Unstable (rustc_private)

Retrieve the Node corresponding to id, returning None if cannot be found.

fn get_parent_node(&self, id: NodeId) -> NodeId

Unstable (rustc_private)

Similar to get_parent, returns the parent node id or id if there is no parent. This function returns the immediate parent in the AST, whereas get_parent returns the enclosing item. Note that this might not be the actual parent node in the AST - some kinds of nodes are not in the map and these will never appear as the parent_node. So you can always walk the parent_nodes from a node to the root of the ast (unless you get the same id back here that can happen if the id is not in the map itself or is just weird).

fn is_argument(&self, id: NodeId) -> bool

Unstable (rustc_private)

Check if the node is an argument. An argument is a local variable whose immediate parent is an item or a closure.

fn get_parent(&self, id: NodeId) -> NodeId

Unstable (rustc_private)

Retrieve the NodeId for id's parent item, or id itself if no parent item is in this map. The "parent item" is the closest parent node in the AST which is recorded by the map and is an item, either an item in a module, trait, or impl.

fn get_module_parent(&self, id: NodeId) -> NodeId

Unstable (rustc_private)

Returns the NodeId of id's nearest module parent, or id itself if no module parent is in this map.

fn get_enclosing_scope(&self, id: NodeId) -> Option<NodeId>

Unstable (rustc_private)

Returns the nearest enclosing scope. A scope is an item or block. FIXME it is not clear to me that all items qualify as scopes - statics and associated types probably shouldn't, for example. Behaviour in this regard should be expected to be highly unstable.

fn get_parent_did(&self, id: NodeId) -> DefId

Unstable (rustc_private)

fn get_foreign_abi(&self, id: NodeId) -> Abi

Unstable (rustc_private)

fn expect_item(&self, id: NodeId) -> &'ast Item

Unstable (rustc_private)

fn expect_trait_item(&self, id: NodeId) -> &'ast TraitItem

Unstable (rustc_private)

fn expect_struct(&self, id: NodeId) -> &'ast VariantData

Unstable (rustc_private)

fn expect_variant(&self, id: NodeId) -> &'ast Variant

Unstable (rustc_private)

fn expect_foreign_item(&self, id: NodeId) -> &'ast ForeignItem

Unstable (rustc_private)

fn expect_expr(&self, id: NodeId) -> &'ast Expr

Unstable (rustc_private)

fn name(&self, id: NodeId) -> Name

Unstable (rustc_private)

Returns the name associated with the given NodeId's AST.

fn attrs(&self, id: NodeId) -> &'ast [Attribute]

Unstable (rustc_private)

Given a node ID, get a list of attributes associated with the AST corresponding to the Node ID

fn nodes_matching_suffix<'a>(&'a self, parts: &'a [String]) -> NodesMatchingSuffix<'a, 'ast>

Unstable (rustc_private)

Returns an iterator that yields the node id's with paths that match parts. (Requires parts is non-empty.)

For example, if given parts equal to ["bar", "quux"], then the iterator will produce node id's for items with paths such as foo::bar::quux, bar::quux, other::bar::quux, and any other such items it can find in the map.

fn opt_span(&self, id: NodeId) -> Option<Span>

Unstable (rustc_private)

fn span(&self, id: NodeId) -> Span

Unstable (rustc_private)

fn span_if_local(&self, id: DefId) -> Option<Span>

Unstable (rustc_private)

fn def_id_span(&self, def_id: DefId, fallback: Span) -> Span

Unstable (rustc_private)

fn node_to_string(&self, id: NodeId) -> String

Unstable (rustc_private)

fn node_to_user_string(&self, id: NodeId) -> String

Unstable (rustc_private)