Module rustc::hir::print

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use self::AnnNode::*;

Structs

NoAnn 	 [
Experimental
]

State 	 [
Experimental
]

Enums

AnnNode 	 [
Experimental
]

Nested 	 [
Experimental
]

Constants

NO_ANN 	 [
Experimental
]

default_columns 	 [
Experimental
]

indent_unit 	 [
Experimental
]

Traits

PpAnn 	 [
Experimental
]

Functions

print_crate 	 [
Experimental
]
Requires you to pass an input filename and reader so that it can scan the input text for comments and literals to copy forward.

to_string 	 [
Experimental
]

visibility_qualified 	 [
Experimental
]