Trait rustc::hir::print::PpAnn

pub trait PpAnn {
    fn nested(&self, _state: &mut State, _nested: Nested) -> Result<()> { ... }

    fn pre(&self, _state: &mut State, _node: AnnNode) -> Result<()> { ... }

    fn post(&self, _state: &mut State, _node: AnnNode) -> Result<()> { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn nested(&self, _state: &mut State, _nested: Nested) -> Result<()>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn pre(&self, _state: &mut State, _node: AnnNode) -> Result<()>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn post(&self, _state: &mut State, _node: AnnNode) -> Result<()>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?