Struct rustc::hir::Crate

pub struct Crate {
    pub module: Mod,
    pub attrs: HirVec<Attribute>,
    pub span: Span,
    pub exported_macros: HirVec<MacroDef>,
    pub items: BTreeMap<NodeId, Item>,
    pub trait_items: BTreeMap<TraitItemId, TraitItem>,
    pub impl_items: BTreeMap<ImplItemId, ImplItem>,
    pub bodies: FxHashMap<BodyId, Body>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

Fields
module: Mod

🔬 This is a nightly-only experimental API. (rustc_private)

attrs: HirVec<Attribute>

🔬 This is a nightly-only experimental API. (rustc_private)

span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

exported_macros: HirVec<MacroDef>

🔬 This is a nightly-only experimental API. (rustc_private)

items: BTreeMap<NodeId, Item>

🔬 This is a nightly-only experimental API. (rustc_private)

trait_items: BTreeMap<TraitItemId, TraitItem>

🔬 This is a nightly-only experimental API. (rustc_private)

impl_items: BTreeMap<ImplItemId, ImplItem>

🔬 This is a nightly-only experimental API. (rustc_private)

bodies: FxHashMap<BodyId, Body>

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl Crate

fn item(&self, id: NodeId) -> &Item

🔬 This is a nightly-only experimental API. (rustc_private)

fn trait_item(&self, id: TraitItemId) -> &TraitItem

🔬 This is a nightly-only experimental API. (rustc_private)

fn impl_item(&self, id: ImplItemId) -> &ImplItem

🔬 This is a nightly-only experimental API. (rustc_private)

fn visit_all_item_likes<'hir, V>(&'hir self, visitor: &mut V) where V: ItemLikeVisitor<'hir>

🔬 This is a nightly-only experimental API. (rustc_private)

Visits all items in the crate in some determinstic (but unspecified) order. If you just need to process every item, but don't care about nesting, this method is the best choice.

If you do care about nesting -- usually because your algorithm follows lexical scoping rules -- then you want a different approach. You should override visit_nested_item in your visitor and then call intravisit::walk_crate instead.

fn body(&self, id: BodyId) -> &Body

🔬 This is a nightly-only experimental API. (rustc_private)