Struct rustc::hir::Pat

pub struct Pat {
    pub id: NodeId,
    pub hir_id: HirId,
    pub node: PatKind,
    pub span: Span,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
id: NodeId

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

hir_id: HirId

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

node: PatKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl Pat

pub fn is_refutable(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn is_const(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn each_binding<F>(&self, f: F) where
F: FnMut(BindingAnnotation, NodeId, Span, &Spanned<Name>),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Call f on every "binding" in a pattern, e.g., on a in match foo() { Some(a) => (), None => () }

pub fn contains_bindings(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Checks if the pattern contains any patterns that bind something to an ident, e.g. foo, or Foo(foo) or foo @ Bar(..).

pub fn contains_bindings_or_wild(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Checks if the pattern contains any patterns that bind something to an ident or wildcard, e.g. foo, or Foo(_), foo @ Bar(..),

pub fn simple_name(&self) -> Option<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn necessary_variants(&self) -> Vec<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Return variants that are necessary to exist for the pattern to match.

pub fn contains_explicit_ref_binding(&self) -> Option<Mutability>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Checks if the pattern contains any ref or ref mut bindings, and if yes whether it contains mutable or just immutables ones.

FIXME(tschottdorf): this is problematic as the HIR is being scraped, but ref bindings are be implicit after #42640 (default match binding modes).

See #44848.

impl Pat

pub fn walk<F>(&self, it: F) -> bool where
F: FnMut(&Pat) -> bool,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?