Struct rustc::infer::at::At

pub struct At<'a, 'gcx: 'tcx, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'gcx, 'tcx> At<'a, 'gcx, 'tcx>

pub fn eq_impl_headers(
self,
expected: &ImplHeader<'tcx>,
actual: &ImplHeader<'tcx>
) -> InferResult<'tcx, ()>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Hacky routine for equating two impl headers in coherence.

pub fn sub_exp<T>(
self,
a_is_expected: bool,
a: T,
b: T
) -> InferResult<'tcx, ()> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make a <: b where a may or may not be expected

pub fn sup<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make actual <: expected. For example, if type-checking a call like foo(x), where foo: fn(i32), you might have sup(i32, x), since the "expected" type is the type that appears in the signature.

pub fn sub<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make expected <: actual

pub fn eq_exp<T>(self, a_is_expected: bool, a: T, b: T) -> InferResult<'tcx, ()> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make expected <: actual

pub fn eq<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make expected <: actual

pub fn lub<T>(self, expected: T, actual: T) -> InferResult<'tcx, T> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Compute the least-upper-bound, or mutual supertype, of two values. The order of the arguments doesn't matter, but since this can result in an error (e.g., if asked to compute LUB of u32 and i32), it is meaningful to call one of them the "expected type".

pub fn glb<T>(self, expected: T, actual: T) -> InferResult<'tcx, T> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Compute the greatest-lower-bound, or mutual subtype, of two values. As with lub order doesn't matter, except for error cases.

pub fn trace<T>(self, expected: T, actual: T) -> Trace<'a, 'gcx, 'tcx> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Sets the "trace" values that will be used for error-reporting, but doesn't actually perform any operation yet (this is useful when you want to set the trace using distinct values from those you wish to operate upon).

pub fn trace_exp<T>(
self,
a_is_expected: bool,
a: T,
b: T
) -> Trace<'a, 'gcx, 'tcx> where
T: ToTrace<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Like trace, but the expected value is determined by the boolean argument (if true, then the first argument a is the "expected" value).