Struct rustc::infer::at::Trace

pub struct Trace<'a, 'gcx: 'tcx, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'gcx, 'tcx> Trace<'a, 'gcx, 'tcx>

pub fn sub<T>(self, a: &T, b: &T) -> InferResult<'tcx, ()> where
T: Relate<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make a <: b where a may or may not be expected (if a_is_expected is true, then a is expected). Make expected <: actual

pub fn eq<T>(self, a: &T, b: &T) -> InferResult<'tcx, ()> where
T: Relate<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make a == b; the expectation is set by the call to trace().

pub fn lub<T>(self, a: &T, b: &T) -> InferResult<'tcx, T> where
T: Relate<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn glb<T>(self, a: &T, b: &T) -> InferResult<'tcx, T> where
T: Relate<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?