Struct rustc::infer::combine::CombineFields

pub struct CombineFields<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> {
    pub infcx: &'a InferCtxt<'a, 'gcx, 'tcx>,
    pub a_is_expected: bool,
    pub trace: TypeTrace<'tcx>,
    pub cause: Option<Cause>,
    pub obligations: PredicateObligations<'tcx>,
}

Unstable (rustc_private)

Fields
infcx: &'a InferCtxt<'a, 'gcx, 'tcx>
Unstable (rustc_private)
a_is_expected: bool
Unstable (rustc_private)
trace: TypeTrace<'tcx>
Unstable (rustc_private)
cause: Option<Cause>
Unstable (rustc_private)
obligations: PredicateObligations<'tcx>
Unstable (rustc_private)

Methods

impl<'a, 'gcx, 'tcx> CombineFields<'a, 'gcx, 'tcx>

fn tcx(&self) -> TyCtxt<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn switch_expected(&self) -> CombineFields<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn equate(&self) -> Equate<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn bivariate(&self) -> Bivariate<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn sub(&self) -> Sub<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn lub(&self) -> Lub<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn glb(&self) -> Glb<'a, 'gcx, 'tcx>

Unstable (rustc_private)

fn instantiate(&self, a_ty: Ty<'tcx>, dir: RelationDir, b_vid: TyVid) -> RelateResult<'tcx, ()>

Unstable (rustc_private)

impl<'a, 'gcx, 'tcx> CombineFields<'a, 'gcx, 'tcx>

fn higher_ranked_sub<T>(&self, a: &Binder<T>, b: &Binder<T>) -> RelateResult<'tcx, Binder<T>> where T: Relate<'tcx>

Unstable (rustc_private)

fn higher_ranked_match<T, U>(&self, span: Span, a_pair: &Binder<(T, U)>, b_match: &T) -> RelateResult<'tcx, HrMatchResult<U>> where T: Relate<'tcx>, U: TypeFoldable<'tcx>

Unstable (rustc_private)

The value consists of a pair (t, u) where t is the matcher and u is a value. The idea is to find a substitution S such that S(t) == b, and then return S(u). In other words, find values for the late-bound regions in a that can make t == b and then replace the LBR in u with those values.

This routine is (as of this writing) used in trait matching, particularly projection.

NB. It should not happen that there are LBR appearing in U that do not appear in T. If that happens, those regions are unconstrained, and this routine replaces them with 'static.

fn higher_ranked_lub<T>(&self, a: &Binder<T>, b: &Binder<T>) -> RelateResult<'tcx, Binder<T>> where T: Relate<'tcx>

Unstable (rustc_private)

fn higher_ranked_glb<T>(&self, a: &Binder<T>, b: &Binder<T>) -> RelateResult<'tcx, Binder<T>> where T: Relate<'tcx>

Unstable (rustc_private)