Module rustc::infer::region_constraints

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

See README.md

Structs

RegionConstraintCollector 	 [
Experimental
]

RegionConstraintData 	 [
Experimental
]
The full set of region constraints gathered up by the collector. Describes constraints between the region variables and other regions, as well as other conditions that must be verified, or assumptions that can be made.

RegionSnapshot 	 [
Experimental
]

TaintDirections 	 [
Experimental
]
When working with skolemized regions, we often wish to find all of the regions that are either reachable from a skolemized region, or which can reach a skolemized region, or both. We call such regions tained regions. This struct allows you to decide what set of tainted regions you want.

Verify 	 [
Experimental
]
VerifyGenericBound(T, _, R, RS): The parameter type T (or associated type) must outlive the region R. T is known to outlive RS. Therefore verify that R <= RS[i] for some i. Inference variables may be involved (but this verification step doesn't influence inference).

Enums

Constraint 	 [
Experimental
]
A constraint that influences the inference process.

GenericKind 	 [
Experimental
]

VerifyBound 	 [
Experimental
]
When we introduce a verification step, we wish to test that a particular region (let's call it 'min) meets some bound. The bound is described the by the following grammar:

Type Definitions

VarOrigins 	 [
Experimental
]