Struct rustc::infer::region_constraints::RegionConstraintData

pub struct RegionConstraintData<'tcx> {
    pub constraints: BTreeMap<Constraint<'tcx>, SubregionOrigin<'tcx>>,
    pub verifys: Vec<Verify<'tcx>>,
    pub givens: FxHashSet<(Region<'tcx>, RegionVid)>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The full set of region constraints gathered up by the collector. Describes constraints between the region variables and other regions, as well as other conditions that must be verified, or assumptions that can be made.

Fields
constraints: BTreeMap<Constraint<'tcx>, SubregionOrigin<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Constraints of the form A <= B, where either A or B can be a region variable (or neither, as it happens).

verifys: Vec<Verify<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A "verify" is something that we need to verify after inference is done, but which does not directly affect inference in any way.

An example is a A <= B where neither A nor B are inference variables.

givens: FxHashSet<(Region<'tcx>, RegionVid)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A "given" is a relationship that is known to hold. In particular, we often know from closure fn signatures that a particular free region must be a subregion of a region variable:

foo.iter().filter(<'a> |x: &'a &'b T| ...)

In situations like this, 'b is in fact a region variable introduced by the call to iter(), and 'a is a bound region on the closure (as indicated by the <'a> prefix). If we are naive, we wind up inferring that 'b must be 'static, because we require that it be greater than 'a and we do not know what 'a is precisely.

This hashmap is used to avoid that naive scenario. Basically we record the fact that 'a <= 'b is implied by the fn signature, and then ignore the constraint when solving equations. This is a bit of a hack but seems to work.

Methods

impl<'tcx> RegionConstraintData<'tcx>

pub fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

True if this region constraint data contains no constraints.