Enum rustc::infer::region_inference::RegionResolutionError

pub enum RegionResolutionError<'tcx> {
    ConcreteFailure(SubregionOrigin<'tcx>, Region<'tcx>, Region<'tcx>),
    GenericBoundFailure(SubregionOrigin<'tcx>, GenericKind<'tcx>, Region<'tcx>),
    SubSupConflict(RegionVariableOrigin, SubregionOrigin<'tcx>, Region<'tcx>, SubregionOrigin<'tcx>, Region<'tcx>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
ConcreteFailure(SubregionOrigin<'tcx>, Region<'tcx>, Region<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

ConcreteFailure(o, a, b):

o requires that a <= b, but this does not hold

GenericBoundFailure(SubregionOrigin<'tcx>, GenericKind<'tcx>, Region<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

`GenericBoundFailure(p, s, a)

The parameter/associated-type p must be known to outlive the lifetime a (but none of the known bounds are sufficient).

SubSupConflict(RegionVariableOrigin, SubregionOrigin<'tcx>, Region<'tcx>, SubregionOrigin<'tcx>, Region<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

SubSupConflict(v, sub_origin, sub_r, sup_origin, sup_r):

Could not infer a value for v because sub_r <= v (due to sub_origin) but v <= sup_r (due to sup_origin) and sub_r <= sup_r does not hold.