Enum rustc::infer::region_inference::VerifyBound

pub enum VerifyBound<'tcx> {
    AnyRegion(Vec<Region<'tcx>>),
    AllRegions(Vec<Region<'tcx>>),
    AnyBound(Vec<VerifyBound<'tcx>>),
    AllBounds(Vec<VerifyBound<'tcx>>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

When we introduce a verification step, we wish to test that a particular region (let's call it 'min) meets some bound. The bound is described the by the following grammar:

Variants
AnyRegion(Vec<Region<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

B = exists {R} --> some 'r in {R} must outlive 'min

Put another way, the subject value is known to outlive all regions in {R}, so if any of those outlives 'min, then the bound is met.

AllRegions(Vec<Region<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

B = forall {R} --> all 'r in {R} must outlive 'min

Put another way, the subject value is known to outlive some region in {R}, so if all of those outlives 'min, then the bound is met.

AnyBound(Vec<VerifyBound<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

B = exists {B} --> 'min must meet some bound b in {B}

AllBounds(Vec<VerifyBound<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

B = forall {B} --> 'min must meet all bounds b in {B}

Methods

impl<'a, 'gcx, 'tcx> VerifyBound<'tcx>

pub fn must_hold(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn cannot_hold(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn or(self, vb: VerifyBound<'tcx>) -> VerifyBound<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn and(self, vb: VerifyBound<'tcx>) -> VerifyBound<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?