Struct rustc::infer::resolve::OpportunisticTypeAndRegionResolver

pub struct OpportunisticTypeAndRegionResolver<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The opportunistic type and region resolver is similar to the opportunistic type resolver, but also opportunistically resolves regions. It is useful for canonicalization.

Methods

impl<'a, 'gcx, 'tcx> OpportunisticTypeAndRegionResolver<'a, 'gcx, 'tcx>

pub fn new(infcx: &'a InferCtxt<'a, 'gcx, 'tcx>) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?