Struct rustc::infer::resolve::OpportunisticTypeResolver

pub struct OpportunisticTypeResolver<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> { /* fields omitted */ }

Unstable (rustc_private)

The opportunistic type resolver can be used at any time. It simply replaces type variables that have been unified with the things they have been unified with (similar to shallow_resolve, but deep). This is useful for printing messages etc but also required at various points for correctness.

Methods

impl<'a, 'gcx, 'tcx> OpportunisticTypeResolver<'a, 'gcx, 'tcx>

fn new(infcx: &'a InferCtxt<'a, 'gcx, 'tcx>) -> Self

Unstable (rustc_private)