Struct rustc::infer::InferCtxtBuilder

pub struct InferCtxtBuilder<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Helper type of a temporary returned by tcx.infer_ctxt(). Necessary because we can't write the following bound: F: for<'b, 'tcx> where 'gcx: 'tcx FnOnce(InferCtxt<'b, 'gcx, 'tcx>).

Methods

impl<'a, 'gcx, 'tcx> InferCtxtBuilder<'a, 'gcx, 'tcx>

pub fn with_fresh_in_progress_tables(self, table_owner: DefId) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Used only by rustc_typeck during body type-checking/inference, will initialize in_progress_tables with fresh TypeckTables.

pub fn enter<F, R>(&'tcx mut self, f: F) -> R where
F: for<'b> FnOnce(InferCtxt<'b, 'gcx, 'tcx>) -> R,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?