Trait rustc::infer::TransNormalize

pub trait TransNormalize<'gcx>: TypeFoldable<'gcx> {
    fn trans_normalize<'a, 'tcx>(
        &self, 
        infcx: &InferCtxt<'a, 'gcx, 'tcx>, 
        param_env: ParamEnv<'tcx>
    ) -> Self;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Helper trait for shortening the lifetimes inside a value for post-type-checking normalization.

Required Methods

fn trans_normalize<'a, 'tcx>(
&self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>,
param_env: ParamEnv<'tcx>
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?