Module rustc::lint::builtin

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Some lints that are built in to the compiler.

These are the built-in lints that are emitted direct in the main compiler code, rather than using their own custom pass. Those lints are all available in rustc_lint::builtin.

Structs

HardwiredLints 	 [
Experimental
]
Does nothing as a lint pass, but registers some Lints which are used by other parts of the compiler.

Statics

CONST_ERR 	 [
Experimental
]

DEAD_CODE 	 [
Experimental
]

DEPRECATED 	 [
Experimental
]

EXTRA_REQUIREMENT_IN_IMPL 	 [
Experimental
]

FAT_PTR_TRANSMUTES 	 [
Experimental
]

INVALID_TYPE_PARAM_DEFAULT 	 [
Experimental
]

LATE_BOUND_LIFETIME_ARGUMENTS 	 [
Experimental
]

LEGACY_CONSTRUCTOR_VISIBILITY 	 [
Experimental
]

LEGACY_DIRECTORY_OWNERSHIP 	 [
Experimental
]

LEGACY_IMPORTS 	 [
Experimental
]

MISSING_FRAGMENT_SPECIFIER 	 [
Experimental
]

PARENTHESIZED_PARAMS_IN_TYPES_AND_MODULES 	 [
Experimental
]

PATTERNS_IN_FNS_WITHOUT_BODY 	 [
Experimental
]

PRIVATE_IN_PUBLIC 	 [
Experimental
]

PUB_USE_OF_PRIVATE_EXTERN_CRATE 	 [
Experimental
]

RENAMED_AND_REMOVED_LINTS 	 [
Experimental
]

RESOLVE_TRAIT_ON_DEFAULTED_UNIT 	 [
Experimental
]

SAFE_EXTERN_STATICS 	 [
Experimental
]

STABLE_FEATURES 	 [
Experimental
]

TRIVIAL_CASTS 	 [
Experimental
]

TRIVIAL_NUMERIC_CASTS 	 [
Experimental
]

UNKNOWN_CRATE_TYPES 	 [
Experimental
]

UNKNOWN_LINTS 	 [
Experimental
]

UNREACHABLE_CODE 	 [
Experimental
]

UNREACHABLE_PATTERNS 	 [
Experimental
]

UNUSED_ASSIGNMENTS 	 [
Experimental
]

UNUSED_EXTERN_CRATES 	 [
Experimental
]

UNUSED_FEATURES 	 [
Experimental
]

UNUSED_IMPORTS 	 [
Experimental
]

UNUSED_MACROS 	 [
Experimental
]

UNUSED_MUT 	 [
Experimental
]

UNUSED_QUALIFICATIONS 	 [
Experimental
]

UNUSED_UNSAFE 	 [
Experimental
]

UNUSED_VARIABLES 	 [
Experimental
]

WARNINGS 	 [
Experimental
]