Struct rustc::lint::builtin::HardwiredLints

pub struct HardwiredLints;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Does nothing as a lint pass, but registers some Lints which are used by other parts of the compiler.