Function rustc::lint::raw_emit_lint

pub fn raw_emit_lint(sess: &Session, lints: &LintStore, lint: &'static Lint, lvlsrc: LevelSource, span: Option<Span>, msg: &str)

Unstable (rustc_private)

Emit a lint as a warning or an error (or not at all) according to level.

This lives outside of Context so it can be used by checks in trans that run after the main lint pass is finished. Most lints elsewhere in the compiler should call Session::add_lint() instead.