Function rustc::lint::struct_lint_level

pub fn struct_lint_level<'a>(
    sess: &'a Session, 
    lint: &'static Lint, 
    level: Level, 
    src: LintSource, 
    span: Option<MultiSpan>, 
    msg: &str
) -> DiagnosticBuilder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?