Struct rustc::lint::Context

pub struct Context<'a, 'tcx> {
    pub tcx: &'a ctxt<'tcx>,
    pub krate: &'a Crate,
    pub exported_items: &'a ExportedItems,
    // some fields omitted
}

Unstable

Context for lint checking.

Fields

tcx
Unstable

Type context we're checking in.

krate
Unstable

The crate being checked.

exported_items
Unstable

Items exported from the crate being checked.

Methods

impl<'a, 'tcx> Context<'a, 'tcx>

fn sess(&'a self) -> &'a Session

Unstable

Get the overall compiler Session object.

fn current_level(&self, lint: &'static Lint) -> Level

Unstable

Get the level of lint at the current position of the lint traversal.

fn lint(&self, lint: &'static Lint, msg: &str)

Unstable

Emit a lint at the appropriate level, with no associated span.

fn span_lint(&self, lint: &'static Lint, span: Span, msg: &str)

Unstable

Emit a lint at the appropriate level, for a particular span.