Struct rustc::lint::EarlyContext

pub struct EarlyContext<'a> {
    pub sess: &'a Session,
    pub krate: &'a Crate,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Context for lint checking of the AST, after expansion, before lowering to HIR.

Fields
sess: &'a Session

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Type context we're checking in.

krate: &'a Crate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The crate being checked.