Struct rustc::lint::LintLevelMap

pub struct LintLevelMap { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl LintLevelMap

pub fn level_and_source(
&self,
lint: &'static Lint,
id: HirId
) -> Option<(Level, LintSource)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If the id was previously registered with register_id when building this LintLevelMap this returns the corresponding lint level and source of the lint level for the lint provided.

If the id was not previously registered, returns None. If None is returned then the parent of id should be acquired and this function should be called again.

pub fn lint_level_set(&self, id: HirId) -> Option<u32>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns if this id has lint level information.