Trait rustc::lint::EarlyLintPass

pub trait EarlyLintPass: LintPass {
    fn check_ident(&mut self, _: &EarlyContext, _: Span, _: Ident) { ... }

    fn check_crate(&mut self, _: &EarlyContext, _: &Crate) { ... }

    fn check_crate_post(&mut self, _: &EarlyContext, _: &Crate) { ... }

    fn check_mod(&mut self, _: &EarlyContext, _: &Mod, _: Span, _: NodeId) { ... }

    fn check_mod_post(&mut self, _: &EarlyContext, _: &Mod, _: Span, _: NodeId) { ... }

    fn check_foreign_item(&mut self, _: &EarlyContext, _: &ForeignItem) { ... }

    fn check_foreign_item_post(&mut self, _: &EarlyContext, _: &ForeignItem) { ... }

    fn check_item(&mut self, _: &EarlyContext, _: &Item) { ... }

    fn check_item_post(&mut self, _: &EarlyContext, _: &Item) { ... }

    fn check_local(&mut self, _: &EarlyContext, _: &Local) { ... }

    fn check_block(&mut self, _: &EarlyContext, _: &Block) { ... }

    fn check_block_post(&mut self, _: &EarlyContext, _: &Block) { ... }

    fn check_stmt(&mut self, _: &EarlyContext, _: &Stmt) { ... }

    fn check_arm(&mut self, _: &EarlyContext, _: &Arm) { ... }

    fn check_pat(&mut self, _: &EarlyContext, _: &Pat) { ... }

    fn check_expr(&mut self, _: &EarlyContext, _: &Expr) { ... }

    fn check_expr_post(&mut self, _: &EarlyContext, _: &Expr) { ... }

    fn check_ty(&mut self, _: &EarlyContext, _: &Ty) { ... }

    fn check_generics(&mut self, _: &EarlyContext, _: &Generics) { ... }

    fn check_fn(
        &mut self, 
        _: &EarlyContext, 
        _: FnKind, 
        _: &FnDecl, 
        _: Span, 
        _: NodeId
    ) { ... }

    fn check_fn_post(
        &mut self, 
        _: &EarlyContext, 
        _: FnKind, 
        _: &FnDecl, 
        _: Span, 
        _: NodeId
    ) { ... }

    fn check_trait_item(&mut self, _: &EarlyContext, _: &TraitItem) { ... }

    fn check_trait_item_post(&mut self, _: &EarlyContext, _: &TraitItem) { ... }

    fn check_impl_item(&mut self, _: &EarlyContext, _: &ImplItem) { ... }

    fn check_impl_item_post(&mut self, _: &EarlyContext, _: &ImplItem) { ... }

    fn check_struct_def(
        &mut self, 
        _: &EarlyContext, 
        _: &VariantData, 
        _: Ident, 
        _: &Generics, 
        _: NodeId
    ) { ... }

    fn check_struct_def_post(
        &mut self, 
        _: &EarlyContext, 
        _: &VariantData, 
        _: Ident, 
        _: &Generics, 
        _: NodeId
    ) { ... }

    fn check_struct_field(&mut self, _: &EarlyContext, _: &StructField) { ... }

    fn check_variant(&mut self, _: &EarlyContext, _: &Variant, _: &Generics) { ... }

    fn check_variant_post(
        &mut self, 
        _: &EarlyContext, 
        _: &Variant, 
        _: &Generics
    ) { ... }

    fn check_lifetime(&mut self, _: &EarlyContext, _: &Lifetime) { ... }

    fn check_lifetime_def(&mut self, _: &EarlyContext, _: &LifetimeDef) { ... }

    fn check_path(&mut self, _: &EarlyContext, _: &Path, _: NodeId) { ... }

    fn check_path_list_item(&mut self, _: &EarlyContext, _: &PathListItem) { ... }

    fn check_attribute(&mut self, _: &EarlyContext, _: &Attribute) { ... }

    fn enter_lint_attrs(&mut self, _: &EarlyContext, _: &[Attribute]) { ... }

    fn exit_lint_attrs(&mut self, _: &EarlyContext, _: &[Attribute]) { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn check_ident(&mut self, _: &EarlyContext, _: Span, _: Ident)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_crate(&mut self, _: &EarlyContext, _: &Crate)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_crate_post(&mut self, _: &EarlyContext, _: &Crate)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_mod(&mut self, _: &EarlyContext, _: &Mod, _: Span, _: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_mod_post(&mut self, _: &EarlyContext, _: &Mod, _: Span, _: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_foreign_item(&mut self, _: &EarlyContext, _: &ForeignItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_foreign_item_post(&mut self, _: &EarlyContext, _: &ForeignItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_item(&mut self, _: &EarlyContext, _: &Item)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_item_post(&mut self, _: &EarlyContext, _: &Item)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_local(&mut self, _: &EarlyContext, _: &Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_block(&mut self, _: &EarlyContext, _: &Block)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_block_post(&mut self, _: &EarlyContext, _: &Block)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_stmt(&mut self, _: &EarlyContext, _: &Stmt)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_arm(&mut self, _: &EarlyContext, _: &Arm)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_pat(&mut self, _: &EarlyContext, _: &Pat)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_expr(&mut self, _: &EarlyContext, _: &Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_expr_post(&mut self, _: &EarlyContext, _: &Expr)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_ty(&mut self, _: &EarlyContext, _: &Ty)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_generics(&mut self, _: &EarlyContext, _: &Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_fn(
&mut self,
_: &EarlyContext,
_: FnKind,
_: &FnDecl,
_: Span,
_: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_fn_post(
&mut self,
_: &EarlyContext,
_: FnKind,
_: &FnDecl,
_: Span,
_: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_trait_item(&mut self, _: &EarlyContext, _: &TraitItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_trait_item_post(&mut self, _: &EarlyContext, _: &TraitItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_impl_item(&mut self, _: &EarlyContext, _: &ImplItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_impl_item_post(&mut self, _: &EarlyContext, _: &ImplItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_struct_def(
&mut self,
_: &EarlyContext,
_: &VariantData,
_: Ident,
_: &Generics,
_: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_struct_def_post(
&mut self,
_: &EarlyContext,
_: &VariantData,
_: Ident,
_: &Generics,
_: NodeId
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_struct_field(&mut self, _: &EarlyContext, _: &StructField)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_variant(&mut self, _: &EarlyContext, _: &Variant, _: &Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_variant_post(&mut self, _: &EarlyContext, _: &Variant, _: &Generics)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_lifetime(&mut self, _: &EarlyContext, _: &Lifetime)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_lifetime_def(&mut self, _: &EarlyContext, _: &LifetimeDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_path(&mut self, _: &EarlyContext, _: &Path, _: NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_path_list_item(&mut self, _: &EarlyContext, _: &PathListItem)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn check_attribute(&mut self, _: &EarlyContext, _: &Attribute)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn enter_lint_attrs(&mut self, _: &EarlyContext, _: &[Attribute])

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Called when entering a syntax node that can have lint attributes such as #[allow(...)]. Called with all the attributes of that node.

fn exit_lint_attrs(&mut self, _: &EarlyContext, _: &[Attribute])

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Counterpart to enter_lint_attrs.