Struct rustc::llvm::archive_ro::ArchiveRO

pub struct ArchiveRO {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl ArchiveRO

fn open(dst: &Path) -> Option<ArchiveRO>

Unstable (rustc_private)

Opens a static archive for read-only purposes. This is more optimized than the open method because it uses LLVM's internal Archive class rather than shelling out to ar for everything.

If this archive is used with a mutable method, then an error will be raised.

fn raw(&self) -> *mut Archive_opaque

Unstable (rustc_private)

fn iter(&self) -> Iter

Unstable (rustc_private)