Struct rustc::middle::borrowck::BorrowCheckResult

pub struct BorrowCheckResult {
    pub used_mut_nodes: FxHashSet<HirId>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
used_mut_nodes: FxHashSet<HirId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?