Struct rustc::middle::const_qualif::ConstQualif

pub struct ConstQualif {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl ConstQualif

const MUTABLE_MEM: ConstQualif = ConstQualif{bits: 1 << 0,}

Unstable (rustc_private)

const NEEDS_DROP: ConstQualif = ConstQualif{bits: 1 << 1,}

Unstable (rustc_private)

const PREFER_IN_PLACE: ConstQualif = ConstQualif{bits: 1 << 2,}

Unstable (rustc_private)

const NON_ZERO_SIZED: ConstQualif = ConstQualif{bits: 1 << 3,}

Unstable (rustc_private)

const HAS_STATIC_BORROWS: ConstQualif = ConstQualif{bits: 1 << 4,}

Unstable (rustc_private)

const NOT_CONST: ConstQualif = ConstQualif{bits: 1 << 5,}

Unstable (rustc_private)

const NON_STATIC_BORROWS: ConstQualif = ConstQualif{bits: ConstQualif::MUTABLE_MEM.bits | ConstQualif::NEEDS_DROP.bits | ConstQualif::NOT_CONST.bits,}

Unstable (rustc_private)

fn empty() -> ConstQualif

Unstable (rustc_private)

Returns an empty set of flags.

fn all() -> ConstQualif

Unstable (rustc_private)

Returns the set containing all flags.

fn bits(&self) -> u8

Unstable (rustc_private)

Returns the raw value of the flags currently stored.

fn from_bits(bits: u8) -> Option<ConstQualif>

Unstable (rustc_private)

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u8) -> ConstQualif

Unstable (rustc_private)

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Unstable (rustc_private)

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Unstable (rustc_private)

Returns true if all flags are currently set.

fn intersects(&self, other: ConstQualif) -> bool

Unstable (rustc_private)

Returns true if there are flags common to both self and other.

fn contains(&self, other: ConstQualif) -> bool

Unstable (rustc_private)

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: ConstQualif)

Unstable (rustc_private)

Inserts the specified flags in-place.

fn remove(&mut self, other: ConstQualif)

Unstable (rustc_private)

Removes the specified flags in-place.

fn toggle(&mut self, other: ConstQualif)

Unstable (rustc_private)

Toggles the specified flags in-place.