Enum rustc::middle::const_val::ConstVal

pub enum ConstVal {
    Float(f64),
    Integral(ConstInt),
    Str(InternedString),
    ByteStr(Rc<Vec<u8>>),
    Bool(bool),
    Struct(NodeId),
    Tuple(NodeId),
    Function(DefId),
    Array(NodeId, u64),
    Repeat(NodeId, u64),
    Char(char),
    Dummy,
}

Unstable (rustc_private)

Variants

Float
Unstable (rustc_private)

Integral
Unstable (rustc_private)

Str
Unstable (rustc_private)

ByteStr
Unstable (rustc_private)

Bool
Unstable (rustc_private)

Struct
Unstable (rustc_private)

Tuple
Unstable (rustc_private)

Function
Unstable (rustc_private)

Array
Unstable (rustc_private)

Repeat
Unstable (rustc_private)

Char
Unstable (rustc_private)

Dummy
Unstable (rustc_private)

A value that only occurs in case eval_const_expr reported an error. You should never handle this case. Its sole purpose is to allow more errors to be reported instead of causing a fatal error.

Methods

impl ConstVal

fn description(&self) -> &'static str

Unstable (rustc_private)