Module rustc::middle::cstore

🔬 This is a nightly-only experimental API. (rustc_private)

Reexports

pub use self::NativeLibraryKind::NativeStatic;

pub use self::NativeLibraryKind::NativeFramework;

pub use self::NativeLibraryKind::NativeUnknown;

Structs

CrateSource 	 [
Experimental
]

DummyCrateStore 	 [
Experimental
]
A dummy crate store that does not support any non-local crates, for test purposes.

ExternCrate 	 [
Experimental
]

LinkMeta 	 [
Experimental
]

NativeLibrary 	 [
Experimental
]

Enums

DepKind 	 [
Experimental
]

LibSource 	 [
Experimental
]

LinkagePreference 	 [
Experimental
]

LoadedMacro 	 [
Experimental
]

NativeLibraryKind 	 [
Experimental
]

Traits

CrateLoader 	 [
Experimental
]

CrateStore 	 [
Experimental
]
A store of Rust crates, through with their metadata can be accessed.

Functions

validate_crate_name 	 [
Experimental
]