Struct rustc::middle::dataflow::DataFlowContext

pub struct DataFlowContext<'a, 'tcx: 'a, O> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx, O: DataFlowOperator> DataFlowContext<'a, 'tcx, O>

pub fn new(
tcx: TyCtxt<'a, 'tcx, 'tcx>,
analysis_name: &'static str,
body: Option<&Body>,
cfg: &CFG,
oper: O,
id_range: IdRange,
bits_per_id: usize
) -> DataFlowContext<'a, 'tcx, O>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn add_gen(&mut self, id: ItemLocalId, bit: usize)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates that id generates bit

pub fn add_kill(&mut self, kind: KillFrom, id: ItemLocalId, bit: usize)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates that id kills bit

pub fn each_bit_on_entry<F>(&self, id: ItemLocalId, f: F) -> bool where
F: FnMut(usize) -> bool,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterates through each bit that is set on entry to id. Only useful after propagate() has been called.

pub fn each_bit_for_node<F>(
&self,
e: EntryOrExit,
cfgidx: CFGIndex,
f: F
) -> bool where
F: FnMut(usize) -> bool,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterates through each bit that is set on entry/exit to cfgidx. Only useful after propagate() has been called.

pub fn each_gen_bit<F>(&self, id: ItemLocalId, f: F) -> bool where
F: FnMut(usize) -> bool,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterates through each bit in the gen set for id.

pub fn add_kills_from_flow_exits(&mut self, cfg: &CFG)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Whenever you have a break or continue statement, flow exits through any number of enclosing scopes on its way to the new destination. This function infers the kill bits of those control operators based on the kill bits associated with those scopes.

This is usually called (if it is called at all), after all add_gen and add_kill calls, but before propagate.

impl<'a, 'tcx, O: DataFlowOperator + Clone + 'static> DataFlowContext<'a, 'tcx, O>

pub fn propagate(&mut self, cfg: &CFG, body: &Body)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Performs the data flow analysis.