Struct rustc::middle::free_region::FreeRegionMap

pub struct FreeRegionMap<'tcx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'tcx> FreeRegionMap<'tcx>

pub fn new() -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn relate_free_regions_from_predicates(
&mut self,
predicates: &[Predicate<'tcx>]
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn relate_regions(&mut self, sub: Region<'tcx>, sup: Region<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Record that 'sup:'sub. Or, put another way, 'sub <= 'sup. (with the exception that 'static: 'x is not notable)

pub fn sub_free_regions<'a, 'gcx>(
&self,
r_a: Region<'tcx>,
r_b: Region<'tcx>
) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

True if r_a <= r_b is known to hold. Both r_a and r_b must be free regions from the function header.

pub fn lub_free_regions<'a, 'gcx>(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
r_a: Region<'tcx>,
r_b: Region<'tcx>
) -> Region<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Compute the least-upper-bound of two free regions. In some cases, this is more conservative than necessary, in order to avoid making arbitrary choices. See TransitiveRelation::postdom_upper_bound for more details.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn regions_that_outlive<'a, 'gcx>(
&self,
r_a: Region<'tcx>
) -> Vec<&Region<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns all regions that are known to outlive r_a. For example, in a function:

fn foo<'a, 'b: 'a, 'c: 'b>() { .. }

if r_a represents 'a, this function would return {'b, 'c}.