Struct rustc::middle::infer::region_inference::SameRegions

pub struct SameRegions {
    pub scope_id: NodeId,
    pub regions: Vec<BoundRegion>,
}

Unstable

SameRegions is used to group regions that we think are the same and would like to indicate so to the user. For example, the following function struct Foo { bar: int } fn foo2<'a, 'b>(x: &'a Foo) -> &'b int { &x.bar } would report an error because we expect 'a and 'b to match, and so we group 'a and 'b together inside a SameRegions struct

Fields

scope_id
Unstable

regions
Unstable

Methods

impl SameRegions

fn contains(&self, other: &BoundRegion) -> bool

Unstable

fn push(&mut self, other: BoundRegion)

Unstable