Struct rustc::middle::infer::resolve::OpportunisticTypeResolver

pub struct OpportunisticTypeResolver<'a, 'tcx> {
    // some fields omitted
}

Unstable (rustc_private)

The opportunistic type resolver can be used at any time. It simply replaces type variables that have been unified with the things they have been unified with (similar to shallow_resolve, but deep). This is useful for printing messages etc but also required at various points for correctness.

Methods

impl<'a, 'tcx> OpportunisticTypeResolver<'a, 'tcx>

fn new(infcx: &'a InferCtxt<'a, 'tcx>) -> OpportunisticTypeResolver<'a, 'tcx>

Unstable (rustc_private)