Enum rustc::middle::mem_categorization::ImmutabilityBlame

pub enum ImmutabilityBlame<'tcx> {
    ImmLocal(NodeId),
    ClosureEnv(DefIndex),
    LocalDeref(NodeId),
    AdtFieldDeref(&'tcx AdtDef, &'tcx FieldDef),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
ImmLocal(NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

ClosureEnv(DefIndex)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

LocalDeref(NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

AdtFieldDeref(&'tcx AdtDef, &'tcx FieldDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?