Module rustc::middle::pat_util

Unstable

Functions

arm_contains_ref_binding 	 [Unstable]
Checks if the patterns for this arm contain any ref or ref mut bindings.

def_to_path 	 [Unstable]

pat_bindings 	 [Unstable]
Call it on every "binding" in a pattern, e.g., on a in match foo() { Some(a) => (), None => () }

pat_contains_bindings 	 [Unstable]
Checks if the pattern contains any patterns that bind something to an ident, e.g. foo, or Foo(foo) or foo @ Bar(..).

pat_contains_bindings_or_wild 	 [Unstable]
Checks if the pattern contains any patterns that bind something to an ident or wildcard, e.g. foo, or Foo(_), foo @ Bar(..),

pat_contains_ref_binding 	 [Unstable]
Checks if the pattern contains any ref or ref mut bindings.

pat_id_map 	 [Unstable]

pat_is_binding 	 [Unstable]

pat_is_binding_or_wild 	 [Unstable]

pat_is_const 	 [Unstable]

pat_is_refutable 	 [Unstable]

pat_is_resolved_const 	 [Unstable]

pat_is_variant_or_struct 	 [Unstable]

simple_identifier 	 [Unstable]

Type Definitions

PatIdMap 	 [Unstable]