Module rustc::middle::region

Unstable

This file actually contains two passes related to regions. The first pass builds up the scope_map, which describes the parent links in the region hierarchy. The second pass infers which types must be region parameterized.

Most of the documentation on regions can be found in middle/typeck/infer/region_inference.rs

Structs

BlockRemainder 	 [Unstable]
Represents a subscope of block for a binding that is introduced by block.stmts[first_statement_index]. Such subscopes represent a suffix of the block. Note that each subscope does not include the initializer expression, if any, for the statement indexed by first_statement_index.

Context 	 [Unstable]

DestructionScopeData 	 [Unstable]
extent of destructors for temporaries of node-id

RegionMaps 	 [Unstable]
The region maps encode information about region relationships.

Enums

CodeExtent 	 [Unstable]
CodeExtent represents a statically-describable extent that can be used to bound the lifetime/region for values.

Functions

resolve_crate 	 [Unstable]

resolve_inlined_item 	 [Unstable]