Struct rustc::middle::region::RegionMaps

pub struct RegionMaps { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The region maps encode information about region relationships.

Methods

impl<'tcx> RegionMaps

fn new() -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn record_code_extent(&mut self, child: CodeExtent, parent: Option<CodeExtent>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn each_encl_scope<E>(&self, e: E) where
E: FnMut(CodeExtent, CodeExtent),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn each_var_scope<E>(&self, e: E) where
E: FnMut(&NodeId, CodeExtent),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn opt_destruction_extent(&self, n: NodeId) -> Option<CodeExtent>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn opt_encl_scope(&self, id: CodeExtent) -> Option<CodeExtent>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the narrowest scope that encloses id, if any.

fn encl_scope(&self, id: CodeExtent) -> CodeExtent

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the narrowest scope that encloses id, if any.

fn var_scope(&self, var_id: NodeId) -> CodeExtent

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the lifetime of the local variable var_id

fn temporary_scope(&self, expr_id: NodeId) -> Option<CodeExtent>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the scope when temp created by expr_id will be cleaned up

fn var_region(&self, id: NodeId) -> RegionKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the lifetime of the variable id.

fn scopes_intersect(&self, scope1: CodeExtent, scope2: CodeExtent) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_subscope_of(&self, subscope: CodeExtent, superscope: CodeExtent) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns true if subscope is equal to or is lexically nested inside superscope and false otherwise.

fn nearest_common_ancestor(
&self,
scope_a: CodeExtent,
scope_b: CodeExtent
) -> CodeExtent

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Finds the nearest common ancestor (if any) of two scopes. That is, finds the smallest scope which is greater than or equal to both scope_a and scope_b.

fn early_free_extent<'a, 'gcx>(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
br: &EarlyBoundRegion
) -> CodeExtent

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Assuming that the provided region was defined within this RegionMaps, returns the outermost CodeExtent that the region outlives.

fn free_extent<'a, 'gcx>(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
fr: &FreeRegion
) -> CodeExtent

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Assuming that the provided region was defined within this RegionMaps, returns the outermost CodeExtent that the region outlives.