Struct rustc::middle::resolve_lifetime::NamedRegionMap

pub struct NamedRegionMap {
    pub defs: NodeMap<Region>,
    pub late_bound: NodeSet,
    pub object_lifetime_defaults: NodeMap<Vec<ObjectLifetimeDefault>>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
defs: NodeMap<Region>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

late_bound: NodeSet

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

object_lifetime_defaults: NodeMap<Vec<ObjectLifetimeDefault>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?