Module rustc::middle::stability

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.

Reexports

pub use self::StabilityLevel::*;

Structs

DeprecationEntry 	 [
Experimental
]
An entry in the depr_map.

Index 	 [
Experimental
]
A stability index, giving the stability level for items and methods.

Enums

StabilityLevel 	 [
Experimental
]

Functions

check_unstable_api_usage 	 [
Experimental
]
Cross-references the feature names of unstable APIs with enabled features and possibly prints errors.

check_unused_or_stable_features 	 [
Experimental
]
Given the list of enabled features that were not language features (i.e. that were expected to be library features), and the list of features used from libraries, identify activated features that don't exist and error about them.