Struct rustc::middle::stability::Index

pub struct Index<'tcx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A stability index, giving the stability level for items and methods.

Methods

impl<'a, 'tcx> Index<'tcx>

pub fn new(tcx: TyCtxt<'a, 'tcx, 'tcx>) -> Index<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn local_stability(&self, id: HirId) -> Option<&'tcx Stability>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn local_deprecation_entry(&self, id: HirId) -> Option<DeprecationEntry>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?