Module rustc::middle::subst

Unstable

Reexports

pub use self::ParamSpace::*;

pub use self::RegionSubsts::*;

Structs

EnumeratedItems 	 [Unstable]

SeparateVecsPerParamSpace 	 [Unstable]
The split function converts one VecPerParamSpace into this SeparateVecsPerParamSpace structure.

Substs 	 [Unstable]
A substitution mapping type/region parameters to new values. We identify each in-scope parameter by an index and a parameter space (which indices where the parameter is defined; see ParamSpace).

VecPerParamSpace 	 [Unstable]
Vector of things sorted by param space. Used to keep the set of things declared on the type, self, or method distinct.

Enums

ParamSpace 	 [Unstable]

RegionSubsts 	 [Unstable]
Represents the values to use when substituting lifetime parameters. If the value is ErasedRegions, then this subst is occurring during trans, and all region parameters will be replaced with ty::ReStatic.

Traits

Subst 	 [Unstable]