Enum rustc::middle::subst::RegionSubsts

pub enum RegionSubsts {
    ErasedRegions,
    NonerasedRegions(VecPerParamSpace<Region>),
}

Unstable (rustc_private)

Represents the values to use when substituting lifetime parameters. If the value is ErasedRegions, then this subst is occurring during trans, and all region parameters will be replaced with ty::ReStatic.

Variants

ErasedRegions
Unstable (rustc_private)

NonerasedRegions
Unstable (rustc_private)

Methods

impl RegionSubsts

fn map<F>(self, op: F) -> RegionSubsts where F: FnOnce(VecPerParamSpace<Region>) -> VecPerParamSpace<Region>

Unstable (rustc_private)

fn is_erased(&self) -> bool

Unstable (rustc_private)