Struct rustc::middle::subst::Substs

pub struct Substs<'tcx> {
    pub types: VecPerParamSpace<Ty<'tcx>>,
    pub regions: RegionSubsts,
}

Unstable (rustc_private)

A substitution mapping type/region parameters to new values. We identify each in-scope parameter by an index and a parameter space (which indices where the parameter is defined; see ParamSpace).

Fields

types
Unstable (rustc_private)

regions
Unstable (rustc_private)

Methods

impl<'tcx> Substs<'tcx>

fn new(t: VecPerParamSpace<Ty<'tcx>>, r: VecPerParamSpace<Region>) -> Substs<'tcx>

Unstable (rustc_private)

fn new_type(t: Vec<Ty<'tcx>>, r: Vec<Region>) -> Substs<'tcx>

Unstable (rustc_private)

fn new_trait(t: Vec<Ty<'tcx>>, r: Vec<Region>, s: Ty<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

fn erased(t: VecPerParamSpace<Ty<'tcx>>) -> Substs<'tcx>

Unstable (rustc_private)

fn empty() -> Substs<'tcx>

Unstable (rustc_private)

fn trans_empty() -> Substs<'tcx>

Unstable (rustc_private)

fn is_noop(&self) -> bool

Unstable (rustc_private)

fn type_for_def(&self, ty_param_def: &TypeParameterDef) -> Ty<'tcx>

Unstable (rustc_private)

fn self_ty(&self) -> Option<Ty<'tcx>>

Unstable (rustc_private)

fn with_self_ty(&self, self_ty: Ty<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

fn erase_regions(self) -> Substs<'tcx>

Unstable (rustc_private)

fn regions<'a>(&'a self) -> &'a VecPerParamSpace<Region>

Unstable (rustc_private)

Since ErasedRegions are only to be used in trans, most of the compiler can use this method to easily access the set of region substitutions.

fn mut_regions<'a>(&'a mut self) -> &'a mut VecPerParamSpace<Region>

Unstable (rustc_private)

Since ErasedRegions are only to be used in trans, most of the compiler can use this method to easily access the set of region substitutions.

fn with_method(self, m_types: Vec<Ty<'tcx>>, m_regions: Vec<Region>) -> Substs<'tcx>

Unstable (rustc_private)

fn with_method_from(self, meth_substs: &Substs<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

fn to_trait_ref(&self, tcx: &ctxt<'tcx>, trait_id: DefId) -> TraitRef<'tcx>

Unstable (rustc_private)

Creates a trait-ref out of this substs, ignoring the FnSpace substs