Struct rustc::middle::subst::VecPerParamSpace

pub struct VecPerParamSpace<T> {
    // some fields omitted
}

Unstable (rustc_private)

Vector of things sorted by param space. Used to keep the set of things declared on the type, self, or method distinct.

Methods

impl<T> VecPerParamSpace<T>

fn empty() -> VecPerParamSpace<T>

Unstable (rustc_private)

fn params_from_type(types: Vec<T>) -> VecPerParamSpace<T>

Unstable (rustc_private)

fn new(t: Vec<T>, s: Vec<T>, f: Vec<T>) -> VecPerParamSpace<T>

Unstable (rustc_private)

t is the type space. s is the self space. f is the fn space.

fn push(&mut self, space: ParamSpace, value: T)

Unstable (rustc_private)

Appends value to the vector associated with space.

Unlike the push method in Vec, this should not be assumed to be a cheap operation (even when amortized over many calls).

fn extend<I: Iterator<Item=T>>(&mut self, space: ParamSpace, values: I)

Unstable (rustc_private)

Appends values to the vector associated with space.

Unlike the extend method in Vec, this should not be assumed to be a cheap operation (even when amortized over many calls).

fn pop(&mut self, space: ParamSpace) -> Option<T>

Unstable (rustc_private)

fn truncate(&mut self, space: ParamSpace, len: usize)

Unstable (rustc_private)

fn replace(&mut self, space: ParamSpace, elems: Vec<T>)

Unstable (rustc_private)

fn get_self<'a>(&'a self) -> Option<&'a T>

Unstable (rustc_private)

fn len(&self, space: ParamSpace) -> usize

Unstable (rustc_private)

fn is_empty_in(&self, space: ParamSpace) -> bool

Unstable (rustc_private)

fn get_slice<'a>(&'a self, space: ParamSpace) -> &'a [T]

Unstable (rustc_private)

fn get_mut_slice<'a>(&'a mut self, space: ParamSpace) -> &'a mut [T]

Unstable (rustc_private)

fn opt_get<'a>(&'a self, space: ParamSpace, index: usize) -> Option<&'a T>

Unstable (rustc_private)

fn get<'a>(&'a self, space: ParamSpace, index: usize) -> &'a T

Unstable (rustc_private)

fn iter<'a>(&'a self) -> Iter<'a, T>

Unstable (rustc_private)

fn into_iter(self) -> IntoIter<T>

Unstable (rustc_private)

fn iter_enumerated<'a>(&'a self) -> EnumeratedItems<'a, T>

Unstable (rustc_private)

fn as_slice(&self) -> &[T]

Unstable (rustc_private)

fn into_vec(self) -> Vec<T>

Unstable (rustc_private)

fn all_vecs<P>(&self, pred: P) -> bool where P: FnMut(&[T]) -> bool

Unstable (rustc_private)

fn all<P>(&self, pred: P) -> bool where P: FnMut(&T) -> bool

Unstable (rustc_private)

fn any<P>(&self, pred: P) -> bool where P: FnMut(&T) -> bool

Unstable (rustc_private)

fn is_empty(&self) -> bool

Unstable (rustc_private)

fn map<U, P>(&self, pred: P) -> VecPerParamSpace<U> where P: FnMut(&T) -> U

Unstable (rustc_private)

fn map_enumerated<U, P>(&self, pred: P) -> VecPerParamSpace<U> where P: FnMut((ParamSpace, usize, &T)) -> U

Unstable (rustc_private)

fn split(self) -> SeparateVecsPerParamSpace<T>

Unstable (rustc_private)

fn with_vec(self, space: ParamSpace, vec: Vec<T>) -> VecPerParamSpace<T>

Unstable (rustc_private)