Enum rustc::middle::traits::ObligationCauseCode

pub enum ObligationCauseCode<'tcx> {
    MiscObligation,
    ItemObligation(DefId),
    ObjectCastObligation(Ty<'tcx>),
    AssignmentLhsSized,
    StructInitializerSized,
    VariableType(NodeId),
    ReturnType,
    RepeatVec,
    ClosureCapture(NodeId, Span, BuiltinBound),
    FieldSized,
    SharedStatic,
    BuiltinDerivedObligation(DerivedObligationCause<'tcx>),
    ImplDerivedObligation(DerivedObligationCause<'tcx>),
    CompareImplMethodObligation,
}

Unstable

Variants

MiscObligation
Unstable

Not well classified or should be obvious from span.

ItemObligation
Unstable

In an impl of trait X for type Y, type Y must also implement all supertraits of X.

ObjectCastObligation
Unstable

Obligation incurred due to an object cast.

AssignmentLhsSized
Unstable

Various cases where expressions must be sized/copy/etc:

StructInitializerSized
Unstable

VariableType
Unstable

ReturnType
Unstable

RepeatVec
Unstable

ClosureCapture
Unstable

FieldSized
Unstable

SharedStatic
Unstable

BuiltinDerivedObligation
Unstable

ImplDerivedObligation
Unstable

CompareImplMethodObligation
Unstable