Function rustc::middle::traits::evaluate_builtin_bound

pub fn evaluate_builtin_bound<'a, 'tcx>(infcx: &InferCtxt<'a, 'tcx>, typer: &ClosureTyper<'tcx>, ty: Ty<'tcx>, bound: BuiltinBound, span: Span) -> SelectionResult<'tcx, ()>

Unstable

Determines whether the type ty is known to meet bound and returns true if so. Returns false if ty either does not meet bound or is not known to meet bound (note that this is conservative towards no impl, which is the opposite of the evaluate methods).