Struct rustc::middle::traits::FulfillmentContext

pub struct FulfillmentContext<'tcx> {
    // some fields omitted
}

Unstable (rustc_private)

The fulfillment context is used to drive trait resolution. It consists of a list of obligations that must be (eventually) satisfied. The job is to track which are satisfied, which yielded errors, and which are still pending. At any point, users can call select_where_possible, and the fulfilment context will try to do selection, retaining only those obligations that remain ambiguous. This may be helpful in pushing type inference along. Once all type inference constraints have been generated, the method select_all_or_error can be used to report any remaining ambiguous cases as errors.

Methods

impl<'tcx> FulfillmentContext<'tcx>

fn new() -> FulfillmentContext<'tcx>

Unstable (rustc_private)

Creates a new fulfillment context.

fn normalize_projection_type<'a>(&mut self, infcx: &InferCtxt<'a, 'tcx>, projection_ty: ProjectionTy<'tcx>, cause: ObligationCause<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

"Normalize" a projection type <SomeType as SomeTrait>::X by creating a fresh type variable $0 as well as a projection predicate <SomeType as SomeTrait>::X == $0. When the inference engine runs, it will attempt to find an impl of SomeTrait or a where clause that lets us unify $0 with something concrete. If this fails, we'll unify $0 with projection_ty again.

fn register_builtin_bound<'a>(&mut self, infcx: &InferCtxt<'a, 'tcx>, ty: Ty<'tcx>, builtin_bound: BuiltinBound, cause: ObligationCause<'tcx>)

Unstable (rustc_private)

fn register_region_obligation<'a>(&mut self, t_a: Ty<'tcx>, r_b: Region, cause: ObligationCause<'tcx>)

Unstable (rustc_private)

fn register_predicate_obligation<'a>(&mut self, infcx: &InferCtxt<'a, 'tcx>, obligation: PredicateObligation<'tcx>)

Unstable (rustc_private)

fn region_obligations(&self, body_id: NodeId) -> &[RegionObligation<'tcx>]

Unstable (rustc_private)

fn select_all_or_error<'a>(&mut self, infcx: &InferCtxt<'a, 'tcx>) -> Result<(), Vec<FulfillmentError<'tcx>>>

Unstable (rustc_private)

fn select_where_possible<'a>(&mut self, infcx: &InferCtxt<'a, 'tcx>) -> Result<(), Vec<FulfillmentError<'tcx>>>

Unstable (rustc_private)

fn pending_obligations(&self) -> Vec<PendingPredicateObligation<'tcx>>

Unstable (rustc_private)