Module rustc::middle::ty

Unstable (rustc_private)

Reexports

pub use self::InferTy::*;

pub use self::ImplOrTraitItemId::*;

pub use self::ClosureKind::*;

pub use self::Variance::*;

pub use self::AutoAdjustment::*;

pub use self::Representability::*;

pub use self::AutoRef::*;

pub use self::DtorKind::*;

pub use self::ExplicitSelfCategory::*;

pub use self::FnOutput::*;

pub use self::Region::*;

pub use self::ImplOrTraitItemContainer::*;

pub use self::BorrowKind::*;

pub use self::ImplOrTraitItem::*;

pub use self::BoundRegion::*;

pub use self::TypeVariants::*;

pub use self::IntVarValue::*;

pub use self::CopyImplementationError::*;

pub use self::BuiltinBound::Send as BoundSend;

pub use self::BuiltinBound::Sized as BoundSized;

pub use self::BuiltinBound::Copy as BoundCopy;

pub use self::BuiltinBound::Sync as BoundSync;

Modules

tls 	 [Unstable]

Structs

AdtDefData 	 [Unstable]
The definition of an abstract data type - a struct or enum.

AdtFlags 	 [Unstable]

AssociatedConst 	 [Unstable]

AssociatedType 	 [Unstable]

AutoDerefRef 	 [Unstable]
Represents coercing a pointer to a different kind of pointer - where 'kind' here means either or both of raw vs borrowed vs unique and fat vs thin.

BareFnTy 	 [Unstable]

Binder 	 [Unstable]
Binder is a binder for higher-ranked lifetimes. It is part of the compiler's representation for things like for<'a> Fn(&'a isize) (which would be represented by the type PolyTraitRef == Binder<TraitRef>). Note that when we skolemize, instantiate, erase, or otherwise "discharge" these bound regions, we change the type from Binder<T> to just T (see e.g. liberate_late_bound_regions).

BuiltinBounds 	 [Unstable]

CReaderCacheKey 	 [Unstable]

ClosureSubsts 	 [Unstable]
A closure can be modeled as a struct that looks like:

ClosureTy 	 [Unstable]

ClosureUpvar 	 [Unstable]

CommonTypes 	 [Unstable]

CrateAnalysis 	 [Unstable]
The complete set of all analyses described in this module. This is produced by the driver and fed to trans and later passes.

CtxtArenas 	 [Unstable]
Internal storage

DebruijnIndex 	 [Unstable]
A [De Bruijn index][dbi] is a standard means of representing regions (and perhaps later types) in a higher-ranked setting. In particular, imagine a type like this:

EarlyBoundRegion 	 [Unstable]

EquatePredicate 	 [Unstable]

ExistentialBounds 	 [Unstable]
Bounds suitable for an existentially quantified type parameter such as those that appear in object types or closure types.

ExpectedFound 	 [Unstable]

FieldDefData 	 [Unstable]

FloatVid 	 [Unstable]

FnSig 	 [Unstable]
Signature of a function type, which I have arbitrarily decided to use to refer to the input/output types.

FreeRegion 	 [Unstable]
A "free" region fr can be interpreted as "some region at least as big as the scope fr.scope".

Freevar 	 [Unstable]
A free variable referred to in a function.

GenericPredicates 	 [Unstable]
Bounds on generics.

Generics 	 [Unstable]
Information about the formal type/lifetime parameters associated with an item or method. Analogous to ast::Generics.

InstantiatedPredicates 	 [Unstable]
Represents the bounds declared on a particular set of type parameters. Should eventually be generalized into a flag list of where clauses. You can obtain a InstantiatedPredicates list from a GenericPredicates by using the instantiate method. Note that this method reflects an important semantic invariant of InstantiatedPredicates: while the GenericPredicates are expressed in terms of the bound type parameters of the impl/trait/whatever, an InstantiatedPredicates instance represented a set of bounds for some particular instantiation, meaning that the generic parameters have been substituted with their values.

IntVid 	 [Unstable]

InternedTy 	 [Unstable]
An entry in the type interner.

ItemSubsts 	 [Unstable]
Records the substitutions used to translate the polytype for an item into the monotype of an item reference.

ItemVariances 	 [Unstable]

Method 	 [Unstable]

MethodCall 	 [Unstable]
With method calls, we store some extra information in side tables (i.e method_map). We use MethodCall as a key to index into these tables instead of just directly using the expression's NodeId. The reason for this being that we may apply adjustments (coercions) with the resulting expression also needing to use the side tables. The problem with this is that we don't assign a separate NodeId to this new expression and so it would clash with the base expression if both needed to add to the side tables. Thus to disambiguate we also keep track of whether there's an adjustment in our key.

MethodCallee 	 [Unstable]

OutlivesPredicate 	 [Unstable]

ParamTy 	 [Unstable]

ParameterEnvironment 	 [Unstable]
When type checking, we use the ParameterEnvironment to track details about the type/lifetime parameters that are in scope. It primarily stores the bounds information.

ProjectionPredicate 	 [Unstable]
This kind of predicate has no direct correspondent in the syntax, but it roughly corresponds to the syntactic forms:

ProjectionTy 	 [Unstable]
Represents the projection of an associated type. In explicit UFCS form this would be written <T as Trait<..>>::N.

RegionParameterDef 	 [Unstable]

RegionVid 	 [Unstable]

SkolemizedRegionVid 	 [Unstable]

Tables 	 [Unstable]

TraitDef 	 [Unstable]
As TypeScheme but for a trait ref.

TraitFlags 	 [Unstable]

TraitPredicate 	 [Unstable]

TraitRef 	 [Unstable]
A complete reference to a trait. These take numerous guises in syntax, but perhaps the most recognizable form is in a where clause:

TraitTy 	 [Unstable]

TransmuteRestriction 	 [Unstable]
A restriction that certain types must be the same size. The use of transmute gives rise to these restrictions. These generally cannot be checked until trans; therefore, each call to transmute will push one or more such restriction into the transmute_restrictions vector during intrinsicck. They are then checked during trans by the fn check_intrinsics.

TyIVar 	 [Unstable]
An IVar that contains a Ty. 'lt is a (reverse-variant) upper bound on the lifetime of the IVar. This is required because of variance problems: the IVar needs to be variant with respect to 'tcx (so it can be referred to from Ty) but can only be modified if its lifetime is exactly 'tcx.

TyS 	 [Unstable]

TyVid 	 [Unstable]

TypeAndMut 	 [Unstable]

TypeContents 	 [Unstable]
Type contents is how the type checker reasons about kinds. They track what kinds of things are found within a type. You can think of them as kind of an "anti-kind". They track the kinds of values and thinks that are contained in types. Having a larger contents for a type tends to rule that type out from various kinds. For example, a type that contains a reference is not sendable.

TypeFlags 	 [Unstable]

TypeParameterDef 	 [Unstable]

TypeScheme 	 [Unstable]
A "type scheme", in ML terminology, is a type combined with some set of generic types that the type is, well, generic over. In Rust terms, it is the "type" of a fn item or struct -- this type will include various generic parameters that must be substituted when the item/struct is referenced. That is called converting the type scheme to a monotype.

UpvarBorrow 	 [Unstable]

UpvarId 	 [Unstable]
Upvars do not get their own node-id. Instead, we use the pair of the original var id (that is, the root variable that is referenced by the upvar) and the id of the closure expression.

VariantDefData 	 [Unstable]

ctxt 	 [Unstable]
The data structure to keep track of all the information that typechecker generates so that so that it can be reused and doesn't have to be redone later on.

Enums

AdtKind 	 [Unstable]

AutoAdjustment 	 [Unstable]

AutoRef 	 [Unstable]

BorrowKind 	 [Unstable]

BoundRegion 	 [Unstable]

BuiltinBound 	 [Unstable]

ClosureKind 	 [Unstable]

CopyImplementationError 	 [Unstable]

CustomCoerceUnsized 	 [Unstable]

DtorKind 	 [Unstable]

ExplicitSelfCategory 	 [Unstable]
The category of explicit self.

FnOutput 	 [Unstable]

FragmentInfo 	 [Unstable]
Describes the fragment-state associated with a NodeId.

ImplOrTraitItem 	 [Unstable]

ImplOrTraitItemContainer 	 [Unstable]

ImplOrTraitItemId 	 [Unstable]

InferTy 	 [Unstable]

IntVarValue 	 [Unstable]

ObjectLifetimeDefault 	 [Unstable]
Default region to use for the bound of objects that are supplied as the value for this type parameter. This is derived from T:'a annotations appearing in the type definition. If this is None, then the default is inherited from the surrounding context. See RFC #599 for details.

Predicate 	 [Unstable]

Region 	 [Unstable]
Representation of regions.

Representability 	 [Unstable]
Describes whether a type is representable. For types that are not representable, 'SelfRecursive' and 'ContainsRecursive' are used to distinguish between types that are recursive with themselves and types that contain a different recursive type. These cases can therefore be treated differently when reporting errors.

TypeError 	 [Unstable]

TypeVariants 	 [Unstable]

UnconstrainedNumeric 	 [Unstable]

UpvarCapture 	 [Unstable]
Information describing the capture of an upvar. This is computed during typeck, specifically by regionck.

Variance 	 [Unstable]

VariantKind 	 [Unstable]

Constants

INITIAL_DISCRIMINANT_VALUE 	 [Unstable]

Traits

HasTypeFlags 	 [Unstable]

IntTypeExt 	 [Unstable]

Lift 	 [Unstable]
A trait implemented for all X<'a> types which can be safely and efficiently converted to X<'tcx> as long as they are part of the provided ty::ctxt<'tcx>. This can be done, for example, for Ty<'tcx> or &'tcx Substs<'tcx> by looking them up in their respective interners. None is returned if the value or one of the components is not part of the provided context. For Ty, None can be returned if either the type interner doesn't contain the TypeVariants key or if the address of the interned pointer differs. The latter case is possible if a primitive type, e.g. () or u8, was interned in a different context.

RegionEscape 	 [Unstable]
An "escaping region" is a bound region whose binder is not part of t.

ToPolyTraitRef 	 [Unstable]

ToPredicate 	 [Unstable]

Functions

sort_bounds_list 	 [Unstable]

Type Definitions

AdtDef 	 [Unstable]

AdtDefMaster 	 [Unstable]

CaptureModeMap 	 [Unstable]

Disr 	 [Unstable]

FieldDef 	 [Unstable]

FieldDefMaster 	 [Unstable]

FreevarMap 	 [Unstable]

GlobMap 	 [Unstable]

MethodMap 	 [Unstable]

PolyEquatePredicate 	 [Unstable]

PolyFnOutput 	 [Unstable]

PolyFnSig 	 [Unstable]

PolyOutlivesPredicate 	 [Unstable]

PolyProjectionPredicate 	 [Unstable]

PolyRegionOutlivesPredicate 	 [Unstable]

PolyTraitPredicate 	 [Unstable]

PolyTraitRef 	 [Unstable]

PolyTypeOutlivesPredicate 	 [Unstable]

TraitMap 	 [Unstable]

Ty 	 [Unstable]

UpvarCaptureMap 	 [Unstable]

VariantDef 	 [Unstable]

VariantDefMaster 	 [Unstable]