Enum rustc::middle::ty::BoundRegionUnstable

pub enum BoundRegion {
    BrAnon(u32),
    BrNamed(DefId, Name),
    BrFresh(u32),
    BrEnv,
}

Variants

BrAnon

An anonymous region parameter for a given fn (&T)

BrNamed

Named region parameters for functions (a in &'a T)

The def-id is needed to distinguish free regions in the event of shadowing.

BrFresh

Fresh bound identifiers created during GLB computations.

BrEnv