Enum rustc::middle::ty::Predicate

pub enum Predicate<'tcx> {
    Trait(PolyTraitPredicate<'tcx>),
    Equate(PolyEquatePredicate<'tcx>),
    RegionOutlives(PolyRegionOutlivesPredicate),
    TypeOutlives(PolyTypeOutlivesPredicate<'tcx>),
    Projection(PolyProjectionPredicate<'tcx>),
}

Unstable

Variants

Trait
Unstable

Corresponds to where Foo : Bar<A,B,C>. Foo here would be the Self type of the trait reference and A, B, and C would be the parameters in the TypeSpace.

Equate
Unstable

where T1 == T2.

RegionOutlives
Unstable

where 'a : 'b

TypeOutlives
Unstable

where T : 'a

Projection
Unstable

where ::Name == X, approximately. See ProjectionPredicate struct for details.

Methods

impl<'tcx> Predicate<'tcx>

fn subst_supertrait(&self, tcx: &ctxt<'tcx>, trait_ref: &PolyTraitRef<'tcx>) -> Predicate<'tcx>

Unstable

Performs a substitution suitable for going from a poly-trait-ref to supertraits that must hold if that poly-trait-ref holds. This is slightly different from a normal substitution in terms of what happens with bound regions. See lengthy comment below for details.

impl<'tcx> Predicate<'tcx>

fn walk_tys(&self) -> IntoIter<Ty<'tcx>>

Unstable

Iterates over the types in this predicate. Note that in all cases this is skipping over a binder, so late-bound regions with depth 0 are bound by the predicate.

fn has_escaping_regions(&self) -> bool

Unstable

fn to_opt_poly_trait_ref(&self) -> Option<PolyTraitRef<'tcx>>

Unstable