Enum rustc::middle::ty::UpvarCapture

pub enum UpvarCapture {
    ByValue,
    ByRef(UpvarBorrow),
}

Unstable

Information describing the capture of an upvar. This is computed during typeck, specifically by regionck.

Variants

ByValue
Unstable

Upvar is captured by value. This is always true when the closure is labeled move, but can also be true in other cases depending on inference.

ByRef
Unstable

Upvar is captured by reference.