Function rustc::middle::ty::erase_late_bound_regions

pub fn erase_late_bound_regions<'tcx, T>(tcx: &ctxt<'tcx>, value: &Binder<T>) -> T where T: TypeFoldable<'tcx>

Unstable

Replace any late-bound regions bound in value with 'static. Useful in trans but also method lookup and a few other places where precise region relationships are not required.