Function rustc::middle::ty::with_ctxt

pub fn with_ctxt<'tcx, F, R>(s: Session, arenas: &'tcx CtxtArenas<'tcx>, def_map: DefMap, named_region_map: NamedRegionMap, map: Map<'tcx>, freevars: RefCell<FreevarMap>, region_maps: RegionMaps, lang_items: LanguageItems, stability: Index<'tcx>, f: F) -> (Session, R) where F: FnOnce(&ctxt<'tcx>) -> R

Unstable

Create a type context and call the closure with a &ty::ctxt reference to the context. The closure enforces that the type context and any interned value (types, substs, etc.) can only be used while ty::tls has a valid reference to the context, to allow formatting values that need it.