Trait rustc::middle::ty::fold::TypeFolder

pub trait TypeFolder<'tcx>: Sized {
    fn tcx<'a>(&'a self) -> &'a ctxt<'tcx>;

    fn enter_region_binder(&mut self) { ... }
    fn exit_region_binder(&mut self) { ... }
    fn fold_binder<T>(&mut self, t: &Binder<T>) -> Binder<T> where T: TypeFoldable<'tcx> { ... }
    fn fold_ty(&mut self, t: Ty<'tcx>) -> Ty<'tcx> { ... }
    fn fold_mt(&mut self, t: &TypeAndMut<'tcx>) -> TypeAndMut<'tcx> { ... }
    fn fold_trait_ref(&mut self, t: &TraitRef<'tcx>) -> TraitRef<'tcx> { ... }
    fn fold_substs(&mut self, substs: &Substs<'tcx>) -> Substs<'tcx> { ... }
    fn fold_fn_sig(&mut self, sig: &FnSig<'tcx>) -> FnSig<'tcx> { ... }
    fn fold_output(&mut self, output: &FnOutput<'tcx>) -> FnOutput<'tcx> { ... }
    fn fold_bare_fn_ty(&mut self, fty: &BareFnTy<'tcx>) -> BareFnTy<'tcx> { ... }
    fn fold_closure_ty(&mut self, fty: &ClosureTy<'tcx>) -> ClosureTy<'tcx> { ... }
    fn fold_region(&mut self, r: Region) -> Region { ... }
    fn fold_existential_bounds(&mut self, s: &ExistentialBounds<'tcx>) -> ExistentialBounds<'tcx> { ... }
    fn fold_autoref(&mut self, ar: &AutoRef<'tcx>) -> AutoRef<'tcx> { ... }
    fn fold_item_substs(&mut self, i: ItemSubsts<'tcx>) -> ItemSubsts<'tcx> { ... }
}

Unstable (rustc_private)

The TypeFolder trait defines the actual folding. There is a method defined for every foldable type. Each of these has a default implementation that does an "identity" fold. Within each identity fold, it should invoke foo.fold_with(self) to fold each sub-item.

Required Methods

fn tcx<'a>(&'a self) -> &'a ctxt<'tcx>

Unstable (rustc_private)

Provided Methods

fn enter_region_binder(&mut self)

Unstable (rustc_private)

Invoked by the super_* routines when we enter a region binding level (for example, when entering a function signature). This is used by clients that want to track the Debruijn index nesting level.

fn exit_region_binder(&mut self)

Unstable (rustc_private)

Invoked by the super_* routines when we exit a region binding level. This is used by clients that want to track the Debruijn index nesting level.

fn fold_binder<T>(&mut self, t: &Binder<T>) -> Binder<T> where T: TypeFoldable<'tcx>

Unstable (rustc_private)

fn fold_ty(&mut self, t: Ty<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn fold_mt(&mut self, t: &TypeAndMut<'tcx>) -> TypeAndMut<'tcx>

Unstable (rustc_private)

fn fold_trait_ref(&mut self, t: &TraitRef<'tcx>) -> TraitRef<'tcx>

Unstable (rustc_private)

fn fold_substs(&mut self, substs: &Substs<'tcx>) -> Substs<'tcx>

Unstable (rustc_private)

fn fold_fn_sig(&mut self, sig: &FnSig<'tcx>) -> FnSig<'tcx>

Unstable (rustc_private)

fn fold_output(&mut self, output: &FnOutput<'tcx>) -> FnOutput<'tcx>

Unstable (rustc_private)

fn fold_bare_fn_ty(&mut self, fty: &BareFnTy<'tcx>) -> BareFnTy<'tcx>

Unstable (rustc_private)

fn fold_closure_ty(&mut self, fty: &ClosureTy<'tcx>) -> ClosureTy<'tcx>

Unstable (rustc_private)

fn fold_region(&mut self, r: Region) -> Region

Unstable (rustc_private)

fn fold_existential_bounds(&mut self, s: &ExistentialBounds<'tcx>) -> ExistentialBounds<'tcx>

Unstable (rustc_private)

fn fold_autoref(&mut self, ar: &AutoRef<'tcx>) -> AutoRef<'tcx>

Unstable (rustc_private)

fn fold_item_substs(&mut self, i: ItemSubsts<'tcx>) -> ItemSubsts<'tcx>

Unstable (rustc_private)