Struct rustc::middle::ty::MethodParam

pub struct MethodParam<'tcx> {
    pub trait_ref: TraitRef<'tcx>,
    pub method_num: usize,
    pub impl_def_id: Option<DefId>,
}

Unstable

Fields

trait_ref
Unstable

method_num
Unstable

impl_def_id
Unstable

The impl for the trait from which the method comes. This should only be used for certain linting/heuristic purposes since there is no guarantee that this is Some in every situation that it could/should be.