Struct rustc::middle::ty::TraitDef

pub struct TraitDef<'tcx> {
    pub unsafety: Unsafety,
    pub paren_sugar: bool,
    pub generics: Generics<'tcx>,
    pub trait_ref: TraitRef<'tcx>,
    pub associated_type_names: Vec<Name>,
    pub nonblanket_impls: RefCell<FnvHashMap<SimplifiedType, Vec<DefId>>>,
    pub blanket_impls: RefCell<Vec<DefId>>,
    pub flags: Cell<TraitFlags>,
}

Unstable (rustc_private)

As TypeScheme but for a trait ref.

Fields

unsafety
Unstable (rustc_private)

paren_sugar
Unstable (rustc_private)

If true, then this trait had the #[rustc_paren_sugar] attribute, indicating that it should be used with Foo() sugar. This is a temporary thing -- eventually any trait wil be usable with the sugar (or without it).

generics
Unstable (rustc_private)

Generic type definitions. Note that Self is listed in here as having a single bound, the trait itself (e.g., in the trait Eq, there is a single bound Self : Eq). This is so that default methods get to assume that the Self parameters implements the trait.

trait_ref
Unstable (rustc_private)

associated_type_names
Unstable (rustc_private)

A list of the associated types defined in this trait. Useful for resolving X::Foo type markers.

nonblanket_impls
Unstable (rustc_private)

Impls of the trait.

blanket_impls
Unstable (rustc_private)

Blanket impls associated with the trait.

flags
Unstable (rustc_private)

Various flags

Methods

impl<'tcx> TraitDef<'tcx>

fn object_safety(&self) -> Option<bool>

Unstable (rustc_private)

fn set_object_safety(&self, is_safe: bool)

Unstable (rustc_private)

fn record_impl(&self, tcx: &ctxt<'tcx>, impl_def_id: DefId, impl_trait_ref: TraitRef<'tcx>)

Unstable (rustc_private)

Records a trait-to-implementation mapping.

fn for_each_impl<F: FnMut(DefId)>(&self, tcx: &ctxt<'tcx>, f: F)

Unstable (rustc_private)

fn for_each_relevant_impl<F: FnMut(DefId)>(&self, tcx: &ctxt<'tcx>, self_ty: Ty<'tcx>, f: F)

Unstable (rustc_private)

Iterate over every impl that could possibly match the self-type self_ty.