Struct rustc::middle::ty::TraitRef

pub struct TraitRef<'tcx> {
    pub def_id: DefId,
    pub substs: &'tcx Substs<'tcx>,
}

Unstable (rustc_private)

A complete reference to a trait. These take numerous guises in syntax, but perhaps the most recognizable form is in a where clause:

T : Foo<U>

This would be represented by a trait-reference where the def-id is the def-id for the trait Foo and the substs defines T as parameter 0 in the SelfSpace and U as parameter 0 in the TypeSpace.

Trait references also appear in object types like Foo<U>, but in that case the Self parameter is absent from the substitutions.

Note that a TraitRef introduces a level of region binding, to account for higher-ranked trait bounds like T : for<'a> Foo<&'a U> or higher-ranked object types.

Fields

def_id
Unstable (rustc_private)

substs
Unstable (rustc_private)

Methods

impl<'tcx> TraitRef<'tcx>

fn new(def_id: DefId, substs: &'tcx Substs<'tcx>) -> TraitRef<'tcx>

Unstable (rustc_private)

fn self_ty(&self) -> Ty<'tcx>

Unstable (rustc_private)

fn input_types(&self) -> &[Ty<'tcx>]

Unstable (rustc_private)