Struct rustc::middle::ty::TyS

pub struct TyS<'tcx> {
    pub sty: TypeVariants<'tcx>,
    pub flags: Cell<TypeFlags>,
    // some fields omitted
}

Unstable

Fields

sty
Unstable

flags
Unstable

Methods

impl<'tcx> TyS<'tcx>

fn walk(&'tcx self) -> TypeWalker<'tcx>

Unstable

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

isize => { isize }
Foo<Bar<isize>> => { Foo<Bar<isize>>, Bar<isize>, isize }
[isize] => { [isize], isize }

fn walk_shallow(&'tcx self) -> IntoIter<Ty<'tcx>>

Unstable

Iterator that walks the immediate children of self. Hence Foo<Bar<i32>, u32> yields the sequence [Bar<i32>, u32] (but not i32, like walk).

fn as_opt_param_ty(&self) -> Option<ParamTy>

Unstable

fn is_param(&self, space: ParamSpace, index: u32) -> bool

Unstable