Struct rustc::middle::ty::ctxt

pub struct ctxt<'tcx> {
    pub types: CommonTypes<'tcx>,
    pub sess: Session,
    pub def_map: DefMap,
    pub named_region_map: NamedRegionMap,
    pub region_maps: RegionMaps,
    pub tables: RefCell<Tables<'tcx>>,
    pub impl_or_trait_items: RefCell<DefIdMap<ImplOrTraitItem<'tcx>>>,
    pub trait_item_def_ids: RefCell<DefIdMap<Rc<Vec<ImplOrTraitItemId>>>>,
    pub trait_items_cache: RefCell<DefIdMap<Rc<Vec<ImplOrTraitItem<'tcx>>>>>,
    pub impl_trait_refs: RefCell<DefIdMap<Option<TraitRef<'tcx>>>>,
    pub trait_defs: RefCell<DefIdMap<&'tcx TraitDef<'tcx>>>,
    pub adt_defs: RefCell<DefIdMap<AdtDefMaster<'tcx>>>,
    pub predicates: RefCell<DefIdMap<GenericPredicates<'tcx>>>,
    pub super_predicates: RefCell<DefIdMap<GenericPredicates<'tcx>>>,
    pub map: Map<'tcx>,
    pub freevars: RefCell<FreevarMap>,
    pub tcache: RefCell<DefIdMap<TypeScheme<'tcx>>>,
    pub rcache: RefCell<FnvHashMap<CReaderCacheKey, Ty<'tcx>>>,
    pub tc_cache: RefCell<FnvHashMap<Ty<'tcx>, TypeContents>>,
    pub ast_ty_to_ty_cache: RefCell<NodeMap<Ty<'tcx>>>,
    pub ty_param_defs: RefCell<NodeMap<TypeParameterDef<'tcx>>>,
    pub normalized_cache: RefCell<FnvHashMap<Ty<'tcx>, Ty<'tcx>>>,
    pub lang_items: LanguageItems,
    pub provided_method_sources: RefCell<DefIdMap<DefId>>,
    pub item_variance_map: RefCell<DefIdMap<Rc<ItemVariances>>>,
    pub variance_computed: Cell<bool>,
    pub destructor_for_type: RefCell<DefIdMap<DefId>>,
    pub destructors: RefCell<DefIdSet>,
    pub inherent_impls: RefCell<DefIdMap<Rc<Vec<DefId>>>>,
    pub impl_items: RefCell<DefIdMap<Vec<ImplOrTraitItemId>>>,
    pub used_unsafe: RefCell<NodeSet>,
    pub used_mut_nodes: RefCell<NodeSet>,
    pub populated_external_types: RefCell<DefIdSet>,
    pub populated_external_primitive_impls: RefCell<DefIdSet>,
    pub extern_const_statics: RefCell<DefIdMap<NodeId>>,
    pub extern_const_variants: RefCell<DefIdMap<NodeId>>,
    pub extern_const_fns: RefCell<DefIdMap<NodeId>>,
    pub dependency_formats: RefCell<Dependencies>,
    pub node_lint_levels: RefCell<FnvHashMap<(NodeId, LintId), LevelSource>>,
    pub transmute_restrictions: RefCell<Vec<TransmuteRestriction<'tcx>>>,
    pub stability: RefCell<Index<'tcx>>,
    pub selection_cache: SelectionCache<'tcx>,
    pub fulfilled_predicates: RefCell<FulfilledPredicates<'tcx>>,
    pub repr_hint_cache: RefCell<DefIdMap<Rc<Vec<ReprAttr>>>>,
    pub const_qualif_map: RefCell<NodeMap<ConstQualif>>,
    pub custom_coerce_unsized_kinds: RefCell<DefIdMap<CustomCoerceUnsized>>,
    pub cast_kinds: RefCell<NodeMap<CastKind>>,
    pub fragment_infos: RefCell<DefIdMap<Vec<FragmentInfo>>>,
    // some fields omitted
}

Unstable

The data structure to keep track of all the information that typechecker generates so that so that it can be reused and doesn't have to be redone later on.

Fields

types
Unstable

Common types, pre-interned for your convenience.

sess
Unstable

def_map
Unstable

named_region_map
Unstable

region_maps
Unstable

tables
Unstable

impl_or_trait_items
Unstable

Maps from a trait item to the trait item "descriptor"

trait_item_def_ids
Unstable

Maps from a trait def-id to a list of the def-ids of its trait items

trait_items_cache
Unstable

A cache for the trait_items() routine

impl_trait_refs
Unstable

trait_defs
Unstable

adt_defs
Unstable

predicates
Unstable

Maps from the def-id of an item (trait/struct/enum/fn) to its associated predicates.

super_predicates
Unstable

Maps from the def-id of a trait to the list of super-predicates. This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that supertraits have additional acyclicity requirements).

map
Unstable

freevars
Unstable

tcache
Unstable

rcache
Unstable

tc_cache
Unstable

ast_ty_to_ty_cache
Unstable

ty_param_defs
Unstable

normalized_cache
Unstable

lang_items
Unstable

provided_method_sources
Unstable

A mapping of fake provided method def_ids to the default implementation

item_variance_map
Unstable

Maps from def-id of a type or region parameter to its (inferred) variance.

variance_computed
Unstable

True if the variance has been computed yet; false otherwise.

destructor_for_type
Unstable

A mapping from the def ID of an enum or struct type to the def ID of the method that implements its destructor. If the type is not present in this map, it does not have a destructor. This map is populated during the coherence phase of typechecking.

destructors
Unstable

A method will be in this list if and only if it is a destructor.

inherent_impls
Unstable

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don't need to be exported.

impl_items
Unstable

Maps a DefId of an impl to a list of its items. Note that this contains all of the impls that we know about, including ones in other crates. It's not clear that this is the best way to do it.

used_unsafe
Unstable

Set of used unsafe nodes (functions or blocks). Unsafe nodes not present in this set can be warned about.

used_mut_nodes
Unstable

Set of nodes which mark locals as mutable which end up getting used at some point. Local variable definitions not in this set can be warned about.

populated_external_types
Unstable

The set of external nominal types whose implementations have been read. This is used for lazy resolution of methods.

populated_external_primitive_impls
Unstable

The set of external primitive types whose implementations have been read. FIXME(arielb1): why is this separate from populated_external_types?

extern_const_statics
Unstable

These caches are used by const_eval when decoding external constants.

extern_const_variants
Unstable

extern_const_fns
Unstable

dependency_formats
Unstable

node_lint_levels
Unstable

transmute_restrictions
Unstable

The types that must be asserted to be the same size for transmute to be valid. We gather up these restrictions in the intrinsicck pass and check them in trans.

stability
Unstable

Maps any item's def-id to its stability index.

selection_cache
Unstable

Caches the results of trait selection. This cache is used for things that do not have to do with the parameters in scope.

fulfilled_predicates
Unstable

A set of predicates that have been fulfilled somewhere. This is used to avoid duplicate work. Predicates are only added to this set when they mention only "global" names (i.e., no type or lifetime parameters).

repr_hint_cache
Unstable

Caches the representation hints for struct definitions.

const_qualif_map
Unstable

Maps Expr NodeId's to their constant qualification.

custom_coerce_unsized_kinds
Unstable

Caches CoerceUnsized kinds for impls on custom types.

cast_kinds
Unstable

Maps a cast expression to its kind. This is keyed on the from expression of the cast, not the cast itself.

fragment_infos
Unstable

Maps Fn items to a collection of fragment infos.

The main goal is to identify data (each of which may be moved or assigned) whose subparts are not moved nor assigned (i.e. their state is unfragmented) and corresponding ast nodes where the path to that data is moved or assigned.

In the long term, unfragmented values will have their destructor entirely driven by a single stack-local drop-flag, and their parents, the collections of the unfragmented values (or more simply, "fragmented values"), are mapped to the corresponding collections of stack-local drop-flags.

(However, in the short term that is not the case; e.g. some unfragmented paths still need to be zeroed, namely when they reference parent data from an outer scope that was not entirely moved, and therefore that needs to be zeroed so that we do not get double-drop when we hit the end of the parent scope.)

Also: currently the table solely holds keys for node-ids of unfragmented values (see FragmentInfo enum definition), but longer-term we will need to also store mappings from fragmented data to the set of unfragmented pieces that constitute it.

Methods

impl<'tcx> ctxt<'tcx>

fn note_and_explain_region(&self, prefix: &str, region: Region, suffix: &str)

Unstable

impl<'tcx> ctxt<'tcx>

fn node_types(&self) -> Ref<NodeMap<Ty<'tcx>>>

Unstable

fn node_type_insert(&self, id: NodeId, ty: Ty<'tcx>)

Unstable

fn intern_trait_def(&self, def: TraitDef<'tcx>) -> &'tcx TraitDef<'tcx>

Unstable

fn intern_adt_def(&self, did: DefId, kind: AdtKind, variants: Vec<VariantDefData<'tcx, 'tcx>>) -> AdtDefMaster<'tcx>

Unstable

fn intern_stability(&self, stab: Stability) -> &'tcx Stability

Unstable

fn store_free_region_map(&self, id: NodeId, map: FreeRegionMap)

Unstable

fn free_region_map(&self, id: NodeId) -> FreeRegionMap

Unstable

fn lift<T: ?Sized + Lift<'tcx>>(&self, value: &T) -> Option<T::Lifted>

Unstable

impl<'tcx> ctxt<'tcx>

fn print_debug_stats(&self)

Unstable

impl<'tcx> ctxt<'tcx>

fn create_and_enter<F, R>(s: Session, arenas: &'tcx CtxtArenas<'tcx>, def_map: DefMap, named_region_map: NamedRegionMap, map: Map<'tcx>, freevars: RefCell<FreevarMap>, region_maps: RegionMaps, lang_items: LanguageItems, stability: Index<'tcx>, f: F) -> (Session, R) where F: FnOnce(&ctxt<'tcx>) -> R

Unstable

Create a type context and call the closure with a &ty::ctxt reference to the context. The closure enforces that the type context and any interned value (types, substs, etc.) can only be used while ty::tls has a valid reference to the context, to allow formatting values that need it.

fn mk_substs(&self, substs: Substs<'tcx>) -> &'tcx Substs<'tcx>

Unstable

fn safe_to_unsafe_fn_ty(&self, bare_fn: &BareFnTy<'tcx>) -> Ty<'tcx>

Unstable

Create an unsafe fn ty based on a safe fn ty.

fn mk_bare_fn(&self, bare_fn: BareFnTy<'tcx>) -> &'tcx BareFnTy<'tcx>

Unstable

fn mk_region(&self, region: Region) -> &'tcx Region

Unstable

fn closure_kind(&self, def_id: DefId) -> ClosureKind

Unstable

fn closure_type(&self, def_id: DefId, substs: &ClosureSubsts<'tcx>) -> ClosureTy<'tcx>

Unstable

fn type_parameter_def(&self, node_id: NodeId) -> TypeParameterDef<'tcx>

Unstable

fn pat_contains_ref_binding(&self, pat: &Pat) -> Option<Mutability>

Unstable

fn arm_contains_ref_binding(&self, arm: &Arm) -> Option<Mutability>

Unstable

fn mk_ty(&self, st: TypeVariants<'tcx>) -> Ty<'tcx>

Unstable

fn mk_mach_int(&self, tm: IntTy) -> Ty<'tcx>

Unstable

fn mk_mach_uint(&self, tm: UintTy) -> Ty<'tcx>

Unstable

fn mk_mach_float(&self, tm: FloatTy) -> Ty<'tcx>

Unstable

fn mk_str(&self) -> Ty<'tcx>

Unstable

fn mk_static_str(&self) -> Ty<'tcx>

Unstable

fn mk_enum(&self, def: AdtDef<'tcx>, substs: &'tcx Substs<'tcx>) -> Ty<'tcx>

Unstable

fn mk_box(&self, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_ptr(&self, tm: TypeAndMut<'tcx>) -> Ty<'tcx>

Unstable

fn mk_ref(&self, r: &'tcx Region, tm: TypeAndMut<'tcx>) -> Ty<'tcx>

Unstable

fn mk_mut_ref(&self, r: &'tcx Region, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_imm_ref(&self, r: &'tcx Region, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_mut_ptr(&self, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_imm_ptr(&self, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_nil_ptr(&self) -> Ty<'tcx>

Unstable

fn mk_array(&self, ty: Ty<'tcx>, n: usize) -> Ty<'tcx>

Unstable

fn mk_slice(&self, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_tup(&self, ts: Vec<Ty<'tcx>>) -> Ty<'tcx>

Unstable

fn mk_nil(&self) -> Ty<'tcx>

Unstable

fn mk_bool(&self) -> Ty<'tcx>

Unstable

fn mk_fn(&self, opt_def_id: Option<DefId>, fty: &'tcx BareFnTy<'tcx>) -> Ty<'tcx>

Unstable

fn mk_ctor_fn(&self, def_id: DefId, input_tys: &[Ty<'tcx>], output: Ty<'tcx>) -> Ty<'tcx>

Unstable

fn mk_trait(&self, principal: PolyTraitRef<'tcx>, bounds: ExistentialBounds<'tcx>) -> Ty<'tcx>

Unstable

fn mk_projection(&self, trait_ref: TraitRef<'tcx>, item_name: Name) -> Ty<'tcx>

Unstable

fn mk_struct(&self, def: AdtDef<'tcx>, substs: &'tcx Substs<'tcx>) -> Ty<'tcx>

Unstable

fn mk_closure(&self, closure_id: DefId, substs: &'tcx Substs<'tcx>, tys: Vec<Ty<'tcx>>) -> Ty<'tcx>

Unstable

fn mk_closure_from_closure_substs(&self, closure_id: DefId, closure_substs: Box<ClosureSubsts<'tcx>>) -> Ty<'tcx>

Unstable

fn mk_var(&self, v: TyVid) -> Ty<'tcx>

Unstable

fn mk_int_var(&self, v: IntVid) -> Ty<'tcx>

Unstable

fn mk_float_var(&self, v: FloatVid) -> Ty<'tcx>

Unstable

fn mk_infer(&self, it: InferTy) -> Ty<'tcx>

Unstable

fn mk_param(&self, space: ParamSpace, index: u32, name: Name) -> Ty<'tcx>

Unstable

fn mk_self_type(&self) -> Ty<'tcx>

Unstable

fn mk_param_from_def(&self, def: &TypeParameterDef) -> Ty<'tcx>

Unstable

impl<'tcx> ctxt<'tcx>

fn positional_element_ty(&self, ty: Ty<'tcx>, i: usize, variant: Option<DefId>) -> Option<Ty<'tcx>>

Unstable

Returns the type of element at index i in tuple or tuple-like type t. For an enum t, variant is None only if t is a univariant enum.

fn named_element_ty(&self, ty: Ty<'tcx>, n: Name, variant: Option<DefId>) -> Option<Ty<'tcx>>

Unstable

Returns the type of element at field n in struct or struct-like type t. For an enum t, variant must be some def id.

fn node_id_to_type(&self, id: NodeId) -> Ty<'tcx>

Unstable

fn node_id_to_type_opt(&self, id: NodeId) -> Option<Ty<'tcx>>

Unstable

fn node_id_item_substs(&self, id: NodeId) -> ItemSubsts<'tcx>

Unstable

fn pat_ty(&self, pat: &Pat) -> Ty<'tcx>

Unstable

fn pat_ty_opt(&self, pat: &Pat) -> Option<Ty<'tcx>>

Unstable

fn expr_ty(&self, expr: &Expr) -> Ty<'tcx>

Unstable

fn expr_ty_opt(&self, expr: &Expr) -> Option<Ty<'tcx>>

Unstable

fn expr_ty_adjusted(&self, expr: &Expr) -> Ty<'tcx>

Unstable

Returns the type of expr, considering any AutoAdjustment entry recorded for that expression.

It would almost certainly be better to store the adjusted ty in with the AutoAdjustment, but I opted not to do this because it would require serializing and deserializing the type and, although that's not hard to do, I just hate that code so much I didn't want to touch it unless it was to fix it properly, which seemed a distraction from the thread at hand! -nmatsakis

fn expr_span(&self, id: NodeId) -> Span

Unstable

fn local_var_name_str(&self, id: NodeId) -> InternedString

Unstable

fn resolve_expr(&self, expr: &Expr) -> Def

Unstable

fn expr_is_lval(&self, expr: &Expr) -> bool

Unstable

fn note_and_explain_type_err(&self, err: &TypeError<'tcx>, sp: Span)

Unstable

fn provided_source(&self, id: DefId) -> Option<DefId>

Unstable

fn provided_trait_methods(&self, id: DefId) -> Vec<Rc<Method<'tcx>>>

Unstable

fn associated_consts(&self, id: DefId) -> Vec<Rc<AssociatedConst<'tcx>>>

Unstable

fn trait_items(&self, trait_did: DefId) -> Rc<Vec<ImplOrTraitItem<'tcx>>>

Unstable

fn trait_impl_polarity(&self, id: DefId) -> Option<ImplPolarity>

Unstable

fn custom_coerce_unsized_kind(&self, did: DefId) -> CustomCoerceUnsized

Unstable

fn impl_or_trait_item(&self, id: DefId) -> ImplOrTraitItem<'tcx>

Unstable

fn trait_item_def_ids(&self, id: DefId) -> Rc<Vec<ImplOrTraitItemId>>

Unstable

fn impl_trait_ref(&self, id: DefId) -> Option<TraitRef<'tcx>>

Unstable

Returns the trait-ref corresponding to a given impl, or None if it is an inherent impl.

fn is_impl(&self, id: DefId) -> bool

Unstable

Returns whether this DefId refers to an impl

fn trait_ref_to_def_id(&self, tr: &TraitRef) -> DefId

Unstable

fn try_add_builtin_trait(&self, trait_def_id: DefId, builtin_bounds: &mut EnumSet<BuiltinBound>) -> bool

Unstable

Checks whether trait_ref refers to one of the builtin traits, like Send, and adds the corresponding bound to the set builtin_bounds if so. Returns true if trait_ref is a builtin trait.

fn item_path_str(&self, id: DefId) -> String

Unstable

fn ty_dtor(&self, struct_id: DefId) -> DtorKind

Unstable

fn with_path<T, F>(&self, id: DefId, f: F) -> T where F: FnOnce(PathElems) -> T

Unstable

fn enum_repr_type(&self, opt_hint: Option<&ReprAttr>) -> (IntType, Ty<'tcx>)

Unstable

Returns (normalized_type, ty), where normalized_type is the IntType representation of one of {i64,i32,i16,i8,u64,u32,u16,u8}, and ty is the original type (i.e. may include isize or usize).

fn register_item_type(&self, did: DefId, ty: TypeScheme<'tcx>)

Unstable

fn lookup_item_type(&self, did: DefId) -> TypeScheme<'tcx>

Unstable

fn lookup_trait_def(&self, did: DefId) -> &'tcx TraitDef<'tcx>

Unstable

Given the did of a trait, returns its canonical trait ref.

fn lookup_adt_def_master(&self, did: DefId) -> AdtDefMaster<'tcx>

Unstable

Given the did of an ADT, return a master reference to its definition. Unless you are planning on fulfilling the ADT's fields, use lookup_adt_def instead.

fn lookup_adt_def(&self, did: DefId) -> AdtDef<'tcx>

Unstable

Given the did of an ADT, return a reference to its definition.

fn lookup_predicates(&self, did: DefId) -> GenericPredicates<'tcx>

Unstable

Given the did of an item, returns its full set of predicates.

fn lookup_super_predicates(&self, did: DefId) -> GenericPredicates<'tcx>

Unstable

Given the did of a trait, returns its superpredicates.

fn get_attrs(&self, did: DefId) -> Cow<'tcx, [Attribute]>

Unstable

Get the attributes of a definition.

fn has_attr(&self, did: DefId, attr: &str) -> bool

Unstable

Determine whether an item is annotated with an attribute

fn lookup_packed(&self, did: DefId) -> bool

Unstable

Determine whether an item is annotated with #[repr(packed)]

fn lookup_simd(&self, did: DefId) -> bool

Unstable

Determine whether an item is annotated with #[simd]

fn lookup_repr_hints(&self, did: DefId) -> Rc<Vec<ReprAttr>>

Unstable

Obtain the representation annotation for a struct definition.

fn struct_tail(&self, ty: Ty<'tcx>) -> Ty<'tcx>

Unstable

Returns the deeply last field of nested structures, or the same type, if not a structure at all. Corresponds to the only possible unsized field, and its type can be used to determine unsizing strategy.

fn struct_lockstep_tails(&self, source: Ty<'tcx>, target: Ty<'tcx>) -> (Ty<'tcx>, Ty<'tcx>)

Unstable

Same as applying struct_tail on source and target, but only keeps going as long as the two types are instances of the same structure definitions. For (Foo<Foo<T>>, Foo<Trait>), the result will be (Foo<T>, Trait), whereas struct_tail produces T, and Trait, respectively.

fn eval_repeat_count(&self, count_expr: &Expr) -> usize

Unstable

fn each_bound_trait_and_supertraits<F>(&self, bounds: &[PolyTraitRef<'tcx>], f: F) -> bool where F: FnMut(PolyTraitRef<'tcx>) -> bool

Unstable

fn required_region_bounds(&self, erased_self_ty: Ty<'tcx>, predicates: Vec<Predicate<'tcx>>) -> Vec<Region>

Unstable

Given a set of predicates that apply to an object type, returns the region bounds that the (erased) Self type must outlive. Precisely because the Self type is erased, the parameter erased_self_ty must be supplied to indicate what type has been used to represent Self in the predicates themselves. This should really be a unique type; FreshTy(0) is a popular choice.

Requires that trait definitions have been processed so that we can elaborate predicates and walk supertraits.

fn item_variances(&self, item_id: DefId) -> Rc<ItemVariances>

Unstable

fn trait_has_default_impl(&self, trait_def_id: DefId) -> bool

Unstable

fn record_trait_has_default_impl(&self, trait_def_id: DefId)

Unstable

Records a trait-to-implementation mapping.

fn populate_implementations_for_primitive_if_necessary(&self, primitive_def_id: DefId)

Unstable

Load primitive inherent implementations if necessary

fn populate_inherent_implementations_for_type_if_necessary(&self, type_id: DefId)

Unstable

Populates the type context with all the inherent implementations for the given type if necessary.

fn populate_implementations_for_trait_if_necessary(&self, trait_id: DefId)

Unstable

Populates the type context with all the implementations for the given trait if necessary.

fn trait_id_of_impl(&self, def_id: DefId) -> Option<DefId>

Unstable

Given the def_id of an impl, return the def_id of the trait it implements. If it implements no trait, return None.

fn impl_of_method(&self, def_id: DefId) -> Option<DefId>

Unstable

If the given def ID describes a method belonging to an impl, return the ID of the impl that the method belongs to. Otherwise, return None.

fn trait_of_item(&self, def_id: DefId) -> Option<DefId>

Unstable

If the given def ID describes an item belonging to a trait (either a default method or an implementation of a trait method), return the ID of the trait that the method belongs to. Otherwise, return None.

fn trait_item_of_item(&self, def_id: DefId) -> Option<ImplOrTraitItemId>

Unstable

If the given def ID describes an item belonging to a trait, (either a default method or an implementation of a trait method), return the ID of the method inside trait definition (this means that if the given def ID is already that of the original trait method, then the return value is the same). Otherwise, return None.

fn hash_crate_independent(&self, ty: Ty<'tcx>, svh: &Svh) -> u64

Unstable

Creates a hash of the type Ty which will be the same no matter what crate context it's calculated within. This is used by the type_id intrinsic.

fn empty_parameter_environment<'a>(&'a self) -> ParameterEnvironment<'a, 'tcx>

Unstable

Construct a parameter environment suitable for static contexts or other contexts where there are no free type/lifetime parameters in scope.

fn construct_free_substs(&self, generics: &Generics<'tcx>, free_id: NodeId) -> Substs<'tcx>

Unstable

Constructs and returns a substitution that can be applied to move from the "outer" view of a type or method to the "inner" view. In general, this means converting from bound parameters to free parameters. Since we currently represent bound/free type parameters in the same way, this only has an effect on regions.

fn construct_parameter_environment<'a>(&'a self, span: Span, generics: &Generics<'tcx>, generic_predicates: &GenericPredicates<'tcx>, free_id: NodeId) -> ParameterEnvironment<'a, 'tcx>

Unstable

See ParameterEnvironment struct def'n for details

fn is_method_call(&self, expr_id: NodeId) -> bool

Unstable

fn is_overloaded_autoderef(&self, expr_id: NodeId, autoderefs: u32) -> bool

Unstable

fn upvar_capture(&self, upvar_id: UpvarId) -> Option<UpvarCapture>

Unstable

impl<'tcx> ctxt<'tcx>

fn with_freevars<T, F>(&self, fid: NodeId, f: F) -> T where F: FnOnce(&[Freevar]) -> T

Unstable

fn liberate_late_bound_regions<T>(&self, all_outlive_scope: DestructionScopeData, value: &Binder<T>) -> T where T: TypeFoldable<'tcx>

Unstable

Replace any late-bound regions bound in value with free variants attached to scope-id scope_id.

fn flatten_late_bound_regions<T>(&self, bound2_value: &Binder<Binder<T>>) -> Binder<T> where T: TypeFoldable<'tcx>

Unstable

Flattens two binding levels into one. So for<'a> for<'b> Foo becomes for<'a,'b> Foo.

fn no_late_bound_regions<T>(&self, value: &Binder<T>) -> Option<T> where T: TypeFoldable<'tcx> + RegionEscape

Unstable

fn erase_late_bound_regions<T>(&self, value: &Binder<T>) -> T where T: TypeFoldable<'tcx>

Unstable

Replace any late-bound regions bound in value with 'static. Useful in trans but also method lookup and a few other places where precise region relationships are not required.

fn anonymize_late_bound_regions<T>(&self, sig: &Binder<T>) -> Binder<T> where T: TypeFoldable<'tcx>

Unstable

Rewrite any late-bound regions so that they are anonymous. Region numbers are assigned starting at 1 and increasing monotonically in the order traversed by the fold operation.

The chief purpose of this function is to canonicalize regions so that two FnSigs or TraitRefs which are equivalent up to region naming will become structurally identical. For example, for<'a, 'b> fn(&'a isize, &'b isize) and for<'a, 'b> fn(&'b isize, &'a isize) will become identical after anonymization.

fn make_substs_for_receiver_types(&self, trait_ref: &TraitRef<'tcx>, method: &Method<'tcx>) -> Substs<'tcx>

Unstable

Substitutes the values for the receiver's type parameters that are found in method, leaving the method's type parameters intact.