Trait rustc::middle::ty::Lift

pub trait Lift<'tcx> {
    type Lifted;
    fn lift_to_tcx(&self, tcx: &ctxt<'tcx>) -> Option<Self::Lifted>;
}

Unstable (rustc_private)

A trait implemented for all X<'a> types which can be safely and efficiently converted to X<'tcx> as long as they are part of the provided ty::ctxt<'tcx>. This can be done, for example, for Ty<'tcx> or &'tcx Substs<'tcx> by looking them up in their respective interners. None is returned if the value or one of the components is not part of the provided context. For Ty, None can be returned if either the type interner doesn't contain the TypeVariants key or if the address of the interned pointer differs. The latter case is possible if a primitive type, e.g. () or u8, was interned in a different context.

Associated Types

type Lifted

Unstable (rustc_private)

Required Methods

fn lift_to_tcx(&self, tcx: &ctxt<'tcx>) -> Option<Self::Lifted>

Unstable (rustc_private)