Module rustc::mir

Unstable (rustc_private)

Modules

tcx 	 [Unstable]
Methods for the various MIR types. These are intended for use after building is complete.

transform 	 [Unstable]

traversal 	 [Unstable]

visit 	 [Unstable]

Structs

BasicBlock 	 [Unstable]

BasicBlockData 	 [Unstable]

Constant 	 [Unstable]
Constants

Field 	 [Unstable]

Local 	 [Unstable]

LocalDecl 	 [Unstable]
A MIR local.

Location 	 [Unstable]

Mir 	 [Unstable]
Lowered representation of a single function.

Projection 	 [Unstable]
The Projection data structure defines things of the form B.x or *B or B[index]. Note that it is parameterized because it is shared between Constant and Lvalue. See the aliases LvalueProjection etc below.

Promoted 	 [Unstable]

SourceInfo 	 [Unstable]
Grouped information about the source code origin of a MIR entity. Intended to be inspected by diagnostics and debuginfo. Most passes can work with it as a whole, within a single function.

Statement 	 [Unstable]

Terminator 	 [Unstable]

TypedConstVal 	 [Unstable]

UpvarDecl 	 [Unstable]
A closure capture, with its name and mode.

VisibilityScope 	 [Unstable]

VisibilityScopeData 	 [Unstable]

Enums

AggregateKind 	 [Unstable]

AssertMessage 	 [Unstable]

BinOp 	 [Unstable]

BorrowKind 	 [Unstable]

CastKind 	 [Unstable]

Literal 	 [Unstable]

LocalKind 	 [Unstable]
Classifies locals into categories. See Mir::local_kind.

Lvalue 	 [Unstable]
A path to a value; something that can be evaluated without changing or disturbing program state.

Mutability 	 [Unstable]

Operand 	 [Unstable]
These are values that can appear inside an rvalue (or an index lvalue). They are intentionally limited to prevent rvalues from being nested in one another.

ProjectionElem 	 [Unstable]

Rvalue 	 [Unstable]
Rvalues

StatementKind 	 [Unstable]

TerminatorKind 	 [Unstable]

UnOp 	 [Unstable]

Constants

ARGUMENT_VISIBILITY_SCOPE 	 [Unstable]

RETURN_POINTER 	 [Unstable]

START_BLOCK 	 [Unstable]
where execution begins

Type Definitions

LvalueElem 	 [Unstable]
Alias for projections as they appear in lvalues, where the base is an lvalue and the index is an operand.

LvalueProjection 	 [Unstable]
Alias for projections as they appear in lvalues, where the base is an lvalue and the index is an operand.