Enum rustc::mir::Lvalue

pub enum Lvalue<'tcx> {
    Local(Local),
    Static(Box<Static<'tcx>>),
    Projection(Box<LvalueProjection<'tcx>>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A path to a value; something that can be evaluated without changing or disturbing program state.

Variants
Local(Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

local variable

Static(Box<Static<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

static or static mut variable

Projection(Box<LvalueProjection<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

projection out of an lvalue (access a field, deref a pointer, etc)

Methods

impl<'tcx> Lvalue<'tcx>

pub fn ty<'a, 'gcx, D>(
&self,
local_decls: &D,
tcx: TyCtxt<'a, 'gcx, 'tcx>
) -> LvalueTy<'tcx> where
D: HasLocalDecls<'tcx>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl<'tcx> Lvalue<'tcx>

pub fn field(self, f: Field, ty: Ty<'tcx>) -> Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn deref(self) -> Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn downcast(
self,
adt_def: &'tcx AdtDef,
variant_index: usize
) -> Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn index(self, index: Local) -> Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn elem(self, elem: LvalueElem<'tcx>) -> Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?