Enum rustc::mir::Rvalue

pub enum Rvalue<'tcx> {
    Use(Operand<'tcx>),
    Repeat(Operand<'tcx>, TypedConstVal<'tcx>),
    Ref(&'tcx Region, BorrowKind, Lvalue<'tcx>),
    Len(Lvalue<'tcx>),
    Cast(CastKind, Operand<'tcx>, Ty<'tcx>),
    BinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>),
    CheckedBinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>),
    UnaryOp(UnOp, Operand<'tcx>),
    Box(Ty<'tcx>),
    Aggregate(AggregateKind<'tcx>, Vec<Operand<'tcx>>),
    InlineAsm {
        asm: InlineAsm,
        outputs: Vec<Lvalue<'tcx>>,
        inputs: Vec<Operand<'tcx>>,
    },
}

This is an experimental API. (rustc_private)

Rvalues

Variants
Use(Operand<'tcx>)

This is an experimental API. (rustc_private)

x (either a move or copy, depending on type of x)

Repeat(Operand<'tcx>, TypedConstVal<'tcx>)

This is an experimental API. (rustc_private)

[x; 32]

Ref(&'tcx Region, BorrowKind, Lvalue<'tcx>)

This is an experimental API. (rustc_private)

&x or &mut x

Len(Lvalue<'tcx>)

This is an experimental API. (rustc_private)

length of a [X] or [X;n] value

Cast(CastKind, Operand<'tcx>, Ty<'tcx>)

This is an experimental API. (rustc_private)

BinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>)

This is an experimental API. (rustc_private)

CheckedBinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>)

This is an experimental API. (rustc_private)

UnaryOp(UnOp, Operand<'tcx>)

This is an experimental API. (rustc_private)

Box(Ty<'tcx>)

This is an experimental API. (rustc_private)

Creates an uninitialized Box

Aggregate(AggregateKind<'tcx>, Vec<Operand<'tcx>>)

This is an experimental API. (rustc_private)

Create an aggregate value, like a tuple or struct. This is only needed because we want to distinguish dest = Foo { x: ..., y: ... } from dest.x = ...; dest.y = ...; in the case that Foo has a destructor. These rvalues can be optimized away after type-checking and before lowering.

InlineAsm

This is an experimental API. (rustc_private)

Fields of InlineAsm

asm: InlineAsm

This is an experimental API. (rustc_private)

outputs: Vec<Lvalue<'tcx>>

This is an experimental API. (rustc_private)

inputs: Vec<Operand<'tcx>>

This is an experimental API. (rustc_private)

Methods

impl<'tcx> Rvalue<'tcx>

fn ty<'a, 'gcx>(&self,
mir: &Mir<'tcx>,
tcx: TyCtxt<'a, 'gcx, 'tcx>)
-> Option<Ty<'tcx>>

This is an experimental API. (rustc_private)