Enum rustc::mir::StatementKind

pub enum StatementKind<'tcx> {
    Assign(Lvalue<'tcx>, Rvalue<'tcx>),
    SetDiscriminant {
        lvalue: Lvalue<'tcx>,
        variant_index: usize,
    },
    StorageLive(Local),
    StorageDead(Local),
    InlineAsm {
        asm: Box<InlineAsm>,
        outputs: Vec<Lvalue<'tcx>>,
        inputs: Vec<Operand<'tcx>>,
    },
    Validate(ValidationOp, Vec<ValidationOperand<'tcx, Lvalue<'tcx>>>),
    EndRegion(Scope),
    Nop,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
Assign(Lvalue<'tcx>, Rvalue<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Write the RHS Rvalue to the LHS Lvalue.

SetDiscriminant

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Write the discriminant for a variant to the enum Lvalue.

Fields of SetDiscriminant

lvalue: Lvalue<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

variant_index: usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

StorageLive(Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Start a live range for the storage of the local.

StorageDead(Local)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

End the current live range for the storage of the local.

InlineAsm

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Execute a piece of inline Assembly.

Fields of InlineAsm

asm: Box<InlineAsm>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

outputs: Vec<Lvalue<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

inputs: Vec<Operand<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Validate(ValidationOp, Vec<ValidationOperand<'tcx, Lvalue<'tcx>>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Assert the given lvalues to be valid inhabitants of their type. These statements are currently only interpreted by miri and only generated when "-Z mir-emit-validate" is passed. See https://internals.rust-lang.org/t/types-as-contracts/5562/73 for more details.

EndRegion(Scope)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Mark one terminating point of a region scope (i.e. static region). (The starting point(s) arise implicitly from borrows.)

Nop

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

No-op. Useful for deleting instructions without affecting statement indices.