Enum rustc::mir::repr::ProjectionElem

pub enum ProjectionElem<'tcx, V> {
    Deref,
    Field(Field, Ty<'tcx>),
    Index(V),
    ConstantIndex {
        offset: u32,
        min_length: u32,
        from_end: bool,
    },
    Downcast(AdtDef<'tcx>, usize),
}

Unstable (rustc_private)

Variants

Deref
Unstable (rustc_private)

Field
Unstable (rustc_private)

Index
Unstable (rustc_private)

ConstantIndex
Unstable (rustc_private)

These indices are generated by slice patterns. Easiest to explain by example:

[X, _, .._, _, _] => { offset: 0, min_length: 4, from_end: false },
[_, X, .._, _, _] => { offset: 1, min_length: 4, from_end: false },
[_, _, .._, X, _] => { offset: 2, min_length: 4, from_end: true },
[_, _, .._, _, X] => { offset: 1, min_length: 4, from_end: true },

Fields

offset
Unstable (rustc_private)

index or -index (in Python terms), depending on from_end

min_length
Unstable (rustc_private)

thing being indexed must be at least this long

from_end
Unstable (rustc_private)

counting backwards from end?

Downcast
Unstable (rustc_private)

"Downcast" to a variant of an ADT. Currently, we only introduce this for ADTs with more than one variant. It may be better to just introduce it always, or always for enums.