Enum rustc::mir::repr::Rvalue

pub enum Rvalue<'tcx> {
    Use(Operand<'tcx>),
    Repeat(Operand<'tcx>, TypedConstVal<'tcx>),
    Ref(Region, BorrowKind, Lvalue<'tcx>),
    Len(Lvalue<'tcx>),
    Cast(CastKind, Operand<'tcx>, Ty<'tcx>),
    BinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>),
    CheckedBinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>),
    UnaryOp(UnOp, Operand<'tcx>),
    Box(Ty<'tcx>),
    Aggregate(AggregateKind<'tcx>, Vec<Operand<'tcx>>),
    InlineAsm {
        asm: InlineAsm,
        outputs: Vec<Lvalue<'tcx>>,
        inputs: Vec<Operand<'tcx>>,
    },
}

Unstable (rustc_private)

Rvalues

Variants
Use(Operand<'tcx>)
Unstable (rustc_private)

x (either a move or copy, depending on type of x)

Repeat(Operand<'tcx>, TypedConstVal<'tcx>)
Unstable (rustc_private)

[x; 32]

Ref(Region, BorrowKind, Lvalue<'tcx>)
Unstable (rustc_private)

&x or &mut x

Len(Lvalue<'tcx>)
Unstable (rustc_private)

length of a [X] or [X;n] value

Cast(CastKind, Operand<'tcx>, Ty<'tcx>)
Unstable (rustc_private)
BinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>)
Unstable (rustc_private)
CheckedBinaryOp(BinOp, Operand<'tcx>, Operand<'tcx>)
Unstable (rustc_private)
UnaryOp(UnOp, Operand<'tcx>)
Unstable (rustc_private)
Box(Ty<'tcx>)
Unstable (rustc_private)

Creates an uninitialized Box

Aggregate(AggregateKind<'tcx>, Vec<Operand<'tcx>>)
Unstable (rustc_private)

Create an aggregate value, like a tuple or struct. This is only needed because we want to distinguish dest = Foo { x: ..., y: ... } from dest.x = ...; dest.y = ...; in the case that Foo has a destructor. These rvalues can be optimized away after type-checking and before lowering.

InlineAsm
Unstable (rustc_private)

Fields

asm: InlineAsm
Unstable (rustc_private)

outputs: Vec<Lvalue<'tcx>>
Unstable (rustc_private)

inputs: Vec<Operand<'tcx>>
Unstable (rustc_private)