Enum rustc::mir::repr::StatementKind

pub enum StatementKind<'tcx> {
    Assign(Lvalue<'tcx>, Rvalue<'tcx>),
    SetDiscriminant {
        lvalue: Lvalue<'tcx>,
        variant_index: usize,
    },
    StorageLive(Lvalue<'tcx>),
    StorageDead(Lvalue<'tcx>),
    Nop,
}

Unstable (rustc_private)

Variants
Assign(Lvalue<'tcx>, Rvalue<'tcx>)
Unstable (rustc_private)

Write the RHS Rvalue to the LHS Lvalue.

SetDiscriminant
Unstable (rustc_private)

Write the discriminant for a variant to the enum Lvalue.

Fields

lvalue: Lvalue<'tcx>
Unstable (rustc_private)

variant_index: usize
Unstable (rustc_private)

StorageLive(Lvalue<'tcx>)
Unstable (rustc_private)

Start a live range for the storage of the local.

StorageDead(Lvalue<'tcx>)
Unstable (rustc_private)

End the current live range for the storage of the local.

Nop
Unstable (rustc_private)

No-op. Useful for deleting instructions without affecting statement indices.