Enum rustc::mir::repr::Terminator

pub enum Terminator<'tcx> {
    Goto {
        target: BasicBlock,
    },
    Panic {
        target: BasicBlock,
    },
    If {
        cond: Operand<'tcx>,
        targets: (BasicBlock, BasicBlock),
    },
    Switch {
        discr: Lvalue<'tcx>,
        adt_def: AdtDef<'tcx>,
        targets: Vec<BasicBlock>,
    },
    SwitchInt {
        discr: Lvalue<'tcx>,
        switch_ty: Ty<'tcx>,
        values: Vec<ConstVal>,
        targets: Vec<BasicBlock>,
    },
    Diverge,
    Return,
    Call {
        data: CallData<'tcx>,
        targets: (BasicBlock, BasicBlock),
    },
}

Unstable (rustc_private)

Variants

Goto
Unstable (rustc_private)

block should have one successor in the graph; we jump there

Fields

target
Unstable (rustc_private)

Panic
Unstable (rustc_private)

block should initiate unwinding; should be one successor that does cleanup and branches to DIVERGE_BLOCK

Fields

target
Unstable (rustc_private)

If
Unstable (rustc_private)

jump to branch 0 if this lvalue evaluates to true

Fields

cond
Unstable (rustc_private)

targets
Unstable (rustc_private)

Switch
Unstable (rustc_private)

lvalue evaluates to some enum; jump depending on the branch

Fields

discr
Unstable (rustc_private)

adt_def
Unstable (rustc_private)

targets
Unstable (rustc_private)

SwitchInt
Unstable (rustc_private)

operand evaluates to an integer; jump depending on its value to one of the targets, and otherwise fallback to otherwise

Fields

discr
Unstable (rustc_private)

discriminant value being tested

switch_ty
Unstable (rustc_private)

type of value being tested

values
Unstable (rustc_private)

Possible values. The locations to branch to in each case are found in the corresponding indices from the targets vector.

targets
Unstable (rustc_private)

Possible branch sites. The length of this vector should be equal to the length of the values vector plus 1 -- the extra item is the block to branch to if none of the values fit.

Diverge
Unstable (rustc_private)

Indicates that the last statement in the block panics, aborts, etc. No successors. This terminator appears on exactly one basic block which we create in advance. However, during construction, we use this value as a sentinel for "terminator not yet assigned", and assert at the end that only the well-known diverging block actually diverges.

Return
Unstable (rustc_private)

Indicates a normal return. The ReturnPointer lvalue should have been filled in by now. This should only occur in the END_BLOCK.

Call
Unstable (rustc_private)

block ends with a call; it should have two successors. The first successor indicates normal return. The second indicates unwinding.

Fields

data
Unstable (rustc_private)

targets
Unstable (rustc_private)

Methods

impl<'tcx> Terminator<'tcx>

fn successors(&self) -> &[BasicBlock]

Unstable (rustc_private)

fn successors_mut(&mut self) -> &mut [BasicBlock]

Unstable (rustc_private)