Struct rustc::mir::repr::ArgDecl

pub struct ArgDecl<'tcx> {
    pub ty: Ty<'tcx>,
    pub spread: bool,
    pub debug_name: Name,
}

Unstable (rustc_private)

A "arg" is one of the function's formal arguments. These are anonymous and distinct from the bindings that the user declares.

For example, in this function:

fn foo((x, y): (i32, u32)) { ... }

there is only one argument, of type (i32, u32), but two bindings (x and y).

Fields
ty: Ty<'tcx>
Unstable (rustc_private)
spread: bool
Unstable (rustc_private)

If true, this argument is a tuple after monomorphization, and has to be collected from multiple actual arguments.

debug_name: Name
Unstable (rustc_private)

Either keywords::Invalid or the name of a single-binding pattern associated with this argument. Useful for debuginfo.