Struct rustc::mir::repr::BasicBlockData

pub struct BasicBlockData<'tcx> {
    pub statements: Vec<Statement<'tcx>>,
    pub terminator: Option<Terminator<'tcx>>,
    pub is_cleanup: bool,
}

Unstable (rustc_private)

Fields
statements: Vec<Statement<'tcx>>
Unstable (rustc_private)

List of statements in this block.

terminator: Option<Terminator<'tcx>>
Unstable (rustc_private)

Terminator for this block.

NB. This should generally ONLY be None during construction. Therefore, you should generally access it via the terminator() or terminator_mut() methods. The only exception is that certain passes, such as simplify_cfg, swap out the terminator temporarily with None while they continue to recurse over the set of basic blocks.

is_cleanup: bool
Unstable (rustc_private)

If true, this block lies on an unwind path. This is used during trans where distinct kinds of basic blocks may be generated (particularly for MSVC cleanup). Unwind blocks must only branch to other unwind blocks.

Methods

impl<'tcx> BasicBlockData<'tcx>

fn new(terminator: Option<Terminator<'tcx>>) -> BasicBlockData<'tcx>

Unstable (rustc_private)

fn terminator(&self) -> &Terminator<'tcx>

Unstable (rustc_private)

Accessor for terminator.

Terminator may not be None after construction of the basic block is complete. This accessor provides a convenience way to reach the terminator.

fn terminator_mut(&mut self) -> &mut Terminator<'tcx>

Unstable (rustc_private)