Struct rustc::mir::repr::Mir

pub struct Mir<'tcx> {
    pub basic_blocks: Vec<BasicBlockData<'tcx>>,
    pub return_ty: FnOutput<'tcx>,
    pub var_decls: Vec<VarDecl<'tcx>>,
    pub arg_decls: Vec<ArgDecl<'tcx>>,
    pub temp_decls: Vec<TempDecl<'tcx>>,
    pub span: Span,
}

Unstable (rustc_private)

Lowered representation of a single function.

Fields

basic_blocks
Unstable (rustc_private)

List of basic blocks. References to basic block use a newtyped index type BasicBlock that indexes into this vector.

return_ty
Unstable (rustc_private)

Return type of the function.

var_decls
Unstable (rustc_private)

Variables: these are stack slots corresponding to user variables. They may be assigned many times.

arg_decls
Unstable (rustc_private)

Args: these are stack slots corresponding to the input arguments.

temp_decls
Unstable (rustc_private)

Temp declarations: stack slots that for temporaries created by the compiler. These are assigned once, but they are not SSA values in that it is possible to borrow them and mutate them through the resulting reference.

span
Unstable (rustc_private)

A span representing this MIR, for error reporting

Methods

impl<'tcx> Mir<'tcx>

fn all_basic_blocks(&self) -> Vec<BasicBlock>

Unstable (rustc_private)

fn basic_block_data(&self, bb: BasicBlock) -> &BasicBlockData<'tcx>

Unstable (rustc_private)

fn basic_block_data_mut(&mut self, bb: BasicBlock) -> &mut BasicBlockData<'tcx>

Unstable (rustc_private)

impl<'tcx> Mir<'tcx>

fn operand_ty(&self, tcx: &TyCtxt<'tcx>, operand: &Operand<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn binop_ty(&self, tcx: &TyCtxt<'tcx>, op: BinOp, lhs_ty: Ty<'tcx>, rhs_ty: Ty<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn lvalue_ty(&self, tcx: &TyCtxt<'tcx>, lvalue: &Lvalue<'tcx>) -> LvalueTy<'tcx>

Unstable (rustc_private)

fn rvalue_ty(&self, tcx: &TyCtxt<'tcx>, rvalue: &Rvalue<'tcx>) -> Option<Ty<'tcx>>

Unstable (rustc_private)