Struct rustc::mir::LocalDecl

pub struct LocalDecl<'tcx> {
    pub mutability: Mutability,
    pub ty: Ty<'tcx>,
    pub name: Option<Name>,
    pub source_info: Option<SourceInfo>,
}

Unstable (rustc_private)

A MIR local.

This can be a binding declared by the user, a temporary inserted by the compiler, a function argument, or the return pointer.

Fields
mutability: Mutability
Unstable (rustc_private)

let mut x vs let x.

Temporaries and the return pointer are always mutable.

ty: Ty<'tcx>
Unstable (rustc_private)

Type of this local.

name: Option<Name>
Unstable (rustc_private)

Name of the local, used in debuginfo and pretty-printing.

Note that function arguments can also have this set to Some(_) to generate better debuginfo.

source_info: Option<SourceInfo>
Unstable (rustc_private)

For user-declared variables, stores their source information.

For temporaries, this is None.

This is the primary way to differentiate between user-declared variables and compiler-generated temporaries.

Methods

impl<'tcx> LocalDecl<'tcx>

fn new_temp(ty: Ty<'tcx>) -> Self

Unstable (rustc_private)

Create a new LocalDecl for a temporary.

fn new_return_pointer(return_ty: Ty) -> LocalDecl

Unstable (rustc_private)

Builds a LocalDecl for the return pointer.

This must be inserted into the local_decls list as the first local.