Struct rustc::mir::Mir

pub struct Mir<'tcx> {
    pub visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>,
    pub promoted: IndexVec<Promoted, Mir<'tcx>>,
    pub return_ty: Ty<'tcx>,
    pub local_decls: IndexVec<Local, LocalDecl<'tcx>>,
    pub arg_count: usize,
    pub upvar_decls: Vec<UpvarDecl>,
    pub spread_arg: Option<Local>,
    pub span: Span,
    // some fields omitted
}

This is an experimental API. (rustc_private)

Lowered representation of a single function.

Fields
visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>

This is an experimental API. (rustc_private)

List of visibility (lexical) scopes; these are referenced by statements and used (eventually) for debuginfo. Indexed by a VisibilityScope.

promoted: IndexVec<Promoted, Mir<'tcx>>

This is an experimental API. (rustc_private)

Rvalues promoted from this function, such as borrows of constants. Each of them is the Mir of a constant with the fn's type parameters in scope, but a separate set of locals.

return_ty: Ty<'tcx>

This is an experimental API. (rustc_private)

Return type of the function.

local_decls: IndexVec<Local, LocalDecl<'tcx>>

This is an experimental API. (rustc_private)

Declarations of locals.

The first local is the return value pointer, followed by arg_count locals for the function arguments, followed by any user-declared variables and temporaries.

arg_count: usize

This is an experimental API. (rustc_private)

Number of arguments this function takes.

Starting at local 1, arg_count locals will be provided by the caller and can be assumed to be initialized.

If this MIR was built for a constant, this will be 0.

upvar_decls: Vec<UpvarDecl>

This is an experimental API. (rustc_private)

Names and capture modes of all the closure upvars, assuming the first argument is either the closure or a reference to it.

spread_arg: Option<Local>

This is an experimental API. (rustc_private)

Mark an argument local (which must be a tuple) as getting passed as its individual components at the LLVM level.

This is used for the "rust-call" ABI.

span: Span

This is an experimental API. (rustc_private)

A span representing this MIR, for error reporting

Methods

impl<'tcx> Mir<'tcx>

fn new(basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>,
visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>,
promoted: IndexVec<Promoted, Mir<'tcx>>,
return_ty: Ty<'tcx>,
local_decls: IndexVec<Local, LocalDecl<'tcx>>,
arg_count: usize,
upvar_decls: Vec<UpvarDecl>,
span: Span)
-> Self

This is an experimental API. (rustc_private)

fn basic_blocks(&self) -> &IndexVec<BasicBlock, BasicBlockData<'tcx>>

This is an experimental API. (rustc_private)

fn basic_blocks_mut(&mut self)
-> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

This is an experimental API. (rustc_private)

fn predecessors(&self) -> Ref<IndexVec<BasicBlock, Vec<BasicBlock>>>

This is an experimental API. (rustc_private)

fn predecessors_for(&self, bb: BasicBlock) -> Ref<Vec<BasicBlock>>

This is an experimental API. (rustc_private)

fn dominators(&self) -> Dominators<BasicBlock>

This is an experimental API. (rustc_private)

fn local_kind(&self, local: Local) -> LocalKind

This is an experimental API. (rustc_private)

fn temps_iter<'a>(&'a self) -> impl Iterator<Item=Local> + 'a

This is an experimental API. (rustc_private)

Returns an iterator over all temporaries.

fn vars_iter<'a>(&'a self) -> impl Iterator<Item=Local> + 'a

This is an experimental API. (rustc_private)

Returns an iterator over all user-declared locals.

fn args_iter(&self) -> impl Iterator<Item=Local>

This is an experimental API. (rustc_private)

Returns an iterator over all function arguments.

fn vars_and_temps_iter(&self) -> impl Iterator<Item=Local>

This is an experimental API. (rustc_private)

Returns an iterator over all user-defined variables and compiler-generated temporaries (all locals that are neither arguments nor the return pointer).

fn make_statement_nop(&mut self, location: Location)

This is an experimental API. (rustc_private)

Changes a statement to a nop. This is both faster than deleting instructions and avoids invalidating statement indices in Locations.