Module rustc::mir::transform

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

See the README for details on writing your own pass.

Structs

MirPassIndex 	 [
Experimental
]

MirSuite 	 [
Experimental
]

Passes 	 [
Experimental
]
A manager for MIR passes.

Enums

MirSource 	 [
Experimental
]
Where a specific Mir comes from.

Constants

MIR_CONST 	 [
Experimental
]
Run the passes we need to do constant qualification and evaluation.

MIR_OPTIMIZED 	 [
Experimental
]
Run the passes we need to consider the MIR optimized.

MIR_SUITES 	 [
Experimental
]
The number of "pass suites" that we have:

MIR_VALIDATED 	 [
Experimental
]
Run the passes we need to consider the MIR validated and ready for borrowck etc.

Traits

MirPass 	 [
Experimental
]
A streamlined trait that you can implement to create a pass; the pass will be named after the type, and it will consist of a main loop that goes over each available MIR and applies run_pass.

PassHook 	 [
Experimental
]
A pass hook is invoked both before and after each pass executes. This is primarily used to dump MIR for debugging.

Functions

default_name 	 [
Experimental
]
Generates a default name for the pass based on the name of the type T.

Type Definitions

PassId 	 [
Experimental
]
The full suite of types that identifies a particular application of a pass to a def-id.