Trait rustc::mir::transform::PassHook

pub trait PassHook {
    fn on_mir_pass<'a, 'tcx: 'a>(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>, 
        suite: MirSuite, 
        pass_num: MirPassIndex, 
        pass_name: &str, 
        source: MirSource, 
        mir: &Mir<'tcx>, 
        is_after: bool
    );
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A pass hook is invoked both before and after each pass executes. This is primarily used to dump MIR for debugging.

You can tell whether this is before or after by inspecting the mir parameter -- before the pass executes, it will be None (in which case you can inspect the MIR from previous pass by executing mir_cx.read_previous_mir()); after the pass executes, it will be Some() with the result of the pass (in which case the output from the previous pass is most likely stolen, so you would not want to try and access it). If the pass is interprocedural, then the hook will be invoked once per output.

Required Methods

fn on_mir_pass<'a, 'tcx: 'a>(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
suite: MirSuite,
pass_num: MirPassIndex,
pass_name: &str,
source: MirSource,
mir: &Mir<'tcx>,
is_after: bool
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?