Trait rustc::mir::visit::Visitor

pub trait Visitor<'tcx> {
    fn visit_mir(&mut self, mir: &Mir<'tcx>) { ... }

    fn visit_basic_block_data(
        &mut self, 
        block: BasicBlock, 
        data: &BasicBlockData<'tcx>
    ) { ... }

    fn visit_visibility_scope_data(&mut self, scope_data: &VisibilityScopeData) { ... }

    fn visit_statement(
        &mut self, 
        block: BasicBlock, 
        statement: &Statement<'tcx>, 
        location: Location
    ) { ... }

    fn visit_assign(
        &mut self, 
        block: BasicBlock, 
        lvalue: &Lvalue<'tcx>, 
        rvalue: &Rvalue<'tcx>, 
        location: Location
    ) { ... }

    fn visit_terminator(
        &mut self, 
        block: BasicBlock, 
        terminator: &Terminator<'tcx>, 
        location: Location
    ) { ... }

    fn visit_terminator_kind(
        &mut self, 
        block: BasicBlock, 
        kind: &TerminatorKind<'tcx>, 
        location: Location
    ) { ... }

    fn visit_assert_message(
        &mut self, 
        msg: &AssertMessage<'tcx>, 
        location: Location
    ) { ... }

    fn visit_rvalue(&mut self, rvalue: &Rvalue<'tcx>, location: Location) { ... }

    fn visit_operand(&mut self, operand: &Operand<'tcx>, location: Location) { ... }

    fn visit_lvalue(
        &mut self, 
        lvalue: &Lvalue<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn visit_static(
        &mut self, 
        static_: &Static<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn visit_projection(
        &mut self, 
        lvalue: &LvalueProjection<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn visit_projection_elem(
        &mut self, 
        lvalue: &LvalueElem<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn visit_branch(&mut self, source: BasicBlock, target: BasicBlock) { ... }

    fn visit_constant(&mut self, constant: &Constant<'tcx>, location: Location) { ... }

    fn visit_literal(&mut self, literal: &Literal<'tcx>, location: Location) { ... }

    fn visit_def_id(&mut self, def_id: &DefId, _: Location) { ... }

    fn visit_span(&mut self, span: &Span) { ... }

    fn visit_source_info(&mut self, source_info: &SourceInfo) { ... }

    fn visit_ty(&mut self, ty: &Ty<'tcx>, _: Lookup) { ... }

    fn visit_region(&mut self, region: &Region<'tcx>, _: Location) { ... }

    fn visit_const(&mut self, constant: &&'tcx Const<'tcx>, _: Location) { ... }

    fn visit_substs(&mut self, substs: &&'tcx Substs<'tcx>, _: Location) { ... }

    fn visit_closure_substs(
        &mut self, 
        substs: &ClosureSubsts<'tcx>, 
        _: Location
    ) { ... }

    fn visit_generator_interior(
        &mut self, 
        interior: &GeneratorInterior<'tcx>, 
        _: Location
    ) { ... }

    fn visit_const_int(&mut self, const_int: &ConstInt, _: Location) { ... }

    fn visit_const_usize(&mut self, const_usize: &ConstUsize, _: Location) { ... }

    fn visit_local_decl(&mut self, local_decl: &LocalDecl<'tcx>) { ... }

    fn visit_local(
        &mut self, 
        _local: &Local, 
        _context: LvalueContext<'tcx>, 
        _location: Location
    ) { ... }

    fn visit_visibility_scope(&mut self, scope: &VisibilityScope) { ... }

    fn super_mir(&mut self, mir: &Mir<'tcx>) { ... }

    fn super_basic_block_data(
        &mut self, 
        block: BasicBlock, 
        data: &BasicBlockData<'tcx>
    ) { ... }

    fn super_visibility_scope_data(&mut self, scope_data: &VisibilityScopeData) { ... }

    fn super_statement(
        &mut self, 
        block: BasicBlock, 
        statement: &Statement<'tcx>, 
        location: Location
    ) { ... }

    fn super_assign(
        &mut self, 
        _block: BasicBlock, 
        lvalue: &Lvalue<'tcx>, 
        rvalue: &Rvalue<'tcx>, 
        location: Location
    ) { ... }

    fn super_terminator(
        &mut self, 
        block: BasicBlock, 
        terminator: &Terminator<'tcx>, 
        location: Location
    ) { ... }

    fn super_terminator_kind(
        &mut self, 
        block: BasicBlock, 
        kind: &TerminatorKind<'tcx>, 
        source_location: Location
    ) { ... }

    fn super_assert_message(
        &mut self, 
        msg: &AssertMessage<'tcx>, 
        location: Location
    ) { ... }

    fn super_rvalue(&mut self, rvalue: &Rvalue<'tcx>, location: Location) { ... }

    fn super_operand(&mut self, operand: &Operand<'tcx>, location: Location) { ... }

    fn super_lvalue(
        &mut self, 
        lvalue: &Lvalue<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn super_static(
        &mut self, 
        static_: &Static<'tcx>, 
        _context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn super_projection(
        &mut self, 
        proj: &LvalueProjection<'tcx>, 
        context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn super_projection_elem(
        &mut self, 
        proj: &LvalueElem<'tcx>, 
        _context: LvalueContext<'tcx>, 
        location: Location
    ) { ... }

    fn super_local_decl(&mut self, local_decl: &LocalDecl<'tcx>) { ... }

    fn super_visibility_scope(&mut self, _scope: &VisibilityScope) { ... }

    fn super_branch(&mut self, _source: BasicBlock, _target: BasicBlock) { ... }

    fn super_constant(&mut self, constant: &Constant<'tcx>, location: Location) { ... }

    fn super_literal(&mut self, literal: &Literal<'tcx>, location: Location) { ... }

    fn super_def_id(&mut self, _def_id: &DefId) { ... }

    fn super_span(&mut self, _span: &Span) { ... }

    fn super_source_info(&mut self, source_info: &SourceInfo) { ... }

    fn super_ty(&mut self, _ty: &Ty<'tcx>) { ... }

    fn super_region(&mut self, _region: &Region<'tcx>) { ... }

    fn super_const(&mut self, _const: &&'tcx Const<'tcx>) { ... }

    fn super_substs(&mut self, _substs: &&'tcx Substs<'tcx>) { ... }

    fn super_generator_interior(&mut self, _interior: &GeneratorInterior<'tcx>) { ... }

    fn super_closure_substs(&mut self, _substs: &ClosureSubsts<'tcx>) { ... }

    fn super_const_int(&mut self, _const_int: &ConstInt) { ... }

    fn super_const_usize(&mut self, _const_usize: &ConstUsize) { ... }

    fn visit_location(&mut self, mir: &Mir<'tcx>, location: Location) { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn visit_mir(&mut self, mir: &Mir<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_basic_block_data(
&mut self,
block: BasicBlock,
data: &BasicBlockData<'tcx>
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_visibility_scope_data(&mut self, scope_data: &VisibilityScopeData)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_statement(
&mut self,
block: BasicBlock,
statement: &Statement<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_assign(
&mut self,
block: BasicBlock,
lvalue: &Lvalue<'tcx>,
rvalue: &Rvalue<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_terminator(
&mut self,
block: BasicBlock,
terminator: &Terminator<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_terminator_kind(
&mut self,
block: BasicBlock,
kind: &TerminatorKind<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_assert_message(
&mut self,
msg: &AssertMessage<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_rvalue(&mut self, rvalue: &Rvalue<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_operand(&mut self, operand: &Operand<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_lvalue(
&mut self,
lvalue: &Lvalue<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_static(
&mut self,
static_: &Static<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_projection(
&mut self,
lvalue: &LvalueProjection<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_projection_elem(
&mut self,
lvalue: &LvalueElem<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_branch(&mut self, source: BasicBlock, target: BasicBlock)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_constant(&mut self, constant: &Constant<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_literal(&mut self, literal: &Literal<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_def_id(&mut self, def_id: &DefId, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_span(&mut self, span: &Span)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_source_info(&mut self, source_info: &SourceInfo)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty(&mut self, ty: &Ty<'tcx>, _: Lookup)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_region(&mut self, region: &Region<'tcx>, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_const(&mut self, constant: &&'tcx Const<'tcx>, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_substs(&mut self, substs: &&'tcx Substs<'tcx>, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_closure_substs(&mut self, substs: &ClosureSubsts<'tcx>, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_generator_interior(
&mut self,
interior: &GeneratorInterior<'tcx>,
_: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_const_int(&mut self, const_int: &ConstInt, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_const_usize(&mut self, const_usize: &ConstUsize, _: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_local_decl(&mut self, local_decl: &LocalDecl<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_local(
&mut self,
_local: &Local,
_context: LvalueContext<'tcx>,
_location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_visibility_scope(&mut self, scope: &VisibilityScope)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_mir(&mut self, mir: &Mir<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_basic_block_data(
&mut self,
block: BasicBlock,
data: &BasicBlockData<'tcx>
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_visibility_scope_data(&mut self, scope_data: &VisibilityScopeData)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_statement(
&mut self,
block: BasicBlock,
statement: &Statement<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_assign(
&mut self,
_block: BasicBlock,
lvalue: &Lvalue<'tcx>,
rvalue: &Rvalue<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_terminator(
&mut self,
block: BasicBlock,
terminator: &Terminator<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_terminator_kind(
&mut self,
block: BasicBlock,
kind: &TerminatorKind<'tcx>,
source_location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_assert_message(
&mut self,
msg: &AssertMessage<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_rvalue(&mut self, rvalue: &Rvalue<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_operand(&mut self, operand: &Operand<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_lvalue(
&mut self,
lvalue: &Lvalue<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_static(
&mut self,
static_: &Static<'tcx>,
_context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_projection(
&mut self,
proj: &LvalueProjection<'tcx>,
context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_projection_elem(
&mut self,
proj: &LvalueElem<'tcx>,
_context: LvalueContext<'tcx>,
location: Location
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_local_decl(&mut self, local_decl: &LocalDecl<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_visibility_scope(&mut self, _scope: &VisibilityScope)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_branch(&mut self, _source: BasicBlock, _target: BasicBlock)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_constant(&mut self, constant: &Constant<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_literal(&mut self, literal: &Literal<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_def_id(&mut self, _def_id: &DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_span(&mut self, _span: &Span)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_source_info(&mut self, source_info: &SourceInfo)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_ty(&mut self, _ty: &Ty<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_region(&mut self, _region: &Region<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_const(&mut self, _const: &&'tcx Const<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_substs(&mut self, _substs: &&'tcx Substs<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_generator_interior(&mut self, _interior: &GeneratorInterior<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_closure_substs(&mut self, _substs: &ClosureSubsts<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_const_int(&mut self, _const_int: &ConstInt)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn super_const_usize(&mut self, _const_usize: &ConstUsize)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_location(&mut self, mir: &Mir<'tcx>, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?