Module rustc::session::config

Unstable (rustc_private)

Contains infrastructure for configuring the compiler, including parsing command line options.

Reexports

pub use self::EntryFnType::*;

pub use self::CrateType::*;

pub use self::Passes::*;

pub use self::DebugInfoLevel::*;

Modules

nightly_options 	 [Unstable]

Structs

CodegenOptions 	 [Unstable]

Config 	 [Unstable]

DebuggingOptions 	 [Unstable]

Options 	 [Unstable]

OutputFilenames 	 [Unstable]

RustcOptGroup 	 [Unstable]

Enums

CrateType 	 [Unstable]

DebugInfoLevel 	 [Unstable]

EntryFnType 	 [Unstable]

ErrorOutputType 	 [Unstable]

Input 	 [Unstable]

OptLevel 	 [Unstable]

OptionStability 	 [Unstable]

OutputType 	 [Unstable]

PanicStrategy 	 [Unstable]

Passes 	 [Unstable]

PrintRequest 	 [Unstable]

Constants

CG_OPTIONS 	 [Unstable]

DB_OPTIONS 	 [Unstable]

Functions

append_configuration 	 [Unstable]

basic_codegen_options 	 [Unstable]

basic_debugging_options 	 [Unstable]

basic_options 	 [Unstable]
Some reasonable defaults

build_codegen_options 	 [Unstable]

build_configuration 	 [Unstable]

build_debugging_options 	 [Unstable]

build_session_options 	 [Unstable]

build_target_config 	 [Unstable]

default_configuration 	 [Unstable]

default_lib_output 	 [Unstable]

get_unstable_features_setting 	 [Unstable]

host_triple 	 [Unstable]

parse_cfgspecs 	 [Unstable]

parse_crate_types_from_list 	 [Unstable]

rustc_optgroups 	 [Unstable]
Returns all rustc command line options, including metadata for each option, such as whether the option is part of the stable long-term interface for rustc.

rustc_short_optgroups 	 [Unstable]
Returns the "short" subset of the rustc command line options, including metadata for each option, such as whether the option is part of the stable long-term interface for rustc.

Type Definitions

CodegenSetter 	 [Unstable]

DebuggingSetter 	 [Unstable]