Struct rustc::session::config::Options

pub struct Options {
    pub crate_types: Vec<CrateType>,
    pub optimize: OptLevel,
    pub debug_assertions: bool,
    pub debuginfo: DebugInfoLevel,
    pub lint_opts: Vec<(String, Level)>,
    pub lint_cap: Option<Level>,
    pub describe_lints: bool,
    pub output_types: HashMap<OutputType, Option<PathBuf>>,
    pub search_paths: SearchPaths,
    pub libs: Vec<(String, NativeLibraryKind)>,
    pub maybe_sysroot: Option<PathBuf>,
    pub target_triple: String,
    pub cfg: CrateConfig,
    pub test: bool,
    pub parse_only: bool,
    pub no_trans: bool,
    pub error_format: ErrorOutputType,
    pub treat_err_as_bug: bool,
    pub continue_parse_after_error: bool,
    pub mir_opt_level: usize,
    pub incremental: Option<PathBuf>,
    pub no_analysis: bool,
    pub debugging_opts: DebuggingOptions,
    pub prints: Vec<PrintRequest>,
    pub cg: CodegenOptions,
    pub externs: HashMap<String, Vec<String>>,
    pub crate_name: Option<String>,
    pub alt_std_name: Option<String>,
    pub unstable_features: UnstableFeatures,
}

Unstable (rustc_private)

Fields
crate_types: Vec<CrateType>
Unstable (rustc_private)
optimize: OptLevel
Unstable (rustc_private)
debug_assertions: bool
Unstable (rustc_private)
debuginfo: DebugInfoLevel
Unstable (rustc_private)
lint_opts: Vec<(String, Level)>
Unstable (rustc_private)
lint_cap: Option<Level>
Unstable (rustc_private)
describe_lints: bool
Unstable (rustc_private)
output_types: HashMap<OutputType, Option<PathBuf>>
Unstable (rustc_private)
search_paths: SearchPaths
Unstable (rustc_private)
libs: Vec<(String, NativeLibraryKind)>
Unstable (rustc_private)
maybe_sysroot: Option<PathBuf>
Unstable (rustc_private)
target_triple: String
Unstable (rustc_private)
cfg: CrateConfig
Unstable (rustc_private)
test: bool
Unstable (rustc_private)
parse_only: bool
Unstable (rustc_private)
no_trans: bool
Unstable (rustc_private)
error_format: ErrorOutputType
Unstable (rustc_private)
treat_err_as_bug: bool
Unstable (rustc_private)
continue_parse_after_error: bool
Unstable (rustc_private)
mir_opt_level: usize
Unstable (rustc_private)
incremental: Option<PathBuf>
Unstable (rustc_private)

if Some, enable incremental compilation, using the given directory to store intermediate results

no_analysis: bool
Unstable (rustc_private)
debugging_opts: DebuggingOptions
Unstable (rustc_private)
prints: Vec<PrintRequest>
Unstable (rustc_private)
cg: CodegenOptions
Unstable (rustc_private)
externs: HashMap<String, Vec<String>>
Unstable (rustc_private)
crate_name: Option<String>
Unstable (rustc_private)
alt_std_name: Option<String>
Unstable (rustc_private)

An optional name to use as the crate for std during std injection, written extern crate std = "name". Default to "std". Used by out-of-tree drivers.

unstable_features: UnstableFeatures
Unstable (rustc_private)

Indicates how the compiler should treat unstable features

Methods

impl Options

fn build_dep_graph(&self) -> bool

Unstable (rustc_private)

True if there is a reason to build the dep graph.

fn single_codegen_unit(&self) -> bool

Unstable (rustc_private)