Struct rustc::session::config::RustcOptGroup

pub struct RustcOptGroup {
    pub apply: Box<Fn(&mut Options) -> &mut Options>,
    pub name: &'static str,
    pub stability: OptionStability,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
apply: Box<Fn(&mut Options) -> &mut Options>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

name: &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

stability: OptionStability

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl RustcOptGroup

pub fn is_stable(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn stable<F>(name: &'static str, f: F) -> RustcOptGroup where
F: Fn(&mut Options) -> &mut Options + 'static,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn unstable<F>(name: &'static str, f: F) -> RustcOptGroup where
F: Fn(&mut Options) -> &mut Options + 'static,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?