Enum rustc::session::IncrCompSession

pub enum IncrCompSession {
    NotInitialized,
    Active {
        session_directory: PathBuf,
        lock_file: Lock,
        load_dep_graph: bool,
    },
    Finalized {
        session_directory: PathBuf,
    },
    InvalidBecauseOfErrors {
        session_directory: PathBuf,
    },
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Holds data on the current incremental compilation session, if there is one.

Variants
NotInitialized

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This is the state the session will be in until the incr. comp. dir is needed.

Active

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This is the state during which the session directory is private and can be modified.

Fields of Active

session_directory: PathBuf

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

lock_file: Lock

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

load_dep_graph: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Finalized

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This is the state after the session directory has been finalized. In this state, the contents of the directory must not be modified any more.

Fields of Finalized

session_directory: PathBuf

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

InvalidBecauseOfErrors

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This is an error state that is reached when some compilation error has occurred. It indicates that the contents of the session directory must not be used, since they might be invalid.

Fields of InvalidBecauseOfErrors

session_directory: PathBuf

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?