Struct rustc::session::PerfStats

pub struct PerfStats {
    pub svh_time: Cell<Duration>,
    pub incr_comp_hashes_time: Cell<Duration>,
    pub incr_comp_hashes_count: Cell<u64>,
    pub incr_comp_bytes_hashed: Cell<u64>,
    pub symbol_hash_time: Cell<Duration>,
    pub decode_def_path_tables_time: Cell<Duration>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
svh_time: Cell<Duration>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

incr_comp_hashes_time: Cell<Duration>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

incr_comp_hashes_count: Cell<u64>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

incr_comp_bytes_hashed: Cell<u64>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

symbol_hash_time: Cell<Duration>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

decode_def_path_tables_time: Cell<Duration>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?