Enum rustc::traits::ObligationCauseCode

pub enum ObligationCauseCode<'tcx> {
    MiscObligation,
    SliceOrArrayElem,
    TupleElem,
    ProjectionWf(ProjectionTy<'tcx>),
    ItemObligation(DefId),
    ReferenceOutlivesReferent(Ty<'tcx>),
    ObjectCastObligation(Ty<'tcx>),
    AssignmentLhsSized,
    StructInitializerSized,
    VariableType(NodeId),
    ReturnType,
    RepeatVec,
    ClosureCapture(NodeId, Span, BuiltinBound),
    FieldSized,
    SharedStatic,
    BuiltinDerivedObligation(DerivedObligationCause<'tcx>),
    ImplDerivedObligation(DerivedObligationCause<'tcx>),
    CompareImplMethodObligation,
}

Unstable (rustc_private)

Variants

MiscObligation
Unstable (rustc_private)

Not well classified or should be obvious from span.

SliceOrArrayElem
Unstable (rustc_private)

A slice or array is WF only if T: Sized

TupleElem
Unstable (rustc_private)

A tuple is WF only if its middle elements are Sized

ProjectionWf
Unstable (rustc_private)

This is the trait reference from the given projection

ItemObligation
Unstable (rustc_private)

In an impl of trait X for type Y, type Y must also implement all supertraits of X.

ReferenceOutlivesReferent
Unstable (rustc_private)

A type like &'a T is WF only if T: 'a.

ObjectCastObligation
Unstable (rustc_private)

Obligation incurred due to an object cast.

AssignmentLhsSized
Unstable (rustc_private)

Various cases where expressions must be sized/copy/etc:

StructInitializerSized
Unstable (rustc_private)

VariableType
Unstable (rustc_private)

ReturnType
Unstable (rustc_private)

RepeatVec
Unstable (rustc_private)

ClosureCapture
Unstable (rustc_private)

FieldSized
Unstable (rustc_private)

SharedStatic
Unstable (rustc_private)

BuiltinDerivedObligation
Unstable (rustc_private)

ImplDerivedObligation
Unstable (rustc_private)

CompareImplMethodObligation
Unstable (rustc_private)