Enum rustc::traits::ProjectionMode

pub enum ProjectionMode {
    Topmost,
    AnyFinal,
    Any,
}

Unstable (rustc_private)

Depending on the stage of compilation, we want projection to be more or less conservative.

Variants

Topmost
Unstable (rustc_private)

FIXME (#32205) At coherence-checking time, we're still constructing the specialization graph, and thus we only project project non-default associated types that are defined directly in the applicable impl. (This behavior should be improved over time, to allow for successful projections modulo cycles between different impls).

Here's an example that will fail due to the restriction:

trait Assoc {
    type Output;
}

impl<T> Assoc for T {
    type Output = bool;
}

impl Assoc for u8 {} // <- inherits the non-default type from above

trait Foo {}
impl Foo for u32 {}
impl Foo for <u8 as Assoc>::Output {}  // <- this projection will fail

The projection would succeed if Output had been defined directly in the impl for u8.

AnyFinal
Unstable (rustc_private)

At type-checking time, we refuse to project any associated type that is marked default. Non-default ("final") types are always projected. This is necessary in general for soundness of specialization. However, we could allow projections in fully-monomorphic cases. We choose not to, because we prefer for default type to force the type definition to be treated abstractly by any consumers of the impl. Concretely, that means that the following example will fail to compile:

trait Assoc {
    type Output;
}

impl<T> Assoc for T {
    default type Output = bool;
}

fn main() {
    let <() as Assoc>::Output = true;
}

Any
Unstable (rustc_private)

At trans time, all projections will succeed.

Methods

impl ProjectionMode

fn is_topmost(&self) -> bool

Unstable (rustc_private)

fn is_any_final(&self) -> bool

Unstable (rustc_private)

fn is_any(&self) -> bool

Unstable (rustc_private)