Enum rustc::traits::Reveal

pub enum Reveal {
    ExactMatch,
    NotSpecializable,
    All,
}

Unstable (rustc_private)

Depending on the stage of compilation, we want projection to be more or less conservative.

Variants
ExactMatch
Unstable (rustc_private)

FIXME (#32205) At coherence-checking time, we're still constructing the specialization graph, and thus we only project non-default associated types that are defined directly in the applicable impl. (This behavior should be improved over time, to allow for successful projections modulo cycles between different impls).

Here's an example that will fail due to the restriction:

trait Assoc {
    type Output;
}

impl<T> Assoc for T {
    type Output = bool;
}

impl Assoc for u8 {} // <- inherits the non-default type from above

trait Foo {}
impl Foo for u32 {}
impl Foo for <u8 as Assoc>::Output {}  // <- this projection will fail

The projection would succeed if Output had been defined directly in the impl for u8.

NotSpecializable
Unstable (rustc_private)

At type-checking time, we refuse to project any associated type that is marked default. Non-default ("final") types are always projected. This is necessary in general for soundness of specialization. However, we could allow projections in fully-monomorphic cases. We choose not to, because we prefer for default type to force the type definition to be treated abstractly by any consumers of the impl. Concretely, that means that the following example will fail to compile:

trait Assoc {
    type Output;
}

impl<T> Assoc for T {
    default type Output = bool;
}

fn main() {
    let <() as Assoc>::Output = true;
}

All
Unstable (rustc_private)

At trans time, all monomorphic projections will succeed. Also, impl Trait is normalized to the concrete type, which has to be already collected by type-checking.

NOTE: As impl Trait's concrete type should never be observable directly by the user, Reveal::All should not be used by checks which may expose type equality or type contents to the user. There are some exceptions, e.g. around OIBITS and transmute-checking, which expose some details, but not the whole concrete type of the impl Trait.