Enum rustc::traits::SelectionError

pub enum SelectionError<'tcx> {
    Unimplemented,
    OutputTypeParameterMismatch(PolyTraitRef<'tcx>, PolyTraitRef<'tcx>, TypeError<'tcx>),
    TraitNotObjectSafe(DefId),
    ConstEvalFailure(ConstEvalErr<'tcx>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
Unimplemented

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

OutputTypeParameterMismatch(PolyTraitRef<'tcx>, PolyTraitRef<'tcx>, TypeError<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

TraitNotObjectSafe(DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

ConstEvalFailure(ConstEvalErr<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?