Function rustc::traits::overlapping_impls

pub fn overlapping_impls<'cx, 'gcx, 'tcx>(
    infcx: &InferCtxt<'cx, 'gcx, 'tcx>, 
    impl1_def_id: DefId, 
    impl2_def_id: DefId
) -> Option<ImplHeader<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If there are types that satisfy both impls, returns a suitably-freshened ImplHeader with those types substituted