Module rustc::traits::specialization_graph

Unstable (rustc_private)

Structs

Ancestors 	 [Unstable]

ConstDefs 	 [Unstable]

FnDefs 	 [Unstable]

Graph 	 [Unstable]
A per-trait graph of impls in specialization order. At the moment, this graph forms a tree rooted with the trait itself, with all other nodes representing impls, and parent-child relationships representing specializations.

NodeItem 	 [Unstable]

TypeDefs 	 [Unstable]

Enums

Node 	 [Unstable]
A node in the specialization graph is either an impl or a trait definition; either can serve as a source of item definitions. There is always exactly one trait definition node: the root.

NodeItems 	 [Unstable]
An iterator over the items defined within a trait or impl.

Functions

ancestors 	 [Unstable]
Walk up the specialization ancestors of a given impl, starting with that impl itself.