Enum rustc::traits::specialization_graph::Node

pub enum Node {
    Impl(DefId),
    Trait(DefId),
}

Unstable (rustc_private)

A node in the specialization graph is either an impl or a trait definition; either can serve as a source of item definitions. There is always exactly one trait definition node: the root.

Variants

Impl
Unstable (rustc_private)

Trait
Unstable (rustc_private)

Methods

impl Node

fn is_from_trait(&self) -> bool

Unstable (rustc_private)

fn items<'a, 'tcx>(&self, tcx: &'a TyCtxt<'tcx>) -> NodeItems<'a, 'tcx>

Unstable (rustc_private)

Iterate over the items defined directly by the given (impl or trait) node.

fn def_id(&self) -> DefId

Unstable (rustc_private)

impl Node

fn is_from_trait(&self) -> bool

Unstable (rustc_private)

fn items<'a, 'tcx>(&self, tcx: &'a TyCtxt<'tcx>) -> NodeItems<'a, 'tcx>

Unstable (rustc_private)

Iterate over the items defined directly by the given (impl or trait) node.

fn def_id(&self) -> DefId

Unstable (rustc_private)