Struct rustc::traits::FulfillmentContext

pub struct FulfillmentContext<'tcx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

The fulfillment context is used to drive trait resolution. It consists of a list of obligations that must be (eventually) satisfied. The job is to track which are satisfied, which yielded errors, and which are still pending. At any point, users can call select_where_possible, and the fulfilment context will try to do selection, retaining only those obligations that remain ambiguous. This may be helpful in pushing type inference along. Once all type inference constraints have been generated, the method select_all_or_error can be used to report any remaining ambiguous cases as errors.

Methods

impl<'a, 'gcx, 'tcx> FulfillmentContext<'tcx>

fn new() -> FulfillmentContext<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

Creates a new fulfillment context.

fn normalize_projection_type(&mut self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>,
projection_ty: ProjectionTy<'tcx>,
cause: ObligationCause<'tcx>)
-> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

"Normalize" a projection type <SomeType as SomeTrait>::X by creating a fresh type variable $0 as well as a projection predicate <SomeType as SomeTrait>::X == $0. When the inference engine runs, it will attempt to find an impl of SomeTrait or a where clause that lets us unify $0 with something concrete. If this fails, we'll unify $0 with projection_ty again.

fn register_bound(&mut self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>,
ty: Ty<'tcx>,
def_id: DefId,
cause: ObligationCause<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

fn register_region_obligation(&mut self,
t_a: Ty<'tcx>,
r_b: &'tcx Region,
cause: ObligationCause<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

fn register_predicate_obligation(&mut self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>,
obligation: PredicateObligation<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

fn region_obligations(&self, body_id: NodeId) -> &[RegionObligation<'tcx>]

🔬 This is a nightly-only experimental API. (rustc_private)

fn select_all_or_error(&mut self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>)
-> Result<(), Vec<FulfillmentError<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

fn select_where_possible(&mut self,
infcx: &InferCtxt<'a, 'gcx, 'tcx>)
-> Result<(), Vec<FulfillmentError<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

fn pending_obligations(&self) -> Vec<PendingPredicateObligation<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

fn take_deferred_obligations(&mut self) -> Vec<DeferredObligation<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)