Enum rustc::ty::cast::CastTy

pub enum CastTy<'tcx> {
    Int(IntTy),
    Float,
    FnPtr,
    Ptr(&'tcx TypeAndMut<'tcx>),
    RPtr(&'tcx TypeAndMut<'tcx>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
Int(IntTy)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Various types that are represented as ints and handled mostly in the same way, merged for easier matching.

Float

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Floating-Point types

FnPtr

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Function Pointers

Ptr(&'tcx TypeAndMut<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Raw pointers

RPtr(&'tcx TypeAndMut<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

References

Methods

impl<'tcx> CastTy<'tcx>

pub fn from_ty(t: Ty<'tcx>) -> Option<CastTy<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?