Module rustc::ty::codec

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Constants

SHORTHAND_OFFSET 	 [
Experimental
]
The shorthand encoding uses an enum's variant index usize and is offset by this value so it never matches a real variant. This offset is also chosen so that the first byte is never < 0x80.

Traits

EncodableWithShorthand 	 [
Experimental
]

TyDecoder 	 [
Experimental
]

TyEncoder 	 [
Experimental
]

Functions

decode_adt_def 	 [
Experimental
]

decode_byte_array 	 [
Experimental
]

decode_cnum 	 [
Experimental
]

decode_const 	 [
Experimental
]

decode_existential_predicate_slice 	 [
Experimental
]

decode_predicates 	 [
Experimental
]

decode_region 	 [
Experimental
]

decode_substs 	 [
Experimental
]

decode_ty 	 [
Experimental
]

decode_ty_slice 	 [
Experimental
]

encode_predicates 	 [
Experimental
]

encode_with_shorthand 	 [
Experimental
]
Encode the given value or a previously cached shorthand.