Enum rustc::ty::Predicate

pub enum Predicate<'tcx> {
    Trait(PolyTraitPredicate<'tcx>),
    Equate(PolyEquatePredicate<'tcx>),
    RegionOutlives(PolyRegionOutlivesPredicate<'tcx>),
    TypeOutlives(PolyTypeOutlivesPredicate<'tcx>),
    Projection(PolyProjectionPredicate<'tcx>),
    WellFormed(Ty<'tcx>),
    ObjectSafe(DefId),
    ClosureKind(DefId, ClosureKind),
}

This is an experimental API. (rustc_private)

Variants
Trait(PolyTraitPredicate<'tcx>)

This is an experimental API. (rustc_private)

Corresponds to where Foo : Bar<A,B,C>. Foo here would be the Self type of the trait reference and A, B, and C would be the type parameters.

Equate(PolyEquatePredicate<'tcx>)

This is an experimental API. (rustc_private)

where T1 == T2.

RegionOutlives(PolyRegionOutlivesPredicate<'tcx>)

This is an experimental API. (rustc_private)

where 'a : 'b

TypeOutlives(PolyTypeOutlivesPredicate<'tcx>)

This is an experimental API. (rustc_private)

where T : 'a

Projection(PolyProjectionPredicate<'tcx>)

This is an experimental API. (rustc_private)

where ::Name == X, approximately. See ProjectionPredicate struct for details.

WellFormed(Ty<'tcx>)

This is an experimental API. (rustc_private)

no syntax: T WF

ObjectSafe(DefId)

This is an experimental API. (rustc_private)

trait must be object-safe

ClosureKind(DefId, ClosureKind)

This is an experimental API. (rustc_private)

No direct syntax. May be thought of as where T : FnFoo<...> for some substitutions ... and T being a closure type. Satisfied (or refuted) once we know the closure's kind.

Methods

impl<'a, 'gcx, 'tcx> Predicate<'tcx>

fn subst_supertrait(&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
trait_ref: &PolyTraitRef<'tcx>)
-> Predicate<'tcx>

This is an experimental API. (rustc_private)

Performs a substitution suitable for going from a poly-trait-ref to supertraits that must hold if that poly-trait-ref holds. This is slightly different from a normal substitution in terms of what happens with bound regions. See lengthy comment below for details.

impl<'tcx> Predicate<'tcx>

fn walk_tys(&self) -> IntoIter<Ty<'tcx>>

This is an experimental API. (rustc_private)

Iterates over the types in this predicate. Note that in all cases this is skipping over a binder, so late-bound regions with depth 0 are bound by the predicate.

fn to_opt_poly_trait_ref(&self) -> Option<PolyTraitRef<'tcx>>

This is an experimental API. (rustc_private)