Enum rustc::ty::TypeVariants

pub enum TypeVariants<'tcx> {
    TyBool,
    TyChar,
    TyInt(IntTy),
    TyUint(UintTy),
    TyFloat(FloatTy),
    TyAdt(AdtDef<'tcx>, &'tcx Substs<'tcx>),
    TyBox(Ty<'tcx>),
    TyStr,
    TyArray(Ty<'tcx>, usize),
    TySlice(Ty<'tcx>),
    TyRawPtr(TypeAndMut<'tcx>),
    TyRef(&'tcx Region, TypeAndMut<'tcx>),
    TyFnDef(DefId, &'tcx Substs<'tcx>, &'tcx BareFnTy<'tcx>),
    TyFnPtr(&'tcx BareFnTy<'tcx>),
    TyTrait(Box<TraitObject<'tcx>>),
    TyClosure(DefId, ClosureSubsts<'tcx>),
    TyNever,
    TyTuple(&'tcx [Ty<'tcx>]),
    TyProjection(ProjectionTy<'tcx>),
    TyAnon(DefId, &'tcx Substs<'tcx>),
    TyParam(ParamTy),
    TyInfer(InferTy),
    TyError,
}

Unstable (rustc_private)

Variants
TyBool
Unstable (rustc_private)

The primitive boolean type. Written as bool.

TyChar
Unstable (rustc_private)

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

TyInt(IntTy)
Unstable (rustc_private)

A primitive signed integer type. For example, i32.

TyUint(UintTy)
Unstable (rustc_private)

A primitive unsigned integer type. For example, u32.

TyFloat(FloatTy)
Unstable (rustc_private)

A primitive floating-point type. For example, f64.

TyAdt(AdtDef<'tcx>, &'tcx Substs<'tcx>)
Unstable (rustc_private)

Structures, enumerations and unions.

Substs here, possibly against intuition, may contain TyParams. That is, even after substitution it is possible that there are type variables. This happens when the TyAdt corresponds to an ADT definition and not a concrete use of it.

TyBox(Ty<'tcx>)
Unstable (rustc_private)

Box<T>; this is nominally a struct in the documentation, but is special-cased internally. For example, it is possible to implicitly move the contents of a box out of that box, and methods of any type can have type Box<Self>.

TyStr
Unstable (rustc_private)

The pointee of a string slice. Written as str.

TyArray(Ty<'tcx>, usize)
Unstable (rustc_private)

An array with the given length. Written as [T; n].

TySlice(Ty<'tcx>)
Unstable (rustc_private)

The pointee of an array slice. Written as [T].

TyRawPtr(TypeAndMut<'tcx>)
Unstable (rustc_private)

A raw pointer. Written as *mut T or *const T

TyRef(&'tcx Region, TypeAndMut<'tcx>)
Unstable (rustc_private)

A reference; a pointer with an associated lifetime. Written as &a mut T or &'a T.

TyFnDef(DefId, &'tcx Substs<'tcx>, &'tcx BareFnTy<'tcx>)
Unstable (rustc_private)

The anonymous type of a function declaration/definition. Each function has a unique type.

TyFnPtr(&'tcx BareFnTy<'tcx>)
Unstable (rustc_private)

A pointer to a function. Written as fn() -> i32. FIXME: This is currently also used to represent the callee of a method; see ty::MethodCallee etc.

TyTrait(Box<TraitObject<'tcx>>)
Unstable (rustc_private)

A trait, defined with trait.

TyClosure(DefId, ClosureSubsts<'tcx>)
Unstable (rustc_private)

The anonymous type of a closure. Used to represent the type of |a| a.

TyNever
Unstable (rustc_private)

The never type !

TyTuple(&'tcx [Ty<'tcx>])
Unstable (rustc_private)

A tuple type. For example, (i32, bool).

TyProjection(ProjectionTy<'tcx>)
Unstable (rustc_private)

The projection of an associated type. For example, <T as Trait<..>>::N.

TyAnon(DefId, &'tcx Substs<'tcx>)
Unstable (rustc_private)

Anonymized (impl Trait) type found in a return type. The DefId comes from the impl Trait ast::Ty node, and the substitutions are for the generics of the function in question. After typeck, the concrete type can be found in the tcache map.

TyParam(ParamTy)
Unstable (rustc_private)

A type parameter; for example, T in `fn f(x: T) {}

TyInfer(InferTy)
Unstable (rustc_private)

A type variable used during type-checking.

TyError
Unstable (rustc_private)

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages.