Trait rustc::ty::fold::TypeFoldable

pub trait TypeFoldable<'tcx>: Debug + Clone {
    fn super_fold_with<'gcx: 'tcx, F: TypeFolder<'gcx, 'tcx>>(&self,
                                                              folder: &mut F)
                                                              -> Self;
    fn super_visit_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool;

    fn fold_with<'gcx: 'tcx, F: TypeFolder<'gcx, 'tcx>>(&self,
                                                        folder: &mut F)
                                                        -> Self { ... }
    fn visit_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool { ... }
    fn has_regions_escaping_depth(&self, depth: u32) -> bool { ... }
    fn has_escaping_regions(&self) -> bool { ... }
    fn has_type_flags(&self, flags: TypeFlags) -> bool { ... }
    fn has_projection_types(&self) -> bool { ... }
    fn references_error(&self) -> bool { ... }
    fn has_param_types(&self) -> bool { ... }
    fn has_self_ty(&self) -> bool { ... }
    fn has_infer_types(&self) -> bool { ... }
    fn needs_infer(&self) -> bool { ... }
    fn needs_subst(&self) -> bool { ... }
    fn has_re_skol(&self) -> bool { ... }
    fn has_closure_types(&self) -> bool { ... }
    fn has_erasable_regions(&self) -> bool { ... }
    fn is_normalized_for_trans(&self) -> bool { ... }
    fn is_global(&self) -> bool { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

The TypeFoldable trait is implemented for every type that can be folded. Basically, every type that has a corresponding method in TypeFolder.

Required Methods

fn super_fold_with<'gcx: 'tcx, F: TypeFolder<'gcx, 'tcx>>(&self,
folder: &mut F)
-> Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn super_visit_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Provided Methods

fn fold_with<'gcx: 'tcx, F: TypeFolder<'gcx, 'tcx>>(&self,
folder: &mut F)
-> Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn visit_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_regions_escaping_depth(&self, depth: u32) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_escaping_regions(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_type_flags(&self, flags: TypeFlags) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_projection_types(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn references_error(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_param_types(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_self_ty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_infer_types(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn needs_infer(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn needs_subst(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_re_skol(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_closure_types(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_erasable_regions(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_normalized_for_trans(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_global(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Indicates whether this value references only 'global' types/lifetimes that are the same regardless of what fn we are in. This is used for caching. Errs on the side of returning false.