Trait rustc::ty::fold::TypeVisitor

pub trait TypeVisitor<'tcx>: Sized {
    fn visit_binder<T: TypeFoldable<'tcx>>(&mut self, t: &Binder<T>) -> bool { ... }

    fn visit_ty(&mut self, t: Ty<'tcx>) -> bool { ... }

    fn visit_region(&mut self, r: Region<'tcx>) -> bool { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn visit_binder<T: TypeFoldable<'tcx>>(&mut self, t: &Binder<T>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_ty(&mut self, t: Ty<'tcx>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visit_region(&mut self, r: Region<'tcx>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?