Module rustc::ty::item_path

Unstable (rustc_private)

Enums

RootMode 	 [Unstable]

Traits

ItemPathBuffer 	 [Unstable]
Unifying Trait for different kinds of item paths we might construct. The basic interface is that components get pushed: the instance can also customize how we handle the root of a crate.

Functions

characteristic_def_id_of_type 	 [Unstable]
As a heuristic, when we see an impl, if we see that the 'self-type' is a type defined in the same module as the impl, we can omit including the path to the impl itself. This function tries to find a "characteristic def-id" for a type. It's just a heuristic so it makes some questionable decisions and we may want to adjust it later.