Function rustc::ty::item_path::with_forced_impl_filename_line

pub fn with_forced_impl_filename_line<F: FnOnce() -> R, R>(f: F) -> R

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Force us to name impls with just the filename/line number. We normally try to use types. But at some points, notably while printing cycle errors, this can result in extra or suboptimal error output, so this variable disables that check.