Enum rustc::ty::layout::Integer

pub enum Integer {
    I1,
    I8,
    I16,
    I32,
    I64,
}

Unstable (rustc_private)

Integers, also used for enum discriminants.

Variants
I1
Unstable (rustc_private)
I8
Unstable (rustc_private)
I16
Unstable (rustc_private)
I32
Unstable (rustc_private)
I64
Unstable (rustc_private)

Methods

impl Integer

fn fit_signed(x: i64) -> Integer

Unstable (rustc_private)

Find the smallest Integer type which can represent the signed value.

fn fit_unsigned(x: u64) -> Integer

Unstable (rustc_private)

Find the smallest Integer type which can represent the unsigned value.

fn from_attr(dl: &TargetDataLayout, ity: IntType) -> Integer

Unstable (rustc_private)

Get the Integer type from an attr::IntType.

fn repr_discr(tcx: TyCtxt, hint: ReprAttr, min: i64, max: i64) -> (Integer, bool)

Unstable (rustc_private)

Find the appropriate Integer type and signedness for the given signed discriminant range and #[repr] attribute. N.B.: u64 values above i64::MAX will be treated as signed, but that shouldn't affect anything, other than maybe debuginfo.