Enum rustc::ty::layout::SizeSkeleton

pub enum SizeSkeleton<'tcx> {
    Known(Size),
    Pointer {
        non_zero: bool,
        tail: Ty<'tcx>,
    },
}

Unstable (rustc_private)

Type size "skeleton", i.e. the only information determining a type's size. While this is conservative, (aside from constant sizes, only pointers, newtypes thereof and null pointer optimized enums are allowed), it is enough to statically check common usecases of transmute.

Variants
Known(Size)
Unstable (rustc_private)

Any statically computable Layout.

Pointer
Unstable (rustc_private)

A potentially-fat pointer.

Fields

non_zero: bool
Unstable (rustc_private)

tail: Ty<'tcx>
Unstable (rustc_private)

Methods

impl<'a, 'gcx, 'tcx> SizeSkeleton<'gcx>

fn compute(ty: Ty<'gcx>, infcx: &InferCtxt<'a, 'gcx, 'tcx>) -> Result<SizeSkeleton<'gcx>, LayoutError<'gcx>>

Unstable (rustc_private)

fn same_size(self, other: SizeSkeleton) -> bool

Unstable (rustc_private)