Struct rustc::ty::layout::LayoutCx

pub struct LayoutCx<'a, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Combines a tcx with the parameter environment so that you can compute layout operations.

Methods

impl<'a, 'tcx> LayoutCx<'a, 'tcx>

pub fn new(tcx: TyCtxt<'a, 'tcx, 'tcx>, param_env: ParamEnv<'tcx>) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?