Struct rustc::ty::layout::Struct

pub struct Struct {
    pub align: Align,
    pub packed: bool,
    pub sized: bool,
    pub offset_after_field: Vec<Size>,
}

Unstable (rustc_private)

A structure, a product type in ADT terms.

Fields
align: Align
Unstable (rustc_private)
packed: bool
Unstable (rustc_private)

If true, no alignment padding is used.

sized: bool
Unstable (rustc_private)

If true, the size is exact, otherwise it's only a lower bound.

offset_after_field: Vec<Size>
Unstable (rustc_private)

Offsets for the first byte after each field. That is, field_offset(i) = offset_after_field[i - 1] and the whole structure's size is the last offset, excluding padding.

Methods

impl<'a, 'gcx, 'tcx> Struct

fn new(dl: &TargetDataLayout, packed: bool) -> Struct

Unstable (rustc_private)

fn extend<I>(&mut self, dl: &TargetDataLayout, fields: I, scapegoat: Ty<'gcx>) -> Result<(), LayoutError<'gcx>> where I: Iterator<Item=Result<&'a Layout, LayoutError<'gcx>>>

Unstable (rustc_private)

Extend the Struct with more fields.

fn min_size(&self) -> Size

Unstable (rustc_private)

Get the size without trailing alignment padding.

fn stride(&self) -> Size

Unstable (rustc_private)

Get the size with trailing aligment padding.

fn would_be_zero_sized<I>(dl: &TargetDataLayout, fields: I) -> Result<bool, LayoutError<'gcx>> where I: Iterator<Item=Result<&'a Layout, LayoutError<'gcx>>>

Unstable (rustc_private)

Determine whether a structure would be zero-sized, given its fields.

fn non_zero_field_in_type(infcx: &InferCtxt<'a, 'gcx, 'tcx>, ty: Ty<'gcx>) -> Result<Option<FieldPath>, LayoutError<'gcx>>

Unstable (rustc_private)

Find the path leading to a non-zero leaf field, starting from the given type and recursing through aggregates.

fn non_zero_field_path<I>(infcx: &InferCtxt<'a, 'gcx, 'tcx>, fields: I) -> Result<Option<FieldPath>, LayoutError<'gcx>> where I: Iterator<Item=Ty<'gcx>>

Unstable (rustc_private)

Find the path leading to a non-zero leaf field, starting from the given set of fields and recursing through aggregates.