Trait rustc::ty::layout::LayoutOf

pub trait LayoutOf<T> {
    type TyLayout;
    fn layout_of(self, ty: T) -> Self::TyLayout;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Associated Types

type TyLayout

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn layout_of(self, ty: T) -> Self::TyLayout

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?