Trait rustc::ty::layout::MaybeResult

pub trait MaybeResult<T> {
    fn from_ok(x: T) -> Self;

    fn map_same<F: FnOnce(T) -> T>(self, f: F) -> Self;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn from_ok(x: T) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn map_same<F: FnOnce(T) -> T>(self, f: F) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?