Module rustc::ty::maps::queries

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

adt_def 	 [
Experimental
]

adt_destructor 	 [
Experimental
]

adt_dtorck_constraint 	 [
Experimental
]

adt_sized_constraint 	 [
Experimental
]

associated_item 	 [
Experimental
]

associated_item_def_ids 	 [
Experimental
]

borrowck 	 [
Experimental
]

closure_kind 	 [
Experimental
]

coerce_unsized_info 	 [
Experimental
]

coherent_trait 	 [
Experimental
]

const_eval 	 [
Experimental
]

const_is_rvalue_promotable_to_static 	 [
Experimental
]

crate_inherent_impls 	 [
Experimental
]

crate_inherent_impls_overlap_check 	 [
Experimental
]

crate_variances 	 [
Experimental
]

def_span 	 [
Experimental
]

def_symbol_name 	 [
Experimental
]

deprecation 	 [
Experimental
]

describe_def 	 [
Experimental
]

dylib_dependency_formats 	 [
Experimental
]

extern_crate 	 [
Experimental
]

fn_arg_names 	 [
Experimental
]

fn_sig 	 [
Experimental
]

generics_of 	 [
Experimental
]

has_typeck_tables 	 [
Experimental
]

impl_parent 	 [
Experimental
]

impl_polarity 	 [
Experimental
]

impl_trait_ref 	 [
Experimental
]

inherent_impls 	 [
Experimental
]

is_allocator 	 [
Experimental
]

is_const_fn 	 [
Experimental
]

is_copy_raw 	 [
Experimental
]

is_default_impl 	 [
Experimental
]

is_exported_symbol 	 [
Experimental
]

is_foreign_item 	 [
Experimental
]

is_freeze_raw 	 [
Experimental
]

is_mir_available 	 [
Experimental
]

is_object_safe 	 [
Experimental
]

is_panic_runtime 	 [
Experimental
]

is_sized_raw 	 [
Experimental
]

item_attrs 	 [
Experimental
]

item_body_nested_bodies 	 [
Experimental
]

layout_raw 	 [
Experimental
]

lint_levels 	 [
Experimental
]

mir_borrowck 	 [
Experimental
]

mir_const 	 [
Experimental
]

mir_const_qualif 	 [
Experimental
]

mir_keys 	 [
Experimental
]

mir_shims 	 [
Experimental
]

mir_validated 	 [
Experimental
]

needs_drop_raw 	 [
Experimental
]

optimized_mir 	 [
Experimental
]

param_env 	 [
Experimental
]

predicates_of 	 [
Experimental
]

privacy_access_levels 	 [
Experimental
]

reachable_set 	 [
Experimental
]

region_maps 	 [
Experimental
]

specialization_graph_of 	 [
Experimental
]

stability 	 [
Experimental
]

super_predicates_of 	 [
Experimental
]

symbol_name 	 [
Experimental
]

trait_def 	 [
Experimental
]

trait_impls_of 	 [
Experimental
]

trait_of_item 	 [
Experimental
]

type_of 	 [
Experimental
]

type_param_predicates 	 [
Experimental
]

typeck_item_bodies 	 [
Experimental
]

typeck_tables_of 	 [
Experimental
]

variances_of 	 [
Experimental
]