Struct rustc::ty::maps::queries::is_copy_raw

pub struct is_copy_raw<'tcx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx, 'lcx> is_copy_raw<'tcx>

pub fn ensure(tcx: TyCtxt<'a, 'tcx, 'lcx>, key: ParamEnvAnd<'tcx, Ty<'tcx>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Ensure that either this query has all green inputs or been executed. Executing query::ensure(D) is considered a read of the dep-node D.

This function is particularly useful when executing passes for their side-effects -- e.g., in order to report errors for erroneous programs.

Note: The optimization is only available during incr. comp.

pub fn try_get(
tcx: TyCtxt<'a, 'tcx, 'lcx>,
span: Span,
key: ParamEnvAnd<'tcx, Ty<'tcx>>
) -> Result<bool, DiagnosticBuilder<'a>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?