Struct rustc::ty::maps::queries::mir

pub struct mir<'tcx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'a, 'tcx, 'lcx> mir<'tcx>

fn try_get(tcx: TyCtxt<'a, 'tcx, 'lcx>,
span: Span,
key: DefId)
-> Result<&'tcx RefCell<Mir<'tcx>>, CycleError<'a, 'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

fn get(tcx: TyCtxt<'a, 'tcx, 'lcx>,
span: Span,
key: DefId)
-> &'tcx RefCell<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from the def-id of a function/method or const/static to its MIR. Mutation is done at an item granularity to allow MIR optimization passes to function and still access cross-crate MIR (e.g. inlining or const eval).

Note that cross-crate MIR appears to be always borrowed (in the RefCell sense) to prevent accidental mutation.

fn force(tcx: TyCtxt<'a, 'tcx, 'lcx>, span: Span, key: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)