Struct rustc::ty::maps::Maps

pub struct Maps<'tcx> {
    pub ty: RefCell<DepTrackingMap<ty<'tcx>>>,
    pub generics: RefCell<DepTrackingMap<generics<'tcx>>>,
    pub predicates: RefCell<DepTrackingMap<predicates<'tcx>>>,
    pub super_predicates: RefCell<DepTrackingMap<super_predicates<'tcx>>>,
    pub type_param_predicates: RefCell<DepTrackingMap<type_param_predicates<'tcx>>>,
    pub trait_def: RefCell<DepTrackingMap<trait_def<'tcx>>>,
    pub adt_def: RefCell<DepTrackingMap<adt_def<'tcx>>>,
    pub adt_destructor: RefCell<DepTrackingMap<adt_destructor<'tcx>>>,
    pub adt_sized_constraint: RefCell<DepTrackingMap<adt_sized_constraint<'tcx>>>,
    pub variances: RefCell<DepTrackingMap<variances<'tcx>>>,
    pub associated_item_def_ids: RefCell<DepTrackingMap<associated_item_def_ids<'tcx>>>,
    pub associated_item: RefCell<DepTrackingMap<associated_item<'tcx>>>,
    pub impl_trait_ref: RefCell<DepTrackingMap<impl_trait_ref<'tcx>>>,
    pub inherent_impls: RefCell<DepTrackingMap<inherent_impls<'tcx>>>,
    pub mir: RefCell<DepTrackingMap<mir<'tcx>>>,
    pub mir_const_qualif: RefCell<DepTrackingMap<mir_const_qualif<'tcx>>>,
    pub closure_kind: RefCell<DepTrackingMap<closure_kind<'tcx>>>,
    pub closure_type: RefCell<DepTrackingMap<closure_type<'tcx>>>,
    pub coerce_unsized_info: RefCell<DepTrackingMap<coerce_unsized_info<'tcx>>>,
    pub typeck_item_bodies: RefCell<DepTrackingMap<typeck_item_bodies<'tcx>>>,
    pub typeck_tables: RefCell<DepTrackingMap<typeck_tables<'tcx>>>,
    pub coherent_trait: RefCell<DepTrackingMap<coherent_trait<'tcx>>>,
    pub crate_inherent_impls: RefCell<DepTrackingMap<crate_inherent_impls<'tcx>>>,
    pub crate_inherent_impls_overlap_check: RefCell<DepTrackingMap<crate_inherent_impls_overlap_check<'tcx>>>,
    pub monomorphic_const_eval: RefCell<DepTrackingMap<monomorphic_const_eval<'tcx>>>,
    pub privacy_access_levels: RefCell<DepTrackingMap<privacy_access_levels<'tcx>>>,
    pub mir_shims: RefCell<DepTrackingMap<mir_shims<'tcx>>>,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

Fields
ty: RefCell<DepTrackingMap<ty<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Records the type of every item.

generics: RefCell<DepTrackingMap<generics<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from the def-id of an item (trait/struct/enum/fn) to its associated generics and predicates.

predicates: RefCell<DepTrackingMap<predicates<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

super_predicates: RefCell<DepTrackingMap<super_predicates<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from the def-id of a trait to the list of super-predicates. This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that supertraits have additional acyclicity requirements).

type_param_predicates: RefCell<DepTrackingMap<type_param_predicates<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

To avoid cycles within the predicates of a single item we compute per-type-parameter predicates for resolving T::AssocTy.

trait_def: RefCell<DepTrackingMap<trait_def<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

adt_def: RefCell<DepTrackingMap<adt_def<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

adt_destructor: RefCell<DepTrackingMap<adt_destructor<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

adt_sized_constraint: RefCell<DepTrackingMap<adt_sized_constraint<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

variances: RefCell<DepTrackingMap<variances<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from def-id of a type or region parameter to its (inferred) variance.

associated_item_def_ids: RefCell<DepTrackingMap<associated_item_def_ids<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from an impl/trait def-id to a list of the def-ids of its items

associated_item: RefCell<DepTrackingMap<associated_item<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from a trait item to the trait item "descriptor"

impl_trait_ref: RefCell<DepTrackingMap<impl_trait_ref<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

inherent_impls: RefCell<DepTrackingMap<inherent_impls<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don't need to be exported.

mir: RefCell<DepTrackingMap<mir<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps from the def-id of a function/method or const/static to its MIR. Mutation is done at an item granularity to allow MIR optimization passes to function and still access cross-crate MIR (e.g. inlining or const eval).

Note that cross-crate MIR appears to be always borrowed (in the RefCell sense) to prevent accidental mutation.

mir_const_qualif: RefCell<DepTrackingMap<mir_const_qualif<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Maps DefId's that have an associated Mir to the result of the MIR qualify_consts pass. The actual meaning of the value isn't known except to the pass itself.

closure_kind: RefCell<DepTrackingMap<closure_kind<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

closure_type: RefCell<DepTrackingMap<closure_type<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

coerce_unsized_info: RefCell<DepTrackingMap<coerce_unsized_info<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Caches CoerceUnsized kinds for impls on custom types.

typeck_item_bodies: RefCell<DepTrackingMap<typeck_item_bodies<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

typeck_tables: RefCell<DepTrackingMap<typeck_tables<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

coherent_trait: RefCell<DepTrackingMap<coherent_trait<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

crate_inherent_impls: RefCell<DepTrackingMap<crate_inherent_impls<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Gets a complete map from all types to their inherent impls. Not meant to be used directly outside of coherence. (Defined only for LOCAL_CRATE)

crate_inherent_impls_overlap_check: RefCell<DepTrackingMap<crate_inherent_impls_overlap_check<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Checks all types in the krate for overlap in their inherent impls. Reports errors. Not meant to be used directly outside of coherence. (Defined only for LOCAL_CRATE)

monomorphic_const_eval: RefCell<DepTrackingMap<monomorphic_const_eval<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Results of evaluating monomorphic constants embedded in other items, such as enum variant explicit discriminants.

privacy_access_levels: RefCell<DepTrackingMap<privacy_access_levels<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Performs the privacy check and computes "access levels".

mir_shims: RefCell<DepTrackingMap<mir_shims<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'tcx> Maps<'tcx>

fn new(dep_graph: DepGraph,
providers: IndexVec<CrateNum, Providers<'tcx>>)
-> Self

🔬 This is a nightly-only experimental API. (rustc_private)