Struct rustc::ty::maps::Providers

pub struct Providers<'tcx> {
    pub type_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Ty<'tcx>,
    pub generics_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Generics,
    pub predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>,
    pub super_predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>,
    pub type_param_predicates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> GenericPredicates<'tcx>,
    pub trait_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TraitDef,
    pub adt_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx AdtDef,
    pub adt_destructor: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Destructor>,
    pub adt_sized_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx [Ty<'tcx>],
    pub adt_dtorck_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> DtorckConstraint<'tcx>,
    pub is_const_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_default_impl: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub crate_variances: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateVariancesMap>,
    pub variances_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Variance>>,
    pub associated_item_def_ids: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>,
    pub associated_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> AssociatedItem,
    pub impl_trait_ref: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<TraitRef<'tcx>>,
    pub impl_polarity: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ImplPolarity,
    pub inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>,
    pub mir_keys: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdSet>,
    pub mir_const_qualif: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> u8,
    pub mir_const: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>,
    pub mir_validated: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>,
    pub optimized_mir: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Mir<'tcx>,
    pub closure_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ClosureKind,
    pub fn_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> PolyFnSig<'tcx>,
    pub coerce_unsized_info: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> CoerceUnsizedInfo,
    pub typeck_item_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CompileResult,
    pub typeck_tables_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TypeckTables<'tcx>,
    pub has_typeck_tables: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub coherent_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId)),
    pub borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId),
    pub mir_borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId),
    pub crate_inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CrateInherentImpls,
    pub crate_inherent_impls_overlap_check: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum),
    pub const_eval: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, (DefId, &'tcx Substs<'tcx>)>) -> EvalResult<'tcx>,
    pub privacy_access_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<AccessLevels>,
    pub reachable_set: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<NodeSet>,
    pub region_maps: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<RegionMaps>,
    pub mir_shims: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: InstanceDef<'tcx>) -> &'tcx Mir<'tcx>,
    pub def_symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> SymbolName,
    pub symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: Instance<'tcx>) -> SymbolName,
    pub describe_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Def>,
    pub def_span: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Span,
    pub stability: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Stability>,
    pub deprecation: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Deprecation>,
    pub item_attrs: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<[Attribute]>,
    pub fn_arg_names: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Vec<Name>,
    pub impl_parent: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>,
    pub trait_of_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>,
    pub is_exported_symbol: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub item_body_nested_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<BTreeMap<BodyId, Body>>,
    pub const_is_rvalue_promotable_to_static: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_mir_available: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub trait_impls_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<TraitImpls>,
    pub specialization_graph_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Graph>,
    pub is_object_safe: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub param_env: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ParamEnv<'tcx>,
    pub is_copy_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub is_sized_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub is_freeze_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub needs_drop_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub layout_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> Result<&'tcx Layout, LayoutError<'tcx>>,
    pub dylib_dependency_formats: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<(CrateNum, LinkagePreference)>>,
    pub is_allocator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_panic_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub extern_crate: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Option<ExternCrate>>,
    pub lint_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LintLevelMap>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
type_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

generics_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Generics

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

super_predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type_param_predicates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TraitDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx AdtDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_destructor: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Destructor>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_sized_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx [Ty<'tcx>]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_dtorck_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> DtorckConstraint<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_const_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_default_impl: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_variances: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateVariancesMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

variances_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Variance>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

associated_item_def_ids: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

associated_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> AssociatedItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_trait_ref: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<TraitRef<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_polarity: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ImplPolarity

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_keys: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdSet>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_const_qualif: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> u8

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_const: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_validated: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

optimized_mir: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Mir<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

closure_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ClosureKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> PolyFnSig<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

coerce_unsized_info: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> CoerceUnsizedInfo

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

typeck_item_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CompileResult

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

typeck_tables_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TypeckTables<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

has_typeck_tables: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

coherent_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId))

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CrateInherentImpls

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_inherent_impls_overlap_check: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

const_eval: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, (DefId, &'tcx Substs<'tcx>)>) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

privacy_access_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<AccessLevels>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

reachable_set: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<NodeSet>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

region_maps: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<RegionMaps>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_shims: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: InstanceDef<'tcx>) -> &'tcx Mir<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> SymbolName

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: Instance<'tcx>) -> SymbolName

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

describe_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Def>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_span: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

stability: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Stability>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

deprecation: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Deprecation>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

item_attrs: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<[Attribute]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn_arg_names: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Vec<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_parent: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_of_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_exported_symbol: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

item_body_nested_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<BTreeMap<BodyId, Body>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

const_is_rvalue_promotable_to_static: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_mir_available: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_impls_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<TraitImpls>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

specialization_graph_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Graph>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_object_safe: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

param_env: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ParamEnv<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_copy_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_sized_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_freeze_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

needs_drop_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

layout_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> Result<&'tcx Layout, LayoutError<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dylib_dependency_formats: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<(CrateNum, LinkagePreference)>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_allocator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_panic_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

extern_crate: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Option<ExternCrate>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

lint_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LintLevelMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?