Module rustc::ty::relate

Unstable (rustc_private)

Generalized type relating mechanism. A type relation R relates a pair of values (A, B). A and B are usually types or regions but can be other things. Examples of type relations are subtyping, type equality, etc.

Enums

Cause 	 [Unstable]

Traits

Relate 	 [Unstable]

TypeRelation 	 [Unstable]

Functions

expected_found 	 [Unstable]

expected_found_bool 	 [Unstable]

relate_substs 	 [Unstable]

super_relate_tys 	 [Unstable]
The main "type relation" routine. Note that this does not handle inference artifacts, so you should filter those out before calling it.

Type Definitions

RelateResult 	 [Unstable]