Function rustc::ty::relate::super_relate_tys

pub fn super_relate_tys<'a, 'gcx, 'tcx, R>(
    relation: &mut R, 
    a: Ty<'tcx>, 
    b: Ty<'tcx>
) -> RelateResult<'tcx, Ty<'tcx>> where
    R: TypeRelation<'a, 'gcx, 'tcx>,
    'gcx: 'a + 'tcx,
    'tcx: 'a, 

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The main "type relation" routine. Note that this does not handle inference artifacts, so you should filter those out before calling it.