Trait rustc::ty::relate::Relate

pub trait Relate<'tcx>: TypeFoldable<'tcx> {
    fn relate<'a, 'gcx, R>(
        relation: &mut R, 
        a: &Self, 
        b: &Self
    ) -> RelateResult<'tcx, Self>
    where
        R: TypeRelation<'a, 'gcx, 'tcx>,
        'gcx: 'a + 'tcx,
        'tcx: 'a;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn relate<'a, 'gcx, R>(
relation: &mut R,
a: &Self,
b: &Self
) -> RelateResult<'tcx, Self> where
R: TypeRelation<'a, 'gcx, 'tcx>,
'gcx: 'a + 'tcx,
'tcx: 'a,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?