Struct rustc::ty::AdtDefData

pub struct AdtDefData<'tcx, 'container> {
    pub did: DefId,
    pub variants: Vec<VariantDefData<'tcx, 'container>>,
    // some fields omitted
}

Unstable (rustc_private)

The definition of an abstract data type - a struct or enum.

These are all interned (by intern_adt_def) into the adt_defs table.

Because of the possibility of nested tcx-s, this type needs 2 lifetimes: the traditional variant lifetime ('tcx) bounding the lifetime of the inner types is of course necessary. However, it is not sufficient - types from a child tcx must not be leaked into the master tcx by being stored in an AdtDefData.

The 'container lifetime ensures that by outliving the container tcx and preventing shorter-lived types from being inserted. When write access is not needed, the 'container lifetime can be erased to 'static, which can be done by the AdtDef wrapper.

Fields

did
Unstable (rustc_private)

variants
Unstable (rustc_private)

Methods

impl<'tcx, 'container> AdtDefData<'tcx, 'container>

fn adt_kind(&self) -> AdtKind

Unstable (rustc_private)

Returns the kind of the ADT - Struct or Enum.

fn is_dtorck(&'tcx self, tcx: &TyCtxt<'tcx>) -> bool

Unstable (rustc_private)

Returns whether this is a dtorck type. If this returns true, this type being safe for destruction requires it to be alive; Otherwise, only the contents are required to be.

fn is_fundamental(&self) -> bool

Unstable (rustc_private)

Returns whether this type is #[fundamental] for the purposes of coherence checking.

fn is_simd(&self) -> bool

Unstable (rustc_private)

fn is_phantom_data(&self) -> bool

Unstable (rustc_private)

Returns true if this is PhantomData.

fn has_dtor(&self) -> bool

Unstable (rustc_private)

Returns whether this type has a destructor.

fn struct_variant(&self) -> &VariantDefData<'tcx, 'container>

Unstable (rustc_private)

Asserts this is a struct and returns the struct's unique variant.

fn type_scheme(&self, tcx: &TyCtxt<'tcx>) -> TypeScheme<'tcx>

Unstable (rustc_private)

fn predicates(&self, tcx: &TyCtxt<'tcx>) -> GenericPredicates<'tcx>

Unstable (rustc_private)

fn all_fields(&self) -> FlatMap<Iter<VariantDefData<'tcx, 'container>>, Iter<FieldDefData<'tcx, 'container>>, fn<'s>(&'s VariantDefData<'tcx, 'container>) -> Iter<'s, FieldDefData<'tcx, 'container>>>

Unstable (rustc_private)

Returns an iterator over all fields contained by this ADT.

fn is_empty(&self) -> bool

Unstable (rustc_private)

fn is_univariant(&self) -> bool

Unstable (rustc_private)

fn is_payloadfree(&self) -> bool

Unstable (rustc_private)

fn variant_with_id(&self, vid: DefId) -> &VariantDefData<'tcx, 'container>

Unstable (rustc_private)

fn variant_index_with_id(&self, vid: DefId) -> usize

Unstable (rustc_private)

fn variant_of_def(&self, def: Def) -> &VariantDefData<'tcx, 'container>

Unstable (rustc_private)

fn destructor(&self) -> Option<DefId>

Unstable (rustc_private)

fn set_destructor(&self, dtor: DefId)

Unstable (rustc_private)

fn dtor_kind(&self) -> DtorKind

Unstable (rustc_private)