Struct rustc::ty::CrateAnalysis

pub struct CrateAnalysis {
    pub access_levels: Rc<AccessLevels>,
    pub name: String,
    pub glob_map: Option<GlobMap>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The complete set of all analyses described in this module. This is produced by the driver and fed to trans and later passes.

NB: These contents are being migrated into queries using the on-demand infrastructure.

Fields
access_levels: Rc<AccessLevels>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

name: String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

glob_map: Option<GlobMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?