Struct rustc::ty::Tables

pub struct Tables<'tcx> {
    pub type_relative_path_defs: NodeMap<Def>,
    pub node_types: NodeMap<Ty<'tcx>>,
    pub item_substs: NodeMap<ItemSubsts<'tcx>>,
    pub adjustments: NodeMap<Adjustment<'tcx>>,
    pub method_map: MethodMap<'tcx>,
    pub upvar_capture_map: UpvarCaptureMap<'tcx>,
    pub closure_tys: DefIdMap<ClosureTy<'tcx>>,
    pub closure_kinds: DefIdMap<ClosureKind>,
    pub liberated_fn_sigs: NodeMap<FnSig<'tcx>>,
    pub fru_field_types: NodeMap<Vec<Ty<'tcx>>>,
}

This is an experimental API. (rustc_private)

Fields
type_relative_path_defs: NodeMap<Def>

This is an experimental API. (rustc_private)

Resolved definitions for <T>::X associated paths.

node_types: NodeMap<Ty<'tcx>>

This is an experimental API. (rustc_private)

Stores the types for various nodes in the AST. Note that this table is not guaranteed to be populated until after typeck. See typeck::check::fn_ctxt for details.

item_substs: NodeMap<ItemSubsts<'tcx>>

This is an experimental API. (rustc_private)

Stores the type parameters which were substituted to obtain the type of this node. This only applies to nodes that refer to entities parameterized by type parameters, such as generic fns, types, or other items.

adjustments: NodeMap<Adjustment<'tcx>>

This is an experimental API. (rustc_private)

method_map: MethodMap<'tcx>

This is an experimental API. (rustc_private)

upvar_capture_map: UpvarCaptureMap<'tcx>

This is an experimental API. (rustc_private)

Borrows

closure_tys: DefIdMap<ClosureTy<'tcx>>

This is an experimental API. (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

closure_kinds: DefIdMap<ClosureKind>

This is an experimental API. (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

liberated_fn_sigs: NodeMap<FnSig<'tcx>>

This is an experimental API. (rustc_private)

For each fn, records the "liberated" types of its arguments and return type. Liberated means that all bound regions (including late-bound regions) are replaced with free equivalents. This table is not used in trans (since regions are erased there) and hence is not serialized to metadata.

fru_field_types: NodeMap<Vec<Ty<'tcx>>>

This is an experimental API. (rustc_private)

For each FRU expression, record the normalized types of the fields of the struct - this is needed because it is non-trivial to normalize while preserving regions. This table is used only in MIR construction and hence is not serialized to metadata.

Methods

impl<'a, 'gcx, 'tcx> Tables<'tcx>

fn empty() -> Tables<'tcx>

This is an experimental API. (rustc_private)

fn qpath_def(&self, qpath: &QPath, id: NodeId) -> Def

This is an experimental API. (rustc_private)

Returns the final resolution of a QPath in an Expr or Pat node.

fn node_id_to_type(&self, id: NodeId) -> Ty<'tcx>

This is an experimental API. (rustc_private)

fn node_id_to_type_opt(&self, id: NodeId) -> Option<Ty<'tcx>>

This is an experimental API. (rustc_private)

fn node_id_item_substs(&self, id: NodeId) -> Option<&'tcx Substs<'tcx>>

This is an experimental API. (rustc_private)

fn pat_ty(&self, pat: &Pat) -> Ty<'tcx>

This is an experimental API. (rustc_private)

fn pat_ty_opt(&self, pat: &Pat) -> Option<Ty<'tcx>>

This is an experimental API. (rustc_private)

fn expr_ty(&self, expr: &Expr) -> Ty<'tcx>

This is an experimental API. (rustc_private)

fn expr_ty_opt(&self, expr: &Expr) -> Option<Ty<'tcx>>

This is an experimental API. (rustc_private)

fn expr_ty_adjusted(&self, expr: &Expr) -> Ty<'tcx>

This is an experimental API. (rustc_private)

Returns the type of expr, considering any Adjustment entry recorded for that expression.

fn expr_ty_adjusted_opt(&self, expr: &Expr) -> Option<Ty<'tcx>>

This is an experimental API. (rustc_private)

fn is_method_call(&self, expr_id: NodeId) -> bool

This is an experimental API. (rustc_private)

fn is_overloaded_autoderef(&self, expr_id: NodeId, autoderefs: u32) -> bool

This is an experimental API. (rustc_private)

fn upvar_capture(&self, upvar_id: UpvarId) -> Option<UpvarCapture<'tcx>>

This is an experimental API. (rustc_private)