Struct rustc::ty::TyS

pub struct TyS<'tcx> {
    pub sty: TypeVariants<'tcx>,
    pub flags: Cell<TypeFlags>,
    // some fields omitted
}

Unstable (rustc_private)

Fields
sty: TypeVariants<'tcx>
Unstable (rustc_private)
flags: Cell<TypeFlags>
Unstable (rustc_private)

Methods

impl<'a, 'gcx, 'tcx> TyS<'tcx>

fn adjust<F>(&'tcx self, tcx: TyCtxt<'a, 'gcx, 'tcx>, span: Span, expr_id: NodeId, adjustment: Option<&AutoAdjustment<'tcx>>, method_type: F) -> Ty<'tcx> where F: FnMut(MethodCall) -> Option<Ty<'tcx>>

Unstable (rustc_private)

See expr_ty_adjusted

fn adjust_for_autoderef<F>(&'tcx self, tcx: TyCtxt<'a, 'gcx, 'tcx>, expr_id: NodeId, expr_span: Span, autoderef: u32, method_type: F) -> Ty<'tcx> where F: FnMut(MethodCall) -> Option<Ty<'tcx>>

Unstable (rustc_private)

fn adjust_for_autoref(&'tcx self, tcx: TyCtxt<'a, 'gcx, 'tcx>, autoref: Option<AutoRef<'tcx>>) -> Ty<'tcx>

Unstable (rustc_private)

impl<'a, 'gcx, 'lcx, 'tcx> TyS<'tcx>

fn sort_string(&self, tcx: TyCtxt<'a, 'gcx, 'lcx>) -> String

Unstable (rustc_private)

impl<'a, 'tcx> TyS<'tcx>

fn moves_by_default(&'tcx self, tcx: TyCtxt<'a, 'tcx, 'tcx>, param_env: &ParameterEnvironment<'tcx>, span: Span) -> bool

Unstable (rustc_private)

fn is_sized(&'tcx self, tcx: TyCtxt<'a, 'tcx, 'tcx>, param_env: &ParameterEnvironment<'tcx>, span: Span) -> bool

Unstable (rustc_private)

fn layout<'lcx>(&'tcx self, infcx: &InferCtxt<'a, 'tcx, 'lcx>) -> Result<&'tcx Layout, LayoutError<'tcx>>

Unstable (rustc_private)

fn is_representable(&'tcx self, tcx: TyCtxt<'a, 'tcx, 'tcx>, sp: Span) -> Representability

Unstable (rustc_private)

Check whether a type is representable. This means it cannot contain unboxed structural recursion. This check is needed for structs and enums.

impl<'a, 'tcx> TyS<'tcx>

fn type_contents(&'tcx self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> TypeContents

Unstable (rustc_private)

impl<'a, 'gcx, 'tcx> TyS<'tcx>

fn as_opt_param_ty(&self) -> Option<ParamTy>

Unstable (rustc_private)

fn is_nil(&self) -> bool

Unstable (rustc_private)

fn is_never(&self) -> bool

Unstable (rustc_private)

fn is_uninhabited(&self, _cx: TyCtxt) -> bool

Unstable (rustc_private)

fn is_primitive(&self) -> bool

Unstable (rustc_private)

fn is_ty_var(&self) -> bool

Unstable (rustc_private)

fn is_phantom_data(&self) -> bool

Unstable (rustc_private)

fn is_bool(&self) -> bool

Unstable (rustc_private)

fn is_param(&self, index: u32) -> bool

Unstable (rustc_private)

fn is_self(&self) -> bool

Unstable (rustc_private)

fn is_slice(&self) -> bool

Unstable (rustc_private)

fn is_structural(&self) -> bool

Unstable (rustc_private)

fn is_simd(&self) -> bool

Unstable (rustc_private)

fn sequence_element_type(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn simd_type(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn simd_size(&self, _cx: TyCtxt) -> usize

Unstable (rustc_private)

fn is_region_ptr(&self) -> bool

Unstable (rustc_private)

fn is_unsafe_ptr(&self) -> bool

Unstable (rustc_private)

fn is_unique(&self) -> bool

Unstable (rustc_private)

fn is_scalar(&self) -> bool

Unstable (rustc_private)

fn is_floating_point(&self) -> bool

Unstable (rustc_private)

Returns true if this type is a floating point type and false otherwise.

fn is_trait(&self) -> bool

Unstable (rustc_private)

fn is_integral(&self) -> bool

Unstable (rustc_private)

fn is_fresh(&self) -> bool

Unstable (rustc_private)

fn is_uint(&self) -> bool

Unstable (rustc_private)

fn is_char(&self) -> bool

Unstable (rustc_private)

fn is_fp(&self) -> bool

Unstable (rustc_private)

fn is_numeric(&self) -> bool

Unstable (rustc_private)

fn is_signed(&self) -> bool

Unstable (rustc_private)

fn is_machine(&self) -> bool

Unstable (rustc_private)

fn has_concrete_skeleton(&self) -> bool

Unstable (rustc_private)

fn builtin_deref(&self, explicit: bool, pref: LvaluePreference) -> Option<TypeAndMut<'tcx>>

Unstable (rustc_private)

fn builtin_index(&self) -> Option<Ty<'tcx>>

Unstable (rustc_private)

fn fn_sig(&self) -> &'tcx PolyFnSig<'tcx>

Unstable (rustc_private)

fn fn_abi(&self) -> Abi

Unstable (rustc_private)

Returns the ABI of the given function.

fn fn_args(&self) -> Binder<Vec<Ty<'tcx>>>

Unstable (rustc_private)

fn fn_ret(&self) -> Binder<Ty<'tcx>>

Unstable (rustc_private)

fn is_fn(&self) -> bool

Unstable (rustc_private)

fn ty_to_def_id(&self) -> Option<DefId>

Unstable (rustc_private)

fn ty_adt_def(&self) -> Option<AdtDef<'tcx>>

Unstable (rustc_private)

fn regions(&self) -> Vec<&'tcx Region>

Unstable (rustc_private)

Returns the regions directly referenced from this type (but not types reachable from this type via walk_tys). This ignores late-bound regions binders.

impl<'tcx> TyS<'tcx>

fn walk(&'tcx self) -> TypeWalker<'tcx>

Unstable (rustc_private)

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

isize => { isize }
Foo<Bar<isize>> => { Foo<Bar<isize>>, Bar<isize>, isize }
[isize] => { [isize], isize }

fn walk_shallow(&'tcx self) -> IntoIter<Ty<'tcx>>

Unstable (rustc_private)

Iterator that walks the immediate children of self. Hence Foo<Bar<i32>, u32> yields the sequence [Bar<i32>, u32] (but not i32, like walk).

fn maybe_walk<F>(&'tcx self, f: F) where F: FnMut(Ty<'tcx>) -> bool

Unstable (rustc_private)

Walks ty and any types appearing within ty, invoking the callback f on each type. If the callback returns false, then the children of the current type are ignored.

Note: prefer ty.walk() where possible.