Struct rustc::ty::TypeContents

pub struct TypeContents {
    pub bits: u64,
}

🔬 This is a nightly-only experimental API. (rustc_private)

Type contents is how the type checker reasons about kinds. They track what kinds of things are found within a type. You can think of them as kind of an "anti-kind". They track the kinds of values and thinks that are contained in types. Having a larger contents for a type tends to rule that type out from various kinds. For example, a type that contains a reference is not sendable.

The reason we compute type contents and not kinds is that it is easier for me (nmatsakis) to think about what is contained within a type than to think about what is not contained within a type.

Fields
bits: u64

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl TypeContents

fn when(&self, cond: bool) -> TypeContents

🔬 This is a nightly-only experimental API. (rustc_private)

fn intersects(&self, tc: TypeContents) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn owns_owned(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn interior_param(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn interior_unsafe(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn needs_drop(&self, _: TyCtxt) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn owned_pointer(&self) -> TypeContents

🔬 This is a nightly-only experimental API. (rustc_private)

Includes only those bits that still apply when indirected through a Box pointer

fn union<I, T, F>(v: I, f: F) -> TypeContents where I: IntoIterator<Item=T>, F: FnMut(T) -> TypeContents

🔬 This is a nightly-only experimental API. (rustc_private)

fn has_dtor(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)