Struct rustc::ty::TypeFlags

pub struct TypeFlags {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl TypeFlags

const HAS_PARAMS: TypeFlags = TypeFlags{bits: 1 << 0,}

const HAS_SELF: TypeFlags = TypeFlags{bits: 1 << 1,}

const HAS_TY_INFER: TypeFlags = TypeFlags{bits: 1 << 2,}

const HAS_RE_INFER: TypeFlags = TypeFlags{bits: 1 << 3,}

const HAS_RE_EARLY_BOUND: TypeFlags = TypeFlags{bits: 1 << 4,}

const HAS_FREE_REGIONS: TypeFlags = TypeFlags{bits: 1 << 5,}

const HAS_TY_ERR: TypeFlags = TypeFlags{bits: 1 << 6,}

const HAS_PROJECTION: TypeFlags = TypeFlags{bits: 1 << 7,}

const HAS_TY_CLOSURE: TypeFlags = TypeFlags{bits: 1 << 8,}

const HAS_LOCAL_NAMES: TypeFlags = TypeFlags{bits: 1 << 9,}

const KEEP_IN_LOCAL_TCX: TypeFlags = TypeFlags{bits: 1 << 10,}

const NEEDS_SUBST: TypeFlags = TypeFlags{bits: TypeFlags::HAS_PARAMS.bits | TypeFlags::HAS_SELF.bits | TypeFlags::HAS_RE_EARLY_BOUND.bits,}

const NOMINAL_FLAGS: TypeFlags = TypeFlags{bits: TypeFlags::HAS_PARAMS.bits | TypeFlags::HAS_SELF.bits | TypeFlags::HAS_TY_INFER.bits | TypeFlags::HAS_RE_INFER.bits | TypeFlags::HAS_RE_EARLY_BOUND.bits | TypeFlags::HAS_FREE_REGIONS.bits | TypeFlags::HAS_TY_ERR.bits | TypeFlags::HAS_PROJECTION.bits | TypeFlags::HAS_TY_CLOSURE.bits | TypeFlags::HAS_LOCAL_NAMES.bits | TypeFlags::KEEP_IN_LOCAL_TCX.bits,}

const SIZEDNESS_CACHED: TypeFlags = TypeFlags{bits: 1 << 16,}

const IS_SIZED: TypeFlags = TypeFlags{bits: 1 << 17,}

const MOVENESS_CACHED: TypeFlags = TypeFlags{bits: 1 << 18,}

const MOVES_BY_DEFAULT: TypeFlags = TypeFlags{bits: 1 << 19,}

fn empty() -> TypeFlags

Unstable (rustc_private)

Returns an empty set of flags.

fn all() -> TypeFlags

Unstable (rustc_private)

Returns the set containing all flags.

fn bits(&self) -> u32

Unstable (rustc_private)

Returns the raw value of the flags currently stored.

fn from_bits(bits: u32) -> Option<TypeFlags>

Unstable (rustc_private)

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u32) -> TypeFlags

Unstable (rustc_private)

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Unstable (rustc_private)

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Unstable (rustc_private)

Returns true if all flags are currently set.

fn intersects(&self, other: TypeFlags) -> bool

Unstable (rustc_private)

Returns true if there are flags common to both self and other.

fn contains(&self, other: TypeFlags) -> bool

Unstable (rustc_private)

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: TypeFlags)

Unstable (rustc_private)

Inserts the specified flags in-place.

fn remove(&mut self, other: TypeFlags)

Unstable (rustc_private)

Removes the specified flags in-place.

fn toggle(&mut self, other: TypeFlags)

Unstable (rustc_private)

Toggles the specified flags in-place.