Struct rustc::ty::TypeParameterDef

pub struct TypeParameterDef {
    pub name: Name,
    pub def_id: DefId,
    pub index: u32,
    pub has_default: bool,
    pub object_lifetime_default: ObjectLifetimeDefault,
    pub pure_wrt_drop: bool,
    pub synthetic: Option<SyntheticTyParamKind>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
name: Name

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_id: DefId

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

index: u32

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

has_default: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

object_lifetime_default: ObjectLifetimeDefault

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pure_wrt_drop: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pure_wrt_drop, set by the (unsafe) #[may_dangle] attribute on generic parameter T, asserts data behind the parameter T won't be accessed during the parent type's Drop impl.

synthetic: Option<SyntheticTyParamKind>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?