Module rustc::ty::subst

Unstable (rustc_private)

Reexports

pub use self::ParamSpace::*;

Structs

EnumeratedItems 	 [Unstable]

SeparateVecsPerParamSpace 	 [Unstable]
The split function converts one VecPerParamSpace into this SeparateVecsPerParamSpace structure.

Substs 	 [Unstable]
A substitution mapping type/region parameters to new values. We identify each in-scope parameter by an index and a parameter space (which indices where the parameter is defined; see ParamSpace).

VecPerParamSpace 	 [Unstable]
Vector of things sorted by param space. Used to keep the set of things declared on the type, self, or method distinct.

Enums

ParamSpace 	 [Unstable]

Traits

Subst 	 [Unstable]