Struct rustc::ty::subst::Kind

pub struct Kind<'tcx> { /* fields omitted */ }

Unstable (rustc_private)

An entity in the Rust typesystem, which can be one of several kinds (only types and lifetimes for now). To reduce memory usage, a Kind is a interned pointer, with the lowest 2 bits being reserved for a tag to indicate the type (Ty or Region) it points to.

Methods

impl<'tcx> Kind<'tcx>

fn as_type(self) -> Option<Ty<'tcx>>

Unstable (rustc_private)

fn as_region(self) -> Option<&'tcx Region>

Unstable (rustc_private)