Trait rustc::ty::subst::Subst

pub trait Subst<'tcx>: Sized {
    fn subst_spanned<'a, 'gcx>(
        &self, 
        tcx: TyCtxt<'a, 'gcx, 'tcx>, 
        substs: &[Kind<'tcx>], 
        span: Option<Span>
    ) -> Self;

    fn subst<'a, 'gcx>(
        &self, 
        tcx: TyCtxt<'a, 'gcx, 'tcx>, 
        substs: &[Kind<'tcx>]
    ) -> Self { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn subst_spanned<'a, 'gcx>(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
substs: &[Kind<'tcx>],
span: Option<Span>
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn subst<'a, 'gcx>(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
substs: &[Kind<'tcx>]
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?