Type Definition rustc::ty::PolyExistentialProjection

type PolyExistentialProjection<'tcx> = Binder<ExistentialProjection<'tcx>>;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx, 'gcx> PolyExistentialProjection<'tcx>

pub fn with_self_ty(
&self,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
self_ty: Ty<'tcx>
) -> PolyProjectionPredicate<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?