Trait rustc::ty::util::IntTypeExt

pub trait IntTypeExt {
    fn to_ty<'a, 'gcx, 'tcx>(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Ty<'tcx>;

    fn disr_incr<'a, 'tcx>(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>, 
        val: Option<ConstInt>
    ) -> Option<ConstInt>;

    fn assert_ty_matches(&self, val: ConstInt);

    fn initial_discriminant<'a, 'tcx>(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>
    ) -> ConstInt;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn to_ty<'a, 'gcx, 'tcx>(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn disr_incr<'a, 'tcx>(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
val: Option<ConstInt>
) -> Option<ConstInt>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn assert_ty_matches(&self, val: ConstInt)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn initial_discriminant<'a, 'tcx>(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> ConstInt

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?