Module rustc::util::common

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

ErrorReported 	 [
Experimental
]

Indenter 	 [
Experimental
]

ProfQDumpParams 	 [
Experimental
]
Parameters to the Dump variant of type ProfileQueriesMsg.

Enums

ProfileQueriesMsg 	 [
Experimental
]
A sequence of these messages induce a trace of query-based incremental compilation. FIXME(matthewhammer): Determine whether we should include cycle detection here or not.

Constants

FN_OUTPUT_NAME 	 [
Experimental
]

Traits

MemoizationMap 	 [
Experimental
]

Functions

duration_to_secs_str 	 [
Experimental
]

indent 	 [
Experimental
]

indenter 	 [
Experimental
]

path2cstr 	 [
Experimental
]

print_time_passes_entry 	 [
Experimental
]

profq_msg 	 [
Experimental
]
If enabled, send a message to the profile-queries thread

profq_set_chan 	 [
Experimental
]
Set channel for profile queries channel

record_time 	 [
Experimental
]

set_time_depth 	 [
Experimental
]
Set the current depth of time() calls. The idea is to call set_time_depth() with the result from time_depth() in the parent thread.

time 	 [
Experimental
]

time_depth 	 [
Experimental
]
Read the current depth of time() calls. This is used to encourage indentation across threads.

to_readable_str 	 [
Experimental
]