Function rustc::util::common::profq_set_chan

pub fn profq_set_chan(s: Sender<ProfileQueriesMsg>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Set channel for profile queries channel