Trait rustc::util::common::MemoizationMap

pub trait MemoizationMap {
    type Key: Clone;
    type Value: Clone;
    fn memoize<OP>(&self, key: Self::Key, op: OP) -> Self::Value
    where
        OP: FnOnce() -> Self::Value;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Associated Types

type Key: Clone

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type Value: Clone

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn memoize<OP>(&self, key: Self::Key, op: OP) -> Self::Value where
OP: FnOnce() -> Self::Value,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If key is present in the map, return the value, otherwise invoke op and store the value in the map.

NB: if the receiver is a DepTrackingMap, special care is needed in the op to ensure that the correct edges are added into the dep graph. See the DepTrackingMap impl for more details!