Trait rustc::util::ppaux::Print

pub trait Print {
    fn print<F: Write>(&self, f: &mut F, cx: &mut PrintContext) -> Result;

    fn print_to_string(&self, cx: &mut PrintContext) -> String { ... }

    fn print_display<F: Write>(
        &self, 
        f: &mut F, 
        cx: &mut PrintContext
    ) -> Result { ... }

    fn print_display_to_string(&self, cx: &mut PrintContext) -> String { ... }

    fn print_debug<F: Write>(&self, f: &mut F, cx: &mut PrintContext) -> Result { ... }

    fn print_debug_to_string(&self, cx: &mut PrintContext) -> String { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn print<F: Write>(&self, f: &mut F, cx: &mut PrintContext) -> Result

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn print_to_string(&self, cx: &mut PrintContext) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn print_display<F: Write>(&self, f: &mut F, cx: &mut PrintContext) -> Result

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn print_display_to_string(&self, cx: &mut PrintContext) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn print_debug<F: Write>(&self, f: &mut F, cx: &mut PrintContext) -> Result

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn print_debug_to_string(&self, cx: &mut PrintContext) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?