Struct rustc::util::snapshot_vec::SnapshotVecUnstable

pub struct SnapshotVec<D: SnapshotVecDelegate> {
    // some fields omitted
}

Methods

impl<D: SnapshotVecDelegate> SnapshotVec<D>

fn new(delegate: D) -> SnapshotVec<D>

fn record(&mut self, action: Undo)

fn push(&mut self, elem: Value) -> usize

fn get<'a>(&'a self, index: usize) -> &'a Value

fn get_mut<'a>(&'a mut self, index: usize) -> &'a mut Value

Returns a mutable pointer into the vec; whatever changes you make here cannot be undone automatically, so you should be sure call record() with some sort of suitable undo action.

fn set(&mut self, index: usize, new_elem: Value)

Updates the element at the given index. The old value will saved (and perhaps restored) if a snapshot is active.

fn start_snapshot(&mut self) -> Snapshot

fn actions_since_snapshot(&self, snapshot: &Snapshot) -> &[UndoLog<D>]

fn rollback_to(&mut self, snapshot: Snapshot)

fn commit(&mut self, snapshot: Snapshot)

Commits all changes since the last snapshot. Of course, they can still be undone if there is a snapshot further out.