Struct rustc::dep_graph::DepTrackingMap

pub struct DepTrackingMap<M: DepTrackingMapConfig> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A DepTrackingMap offers a subset of the Map API and ensures that we make calls to read and write as appropriate. We key the maps with a unique type for brevity.

Methods

impl<M: DepTrackingMapConfig> DepTrackingMap<M>

pub fn new(graph: DepGraph) -> DepTrackingMap<M>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?