Module rustc::hir::map

Unstable (rustc_private)

Reexports

pub use self::Node::*;

pub use self::definitions::{Definitions, DefKey, DefPath, DefPathData, DisambiguatedDefPathData, InlinedRootPath};

Modules

blocks 	 [Unstable]
This module provides a simplified abstraction for working with code blocks identified by their integer node-id. In particular, it captures a common set of attributes that all "function-like things" (represented by FnLike instances) share. For example, all FnLike instances have a type signature (be it explicit or inferred). And all FnLike instances have a body, i.e. the code that is run when the function-like thing it represents is invoked.

definitions 	 [Unstable]

Structs

Forest 	 [Unstable]
Stores a crate and any number of inlined items from other crates.

Map 	 [Unstable]
Represents a mapping from Node IDs to AST elements and their parent Node IDs

NodesMatchingSuffix 	 [Unstable]

Enums

MapEntry 	 [Unstable]
Represents an entry and its parent NodeID. The odd layout is to bring down the total size.

Node 	 [Unstable]

Traits

FoldOps 	 [Unstable]

NodePrinter 	 [Unstable]

Functions

collect_definitions 	 [Unstable]

map_crate 	 [Unstable]

map_decoded_item 	 [Unstable]
Used for items loaded from external crate that are being inlined into this crate.