Module rustc::hir::map::definitions

Unstable (rustc_private)

Structs

DefData 	 [Unstable]
For each definition, we track the following data. A definition here is defined somewhat circularly as "something with a def-id", but it generally corresponds to things like structs, enums, etc. There are also some rather random cases (like const initializer expressions) that are mostly just leftovers.

DefKey 	 [Unstable]
A unique identifier that we can use to lookup a definition precisely. It combines the index of the definition's parent (if any) with a DisambiguatedDefPathData.

DefPath 	 [Unstable]

Definitions 	 [Unstable]
The definition table containing node definitions

DisambiguatedDefPathData 	 [Unstable]
Pair of DefPathData and an integer disambiguator. The integer is normally 0, but in the event that there are multiple defs with the same parent and data, we use this field to disambiguate between them. This introduces some artificial ordering dependency but means that if you have (e.g.) two impls for the same type in the same module, they do get distinct def-ids.

InlinedRootPath 	 [Unstable]
Root of an inlined item. We track the DefPath of the item within the original crate but also its def-id. This is kind of an augmented version of a DefPath that includes a DefId. This is all sort of ugly but the hope is that inlined items will be going away soon anyway.

Enums

DefPathData 	 [Unstable]