Module rustc::hir::print

Unstable (rustc_private)

Reexports

pub use self::AnnNode::*;

Structs

NoAnn 	 [Unstable]

State 	 [Unstable]

Enums

AnnNode 	 [Unstable]

Constants

default_columns 	 [Unstable]

indent_unit 	 [Unstable]

Traits

PpAnn 	 [Unstable]

Functions

arg_to_string 	 [Unstable]

arm_to_string 	 [Unstable]

binop_to_string 	 [Unstable]

block_to_string 	 [Unstable]

bounds_to_string 	 [Unstable]

expr_to_string 	 [Unstable]

fn_block_to_string 	 [Unstable]

fun_to_string 	 [Unstable]

generics_to_string 	 [Unstable]

impl_item_to_string 	 [Unstable]

item_to_string 	 [Unstable]

lifetime_to_string 	 [Unstable]

name_to_string 	 [Unstable]

pat_to_string 	 [Unstable]

path_to_string 	 [Unstable]

print_crate 	 [Unstable]
Requires you to pass an input filename and reader so that it can scan the input text for comments and literals to copy forward.

rust_printer 	 [Unstable]

rust_printer_annotated 	 [Unstable]

stmt_to_string 	 [Unstable]

to_string 	 [Unstable]

trait_item_to_string 	 [Unstable]

ty_to_string 	 [Unstable]

variant_to_string 	 [Unstable]

visibility_qualified 	 [Unstable]

where_clause_to_string 	 [Unstable]