Struct rustc::hir::Crate

pub struct Crate {
    pub module: Mod,
    pub attrs: HirVec<Attribute>,
    pub config: CrateConfig,
    pub span: Span,
    pub exported_macros: HirVec<MacroDef>,
    pub items: BTreeMap<NodeId, Item>,
}

Unstable (rustc_private)

Fields

module
Unstable (rustc_private)

attrs
Unstable (rustc_private)

config
Unstable (rustc_private)

span
Unstable (rustc_private)

exported_macros
Unstable (rustc_private)

items
Unstable (rustc_private)

Methods

impl Crate

fn item(&self, id: NodeId) -> &Item

Unstable (rustc_private)

fn visit_all_items<'hir, V>(&'hir self, visitor: &mut V) where V: Visitor<'hir>

Unstable (rustc_private)

Visits all items in the crate in some determinstic (but unspecified) order. If you just need to process every item, but don't care about nesting, this method is the best choice.

If you do care about nesting -- usually because your algorithm follows lexical scoping rules -- then you want a different approach. You should override visit_nested_item in your visitor and then call intravisit::walk_crate instead.