Module rustc::infer::region_inference

Unstable (rustc_private)

See README.md

Reexports

pub use self::Constraint::*;

pub use self::Verify::*;

pub use self::UndoLogEntry::*;

pub use self::CombineMapType::*;

pub use self::RegionResolutionError::*;

pub use self::VarValue::*;

Structs

RegionSnapshot 	 [Unstable]

RegionVarBindings 	 [Unstable]

SameRegions 	 [Unstable]
SameRegions is used to group regions that we think are the same and would like to indicate so to the user. For example, the following function struct Foo { bar: i32 } fn foo2<'a, 'b>(x: &'a Foo) -> &'b i32 { &x.bar } would report an error because we expect 'a and 'b to match, and so we group 'a and 'b together inside a SameRegions struct

TwoRegions 	 [Unstable]

Enums

CombineMapType 	 [Unstable]

Constraint 	 [Unstable]

GenericKind 	 [Unstable]

ProcessedErrorOrigin 	 [Unstable]

RegionResolutionError 	 [Unstable]

UndoLogEntry 	 [Unstable]

VarValue 	 [Unstable]

Verify 	 [Unstable]

VerifyBound 	 [Unstable]

Type Definitions

CombineMap 	 [Unstable]