Struct rustc::lint::Lint

pub struct Lint {
    pub name: &'static str,
    pub default_level: Level,
    pub desc: &'static str,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Specification of a single lint.

Fields
name: &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A string identifier for the lint.

This identifies the lint in attributes and in command-line arguments. In those contexts it is always lowercase, but this field is compared in a way which is case-insensitive for ASCII characters. This allows declare_lint!() invocations to follow the convention of upper-case statics without repeating the name.

The name is written with underscores, e.g. "unused_imports". On the command line, underscores become dashes.

default_level: Level

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Default level for the lint.

desc: &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Description of the lint or the issue it detects.

e.g. "imports that are never used"

Methods

impl Lint

pub fn name_lower(&self) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Get the lint's name, with ASCII letters converted to lowercase.