Struct rustc::llvm::archive_ro::ArchiveROUnstable

pub struct ArchiveRO {
    // some fields omitted
}

Methods

impl ArchiveRO

fn open(dst: &Path) -> Option<ArchiveRO>

Opens a static archive for read-only purposes. This is more optimized than the open method because it uses LLVM's internal Archive class rather than shelling out to ar for everything.

If this archive is used with a mutable method, then an error will be raised.

fn read(&'a self, file: &str) -> Option<&'a [u8]>

Reads a file in the archive