Struct rustc::middle::cstore::DummyCrateStore

pub struct DummyCrateStore;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A dummy crate store that does not support any non-local crates, for test purposes.