Trait rustc::middle::cstore::CrateStore

pub trait CrateStore {
    fn crate_data_as_rc_any(&self, krate: CrateNum) -> Rc<Any>;

    fn metadata_loader(&self) -> &MetadataLoader;

    fn visibility(&self, def: DefId) -> Visibility;

    fn visible_parent_map<'a>(
        &'a self, 
        sess: &Session
    ) -> Ref<'a, DefIdMap<DefId>>;

    fn item_generics_cloned(&self, def: DefId) -> Generics;

    fn implementations_of_trait(&self, filter: Option<DefId>) -> Vec<DefId>;

    fn impl_defaultness(&self, def: DefId) -> Defaultness;

    fn associated_item_cloned(&self, def: DefId) -> AssociatedItem;

    fn is_dllimport_foreign_item(&self, def: DefId) -> bool;

    fn is_statically_included_foreign_item(&self, def_id: DefId) -> bool;

    fn dep_kind(&self, cnum: CrateNum) -> DepKind;

    fn export_macros(&self, cnum: CrateNum);

    fn lang_items(&self, cnum: CrateNum) -> Vec<(DefIndex, usize)>;

    fn missing_lang_items(&self, cnum: CrateNum) -> Vec<LangItem>;

    fn is_compiler_builtins(&self, cnum: CrateNum) -> bool;

    fn is_sanitizer_runtime(&self, cnum: CrateNum) -> bool;

    fn is_profiler_runtime(&self, cnum: CrateNum) -> bool;

    fn panic_strategy(&self, cnum: CrateNum) -> PanicStrategy;

    fn crate_name(&self, cnum: CrateNum) -> Symbol;

    fn original_crate_name(&self, cnum: CrateNum) -> Symbol;

    fn crate_hash(&self, cnum: CrateNum) -> Svh;

    fn crate_disambiguator(&self, cnum: CrateNum) -> Symbol;

    fn plugin_registrar_fn(&self, cnum: CrateNum) -> Option<DefId>;

    fn derive_registrar_fn(&self, cnum: CrateNum) -> Option<DefId>;

    fn native_libraries(&self, cnum: CrateNum) -> Vec<NativeLibrary>;

    fn exported_symbols(&self, cnum: CrateNum) -> Vec<DefId>;

    fn is_no_builtins(&self, cnum: CrateNum) -> bool;

    fn def_key(&self, def: DefId) -> DefKey;

    fn def_path(&self, def: DefId) -> DefPath;

    fn def_path_hash(&self, def: DefId) -> DefPathHash;

    fn def_path_table(&self, cnum: CrateNum) -> Rc<DefPathTable>;

    fn struct_field_names(&self, def: DefId) -> Vec<Name>;

    fn item_children(&self, did: DefId, sess: &Session) -> Vec<Export>;

    fn load_macro(&self, did: DefId, sess: &Session) -> LoadedMacro;

    fn item_body<'a, 'tcx>(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>, 
        def: DefId
    ) -> &'tcx Body;

    fn crates(&self) -> Vec<CrateNum>;

    fn used_libraries(&self) -> Vec<NativeLibrary>;

    fn used_link_args(&self) -> Vec<String>;

    fn used_crates(
        &self, 
        prefer: LinkagePreference
    ) -> Vec<(CrateNum, LibSource)>;

    fn used_crate_source(&self, cnum: CrateNum) -> CrateSource;

    fn extern_mod_stmt_cnum(&self, emod_id: NodeId) -> Option<CrateNum>;

    fn encode_metadata<'a, 'tcx>(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>, 
        link_meta: &LinkMeta, 
        reachable: &NodeSet
    ) -> (EncodedMetadata, EncodedMetadataHashes);

    fn metadata_encoding_version(&self) -> &[u8];
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A store of Rust crates, through with their metadata can be accessed.

Required Methods

fn crate_data_as_rc_any(&self, krate: CrateNum) -> Rc<Any>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn metadata_loader(&self) -> &MetadataLoader

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visibility(&self, def: DefId) -> Visibility

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn visible_parent_map<'a>(&'a self, sess: &Session) -> Ref<'a, DefIdMap<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn item_generics_cloned(&self, def: DefId) -> Generics

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn implementations_of_trait(&self, filter: Option<DefId>) -> Vec<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn impl_defaultness(&self, def: DefId) -> Defaultness

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn associated_item_cloned(&self, def: DefId) -> AssociatedItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_dllimport_foreign_item(&self, def: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_statically_included_foreign_item(&self, def_id: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn dep_kind(&self, cnum: CrateNum) -> DepKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn export_macros(&self, cnum: CrateNum)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn lang_items(&self, cnum: CrateNum) -> Vec<(DefIndex, usize)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn missing_lang_items(&self, cnum: CrateNum) -> Vec<LangItem>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_compiler_builtins(&self, cnum: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_sanitizer_runtime(&self, cnum: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_profiler_runtime(&self, cnum: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn panic_strategy(&self, cnum: CrateNum) -> PanicStrategy

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn crate_name(&self, cnum: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The name of the crate as it is referred to in source code of the current crate.

fn original_crate_name(&self, cnum: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The name of the crate as it is stored in the crate's metadata.

fn crate_hash(&self, cnum: CrateNum) -> Svh

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn crate_disambiguator(&self, cnum: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn plugin_registrar_fn(&self, cnum: CrateNum) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn derive_registrar_fn(&self, cnum: CrateNum) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn native_libraries(&self, cnum: CrateNum) -> Vec<NativeLibrary>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn exported_symbols(&self, cnum: CrateNum) -> Vec<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_no_builtins(&self, cnum: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_key(&self, def: DefId) -> DefKey

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_path(&self, def: DefId) -> DefPath

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_path_hash(&self, def: DefId) -> DefPathHash

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_path_table(&self, cnum: CrateNum) -> Rc<DefPathTable>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn struct_field_names(&self, def: DefId) -> Vec<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn item_children(&self, did: DefId, sess: &Session) -> Vec<Export>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn load_macro(&self, did: DefId, sess: &Session) -> LoadedMacro

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn item_body<'a, 'tcx>(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
def: DefId
) -> &'tcx Body

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn crates(&self) -> Vec<CrateNum>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn used_libraries(&self) -> Vec<NativeLibrary>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn used_link_args(&self) -> Vec<String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn used_crates(&self, prefer: LinkagePreference) -> Vec<(CrateNum, LibSource)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn used_crate_source(&self, cnum: CrateNum) -> CrateSource

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn extern_mod_stmt_cnum(&self, emod_id: NodeId) -> Option<CrateNum>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn encode_metadata<'a, 'tcx>(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
link_meta: &LinkMeta,
reachable: &NodeSet
) -> (EncodedMetadata, EncodedMetadataHashes)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn metadata_encoding_version(&self) -> &[u8]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?