Struct rustc::middle::free_region::FreeRegionMap

pub struct FreeRegionMap {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl FreeRegionMap

fn new() -> FreeRegionMap

fn relate_free_regions_from_implied_bounds<'tcx>(&mut self, implied_bounds: &[ImpliedBound<'tcx>])

fn relate_free_regions_from_predicates(&mut self, predicates: &[Predicate])

fn sub_free_region(&self, sub: FreeRegion, sup: FreeRegion) -> bool

Unstable (rustc_private)

Determines whether two free regions have a subregion relationship by walking the graph encoded in map. Note that it is possible that sub != sup and sub <= sup and sup <= sub (that is, the user can give two different names to the same lifetime).

fn lub_free_regions(&self, fr_a: FreeRegion, fr_b: FreeRegion) -> Region

fn is_subregion_of(&self, tcx: TyCtxt, sub_region: Region, super_region: Region) -> bool

Unstable (rustc_private)

Determines whether one region is a subregion of another. This is intended to run after inference and sadly the logic is somewhat duplicated with the code in infer.rs.

fn is_static(&self, super_region: FreeRegion) -> bool

Unstable (rustc_private)

Determines whether this free-region is required to be 'static