Module rustc::middle::stability

Unstable (rustc_private)

A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.

Reexports

pub use self::StabilityLevel::*;

Structs

Index 	 [Unstable]
A stability index, giving the stability level for items and methods.

Enums

StabilityLevel 	 [Unstable]

Functions

check_expr 	 [Unstable]
Helper for discovering nodes to check for stability

check_item 	 [Unstable]
Helper for discovering nodes to check for stability

check_pat 	 [Unstable]

check_path 	 [Unstable]

check_path_list_item 	 [Unstable]

check_unstable_api_usage 	 [Unstable]
Cross-references the feature names of unstable APIs with enabled features and possibly prints errors. Returns a list of all features used.

check_unused_or_stable_features 	 [Unstable]
Given the list of enabled features that were not language features (i.e. that were expected to be library features), and the list of features used from libraries, identify activated features that don't exist and error about them.