Enum rustc::middle::subst::RegionSubstsUnstable

pub enum RegionSubsts {
    ErasedRegions,
    NonerasedRegions(VecPerParamSpace<Region>),
}

Represents the values to use when substituting lifetime parameters. If the value is ErasedRegions, then this subst is occurring during trans, and all region parameters will be replaced with ty::ReStatic.

Variants

ErasedRegions

NonerasedRegions

Methods

impl RegionSubsts

fn is_erased(&self) -> bool