Struct rustc::middle::ty::MethodParamUnstable

pub struct MethodParam<'tcx> {
    pub trait_ref: Rc<TraitRef<'tcx>>,
    pub method_num: usize,
    pub impl_def_id: Option<DefId>,
}

Fields

trait_ref

method_num

impl_def_id

The impl for the trait from which the method comes. This should only be used for certain linting/heuristic purposes since there is no guarantee that this is Some in every situation that it could/should be.