Enum rustc::mir::repr::Terminator

pub enum Terminator<'tcx> {
    Goto {
        target: BasicBlock,
    },
    If {
        cond: Operand<'tcx>,
        targets: (BasicBlock, BasicBlock),
    },
    Switch {
        discr: Lvalue<'tcx>,
        adt_def: AdtDef<'tcx>,
        targets: Vec<BasicBlock>,
    },
    SwitchInt {
        discr: Lvalue<'tcx>,
        switch_ty: Ty<'tcx>,
        values: Vec<ConstVal>,
        targets: Vec<BasicBlock>,
    },
    Resume,
    Return,
    Drop {
        value: Lvalue<'tcx>,
        target: BasicBlock,
        unwind: Option<BasicBlock>,
    },
    Call {
        func: Operand<'tcx>,
        args: Vec<Operand<'tcx>>,
        destination: Option<(Lvalue<'tcx>, BasicBlock)>,
        cleanup: Option<BasicBlock>,
    },
}

Unstable (rustc_private)

Variants

Goto
Unstable (rustc_private)

block should have one successor in the graph; we jump there

Fields

target
Unstable (rustc_private)

If
Unstable (rustc_private)

jump to branch 0 if this lvalue evaluates to true

Fields

cond
Unstable (rustc_private)

targets
Unstable (rustc_private)

Switch
Unstable (rustc_private)

lvalue evaluates to some enum; jump depending on the branch

Fields

discr
Unstable (rustc_private)

adt_def
Unstable (rustc_private)

targets
Unstable (rustc_private)

SwitchInt
Unstable (rustc_private)

operand evaluates to an integer; jump depending on its value to one of the targets, and otherwise fallback to otherwise

Fields

discr
Unstable (rustc_private)

discriminant value being tested

switch_ty
Unstable (rustc_private)

type of value being tested

values
Unstable (rustc_private)

Possible values. The locations to branch to in each case are found in the corresponding indices from the targets vector.

targets
Unstable (rustc_private)

Possible branch sites. The length of this vector should be equal to the length of the values vector plus 1 -- the extra item is the block to branch to if none of the values fit.

Resume
Unstable (rustc_private)

Indicates that the landing pad is finished and unwinding should continue. Emitted by build::scope::diverge_cleanup.

Return
Unstable (rustc_private)

Indicates a normal return. The ReturnPointer lvalue should have been filled in by now. This should only occur in the END_BLOCK.

Drop
Unstable (rustc_private)

Drop the Lvalue

Fields

value
Unstable (rustc_private)

target
Unstable (rustc_private)

unwind
Unstable (rustc_private)

Call
Unstable (rustc_private)

Block ends with a call of a converging function

Fields

func
Unstable (rustc_private)

The function that’s being called

args
Unstable (rustc_private)

Arguments the function is called with

destination
Unstable (rustc_private)

Destination for the return value. If some, the call is converging.

cleanup
Unstable (rustc_private)

Cleanups to be done if the call unwinds.

Methods

impl<'tcx> Terminator<'tcx>

fn successors(&self) -> Cow<[BasicBlock]>

Unstable (rustc_private)

fn successors_mut(&mut self) -> Vec<&mut BasicBlock>

Unstable (rustc_private)

impl<'tcx> Terminator<'tcx>

fn fmt_head<W: Write>(&self, fmt: &mut W) -> Result

Unstable (rustc_private)

Write the "head" part of the terminator; that is, its name and the data it uses to pick the successor basic block, if any. The only information not inlcuded is the list of possible successors, which may be rendered differently between the text and the graphviz format.

fn fmt_successor_labels(&self) -> Vec<Cow<'static, str>>

Unstable (rustc_private)

Return the list of labels for the edges to the successor basic blocks.