Struct rustc::mir::repr::BasicBlockData

pub struct BasicBlockData<'tcx> {
    pub statements: Vec<Statement<'tcx>>,
    pub terminator: Option<Terminator<'tcx>>,
    pub is_cleanup: bool,
}

Unstable (rustc_private)

Fields

statements
Unstable (rustc_private)

terminator
Unstable (rustc_private)

is_cleanup
Unstable (rustc_private)

Methods

impl<'tcx> BasicBlockData<'tcx>

fn new(terminator: Option<Terminator<'tcx>>) -> BasicBlockData<'tcx>

Unstable (rustc_private)

fn terminator(&self) -> &Terminator<'tcx>

Unstable (rustc_private)

Accessor for terminator.

Terminator may not be None after construction of the basic block is complete. This accessor provides a convenience way to reach the terminator.

fn terminator_mut(&mut self) -> &mut Terminator<'tcx>

Unstable (rustc_private)