Struct rustc::mir::repr::Mir

pub struct Mir<'tcx> {
    pub visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>,
    pub promoted: IndexVec<Promoted, Mir<'tcx>>,
    pub return_ty: FnOutput<'tcx>,
    pub var_decls: IndexVec<Var, VarDecl<'tcx>>,
    pub arg_decls: IndexVec<Arg, ArgDecl<'tcx>>,
    pub temp_decls: IndexVec<Temp, TempDecl<'tcx>>,
    pub upvar_decls: Vec<UpvarDecl>,
    pub span: Span,
    // some fields omitted
}

Unstable (rustc_private)

Lowered representation of a single function.

Fields
visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>
Unstable (rustc_private)

List of visibility (lexical) scopes; these are referenced by statements and used (eventually) for debuginfo. Indexed by a VisibilityScope.

promoted: IndexVec<Promoted, Mir<'tcx>>
Unstable (rustc_private)

Rvalues promoted from this function, such as borrows of constants. Each of them is the Mir of a constant with the fn's type parameters in scope, but no vars or args and a separate set of temps.

return_ty: FnOutput<'tcx>
Unstable (rustc_private)

Return type of the function.

var_decls: IndexVec<Var, VarDecl<'tcx>>
Unstable (rustc_private)

Variables: these are stack slots corresponding to user variables. They may be assigned many times.

arg_decls: IndexVec<Arg, ArgDecl<'tcx>>
Unstable (rustc_private)

Args: these are stack slots corresponding to the input arguments.

temp_decls: IndexVec<Temp, TempDecl<'tcx>>
Unstable (rustc_private)

Temp declarations: stack slots that for temporaries created by the compiler. These are assigned once, but they are not SSA values in that it is possible to borrow them and mutate them through the resulting reference.

upvar_decls: Vec<UpvarDecl>
Unstable (rustc_private)

Names and capture modes of all the closure upvars, assuming the first argument is either the closure or a reference to it.

span: Span
Unstable (rustc_private)

A span representing this MIR, for error reporting

Methods

impl<'tcx> Mir<'tcx>

fn new(basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>, visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>, promoted: IndexVec<Promoted, Mir<'tcx>>, return_ty: FnOutput<'tcx>, var_decls: IndexVec<Var, VarDecl<'tcx>>, arg_decls: IndexVec<Arg, ArgDecl<'tcx>>, temp_decls: IndexVec<Temp, TempDecl<'tcx>>, upvar_decls: Vec<UpvarDecl>, span: Span) -> Self

Unstable (rustc_private)

fn basic_blocks(&self) -> &IndexVec<BasicBlock, BasicBlockData<'tcx>>

Unstable (rustc_private)

fn basic_blocks_mut(&mut self) -> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

Unstable (rustc_private)

fn predecessors(&self) -> Ref<IndexVec<BasicBlock, Vec<BasicBlock>>>

Unstable (rustc_private)

fn predecessors_for(&self, bb: BasicBlock) -> Ref<Vec<BasicBlock>>

Unstable (rustc_private)

fn dominators(&self) -> Dominators<BasicBlock>

Unstable (rustc_private)

fn local_index(&self, lvalue: &Lvalue<'tcx>) -> Option<Local>

Unstable (rustc_private)

Maps locals (Arg's, Var's, Temp's and ReturnPointer, in that order) to their index in the whole list of locals. This is useful if you want to treat all locals the same instead of repeating yourself.

fn count_locals(&self) -> usize

Unstable (rustc_private)

Counts the number of locals, such that that local_index will always return an index smaller than this count.

impl<'a, 'gcx, 'tcx> Mir<'tcx>

fn operand_ty(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>, operand: &Operand<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn binop_ty(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>, op: BinOp, lhs_ty: Ty<'tcx>, rhs_ty: Ty<'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn lvalue_ty(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>, lvalue: &Lvalue<'tcx>) -> LvalueTy<'tcx>

Unstable (rustc_private)

fn rvalue_ty(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>, rvalue: &Rvalue<'tcx>) -> Option<Ty<'tcx>>

Unstable (rustc_private)