Struct rustc::mir::LocalDecl

pub struct LocalDecl<'tcx> {
    pub mutability: Mutability,
    pub is_user_variable: bool,
    pub internal: bool,
    pub ty: Ty<'tcx>,
    pub name: Option<Name>,
    pub source_info: SourceInfo,
    pub lexical_scope: VisibilityScope,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A MIR local.

This can be a binding declared by the user, a temporary inserted by the compiler, a function argument, or the return pointer.

Fields
mutability: Mutability

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

let mut x vs let x.

Temporaries and the return pointer are always mutable.

is_user_variable: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

True if this corresponds to a user-declared local variable.

internal: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

True if this is an internal local

These locals are not based on types in the source code and are only used for a few desugarings at the moment.

The generator transformation will sanity check the locals which are live across a suspension point against the type components of the generator which type checking knows are live across a suspension point. We need to flag drop flags to avoid triggering this check as they are introduced after typeck.

Unsafety checking will also ignore dereferences of these locals, so they can be used for raw pointers only used in a desugaring.

This should be sound because the drop flags are fully algebraic, and therefore don't affect the OIBIT or outlives properties of the generator.

ty: Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Type of this local.

name: Option<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Name of the local, used in debuginfo and pretty-printing.

Note that function arguments can also have this set to Some(_) to generate better debuginfo.

source_info: SourceInfo

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Source info of the local.

lexical_scope: VisibilityScope

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The lexical visibility scope the local is defined in. If the local was defined in a let-statement, this is within the let-statement, rather than outside of it.

Methods

impl<'tcx> LocalDecl<'tcx>

pub fn new_temp(ty: Ty<'tcx>, span: Span) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new LocalDecl for a temporary.

pub fn new_internal(ty: Ty<'tcx>, span: Span) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new LocalDecl for a internal temporary.

pub fn new_return_pointer(return_ty: Ty, span: Span) -> LocalDecl

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Builds a LocalDecl for the return pointer.

This must be inserted into the local_decls list as the first local.