Struct rustc::mir::Mir

pub struct Mir<'tcx> {
    pub visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>,
    pub visibility_scope_info: ClearOnDecode<IndexVec<VisibilityScope, VisibilityScopeInfo>>,
    pub promoted: IndexVec<Promoted, Mir<'tcx>>,
    pub return_ty: Ty<'tcx>,
    pub yield_ty: Option<Ty<'tcx>>,
    pub generator_drop: Option<Box<Mir<'tcx>>>,
    pub generator_layout: Option<GeneratorLayout<'tcx>>,
    pub local_decls: IndexVec<Local, LocalDecl<'tcx>>,
    pub arg_count: usize,
    pub upvar_decls: Vec<UpvarDecl>,
    pub spread_arg: Option<Local>,
    pub span: Span,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Lowered representation of a single function.

Fields
visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

List of visibility (lexical) scopes; these are referenced by statements and used (eventually) for debuginfo. Indexed by a VisibilityScope.

visibility_scope_info: ClearOnDecode<IndexVec<VisibilityScope, VisibilityScopeInfo>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Crate-local information for each visibility scope, that can't (and needn't) be tracked across crates.

promoted: IndexVec<Promoted, Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Rvalues promoted from this function, such as borrows of constants. Each of them is the Mir of a constant with the fn's type parameters in scope, but a separate set of locals.

return_ty: Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Return type of the function.

yield_ty: Option<Ty<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Yield type of the function, if it is a generator.

generator_drop: Option<Box<Mir<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Generator drop glue

generator_layout: Option<GeneratorLayout<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The layout of a generator. Produced by the state transformation.

local_decls: IndexVec<Local, LocalDecl<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Declarations of locals.

The first local is the return value pointer, followed by arg_count locals for the function arguments, followed by any user-declared variables and temporaries.

arg_count: usize

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Number of arguments this function takes.

Starting at local 1, arg_count locals will be provided by the caller and can be assumed to be initialized.

If this MIR was built for a constant, this will be 0.

upvar_decls: Vec<UpvarDecl>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Names and capture modes of all the closure upvars, assuming the first argument is either the closure or a reference to it.

spread_arg: Option<Local>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Mark an argument local (which must be a tuple) as getting passed as its individual components at the LLVM level.

This is used for the "rust-call" ABI.

span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A span representing this MIR, for error reporting

Methods

impl<'tcx> Mir<'tcx>

pub fn new(
basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>,
visibility_scopes: IndexVec<VisibilityScope, VisibilityScopeData>,
visibility_scope_info: ClearOnDecode<IndexVec<VisibilityScope, VisibilityScopeInfo>>,
promoted: IndexVec<Promoted, Mir<'tcx>>,
return_ty: Ty<'tcx>,
yield_ty: Option<Ty<'tcx>>,
local_decls: IndexVec<Local, LocalDecl<'tcx>>,
arg_count: usize,
upvar_decls: Vec<UpvarDecl>,
span: Span
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn basic_blocks(&self) -> &IndexVec<BasicBlock, BasicBlockData<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn basic_blocks_mut(
&mut self
) -> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn predecessors(&self) -> Ref<IndexVec<BasicBlock, Vec<BasicBlock>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn predecessors_for(&self, bb: BasicBlock) -> Ref<Vec<BasicBlock>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn dominators(&self) -> Dominators<BasicBlock>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn local_kind(&self, local: Local) -> LocalKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn temps_iter<'a>(&'a self) -> impl Iterator<Item = Local> + 'a

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns an iterator over all temporaries.

pub fn vars_iter<'a>(&'a self) -> impl Iterator<Item = Local> + 'a

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns an iterator over all user-declared locals.

pub fn args_iter(&self) -> impl Iterator<Item = Local>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns an iterator over all function arguments.

pub fn vars_and_temps_iter(&self) -> impl Iterator<Item = Local>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns an iterator over all user-defined variables and compiler-generated temporaries (all locals that are neither arguments nor the return pointer).

pub fn make_statement_nop(&mut self, location: Location)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Changes a statement to a nop. This is both faster than deleting instructions and avoids invalidating statement indices in Locations.

pub fn source_info(&self, location: Location) -> &SourceInfo

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the source info associated with location.