Struct rustc::session::config::Options

pub struct Options {
    pub crate_types: Vec<CrateType>,
    pub gc: bool,
    pub optimize: OptLevel,
    pub debug_assertions: bool,
    pub debuginfo: DebugInfoLevel,
    pub lint_opts: Vec<(String, Level)>,
    pub lint_cap: Option<Level>,
    pub describe_lints: bool,
    pub output_types: HashMap<OutputType, Option<PathBuf>>,
    pub search_paths: SearchPaths,
    pub libs: Vec<(String, NativeLibraryKind)>,
    pub maybe_sysroot: Option<PathBuf>,
    pub target_triple: String,
    pub cfg: CrateConfig,
    pub test: bool,
    pub parse_only: bool,
    pub no_trans: bool,
    pub treat_err_as_bug: bool,
    pub no_analysis: bool,
    pub debugging_opts: DebuggingOptions,
    pub prints: Vec<PrintRequest>,
    pub cg: CodegenOptions,
    pub color: ColorConfig,
    pub show_span: Option<String>,
    pub externs: HashMap<String, Vec<String>>,
    pub crate_name: Option<String>,
    pub alt_std_name: Option<String>,
    pub unstable_features: UnstableFeatures,
}

Unstable (rustc_private)

Fields

crate_types
Unstable (rustc_private)

gc
Unstable (rustc_private)

optimize
Unstable (rustc_private)

debug_assertions
Unstable (rustc_private)

debuginfo
Unstable (rustc_private)

lint_opts
Unstable (rustc_private)

lint_cap
Unstable (rustc_private)

describe_lints
Unstable (rustc_private)

output_types
Unstable (rustc_private)

search_paths
Unstable (rustc_private)

libs
Unstable (rustc_private)

maybe_sysroot
Unstable (rustc_private)

target_triple
Unstable (rustc_private)

cfg
Unstable (rustc_private)

test
Unstable (rustc_private)

parse_only
Unstable (rustc_private)

no_trans
Unstable (rustc_private)

treat_err_as_bug
Unstable (rustc_private)

no_analysis
Unstable (rustc_private)

debugging_opts
Unstable (rustc_private)

prints
Unstable (rustc_private)

cg
Unstable (rustc_private)

color
Unstable (rustc_private)

show_span
Unstable (rustc_private)

externs
Unstable (rustc_private)

crate_name
Unstable (rustc_private)

alt_std_name
Unstable (rustc_private)

An optional name to use as the crate for std during std injection, written extern crate std = "name". Default to "std". Used by out-of-tree drivers.

unstable_features
Unstable (rustc_private)

Indicates how the compiler should treat unstable features