Struct rustc::session::Session

pub struct Session {
    pub target: Config,
    pub host: Target,
    pub opts: Options,
    pub cstore: Rc<for<'a> CrateStore<'a>>,
    pub parse_sess: ParseSess,
    pub entry_fn: RefCell<Option<(NodeId, Span)>>,
    pub entry_type: Cell<Option<EntryFnType>>,
    pub plugin_registrar_fn: Cell<Option<NodeId>>,
    pub default_sysroot: Option<PathBuf>,
    pub local_crate_source_file: Option<PathBuf>,
    pub working_dir: PathBuf,
    pub lint_store: RefCell<LintStore>,
    pub lints: RefCell<NodeMap<Vec<(LintId, Span, String)>>>,
    pub plugin_llvm_passes: RefCell<Vec<String>>,
    pub plugin_mir_passes: RefCell<Vec<Box<MirPass>>>,
    pub plugin_attributes: RefCell<Vec<(String, AttributeType)>>,
    pub crate_types: RefCell<Vec<CrateType>>,
    pub dependency_formats: RefCell<Dependencies>,
    pub crate_metadata: RefCell<Vec<String>>,
    pub features: RefCell<Features>,
    pub recursion_limit: Cell<usize>,
    pub injected_allocator: Cell<Option<CrateNum>>,
    pub available_macros: RefCell<HashSet<Name>>,
    pub imported_macro_spans: RefCell<HashMap<Span, (String, Span)>>,
    // some fields omitted
}

Unstable (rustc_private)

Fields

target
Unstable (rustc_private)

host
Unstable (rustc_private)

opts
Unstable (rustc_private)

cstore
Unstable (rustc_private)

parse_sess
Unstable (rustc_private)

entry_fn
Unstable (rustc_private)

entry_type
Unstable (rustc_private)

plugin_registrar_fn
Unstable (rustc_private)

default_sysroot
Unstable (rustc_private)

local_crate_source_file
Unstable (rustc_private)

working_dir
Unstable (rustc_private)

lint_store
Unstable (rustc_private)

lints
Unstable (rustc_private)

plugin_llvm_passes
Unstable (rustc_private)

plugin_mir_passes
Unstable (rustc_private)

plugin_attributes
Unstable (rustc_private)

crate_types
Unstable (rustc_private)

dependency_formats
Unstable (rustc_private)

crate_metadata
Unstable (rustc_private)

features
Unstable (rustc_private)

recursion_limit
Unstable (rustc_private)

The maximum recursion limit for potentially infinitely recursive operations such as auto-dereference and monomorphization.

injected_allocator
Unstable (rustc_private)

The metadata::creader module may inject an allocator dependency if it didn't already find one, and this tracks what was injected.

available_macros
Unstable (rustc_private)

Names of all bang-style macros and syntax extensions available in this crate

imported_macro_spans
Unstable (rustc_private)

Map from imported macro spans (which consist of the localized span for the macro body) to the macro name and defintion span in the source crate.

Methods

impl Session

fn struct_span_warn<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_span_warn_with_code<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str, code: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_warn<'a>(&'a self, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_span_err<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_span_err_with_code<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str, code: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_err<'a>(&'a self, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_span_fatal<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_span_fatal_with_code<'a, S: Into<MultiSpan>>(&'a self, sp: S, msg: &str, code: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn struct_fatal<'a>(&'a self, msg: &str) -> DiagnosticBuilder<'a>

Unstable (rustc_private)

fn span_fatal<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

Unstable (rustc_private)

fn span_fatal_with_code<S: Into<MultiSpan>>(&self, sp: S, msg: &str, code: &str) -> !

Unstable (rustc_private)

fn fatal(&self, msg: &str) -> !

Unstable (rustc_private)

fn span_err_or_warn<S: Into<MultiSpan>>(&self, is_warning: bool, sp: S, msg: &str)

Unstable (rustc_private)

fn span_err<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

Unstable (rustc_private)

fn span_err_with_code<S: Into<MultiSpan>>(&self, sp: S, msg: &str, code: &str)

Unstable (rustc_private)

fn err(&self, msg: &str)

Unstable (rustc_private)

fn err_count(&self) -> usize

Unstable (rustc_private)

fn has_errors(&self) -> bool

Unstable (rustc_private)

fn abort_if_errors(&self)

Unstable (rustc_private)

fn track_errors<F, T>(&self, f: F) -> Result<T, usize> where F: FnOnce() -> T

Unstable (rustc_private)

fn span_warn<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

Unstable (rustc_private)

fn span_warn_with_code<S: Into<MultiSpan>>(&self, sp: S, msg: &str, code: &str)

Unstable (rustc_private)

fn warn(&self, msg: &str)

Unstable (rustc_private)

fn opt_span_warn<S: Into<MultiSpan>>(&self, opt_sp: Option<S>, msg: &str)

Unstable (rustc_private)

fn opt_span_bug<S: Into<MultiSpan>>(&self, opt_sp: Option<S>, msg: &str) -> !

Unstable (rustc_private)

fn delay_span_bug<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

Unstable (rustc_private)

Delay a span_bug() call until abort_if_errors()

fn span_bug<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

Unstable (rustc_private)

fn bug(&self, msg: &str) -> !

Unstable (rustc_private)

fn note_without_error(&self, msg: &str)

Unstable (rustc_private)

fn span_note_without_error<S: Into<MultiSpan>>(&self, sp: S, msg: &str)

Unstable (rustc_private)

fn span_unimpl<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

Unstable (rustc_private)

fn unimpl(&self, msg: &str) -> !

Unstable (rustc_private)

fn add_lint(&self, lint: &'static Lint, id: NodeId, sp: Span, msg: String)

Unstable (rustc_private)

fn reserve_node_ids(&self, count: NodeId) -> NodeId

Unstable (rustc_private)

fn diagnostic<'a>(&'a self) -> &'a Handler

Unstable (rustc_private)

fn codemap<'a>(&'a self) -> &'a CodeMap

Unstable (rustc_private)

fn impossible_case<S: Into<MultiSpan>>(&self, sp: S, msg: &str) -> !

Unstable (rustc_private)

fn verbose(&self) -> bool

Unstable (rustc_private)

fn time_passes(&self) -> bool

Unstable (rustc_private)

fn count_llvm_insns(&self) -> bool

Unstable (rustc_private)

fn count_type_sizes(&self) -> bool

Unstable (rustc_private)

fn time_llvm_passes(&self) -> bool

Unstable (rustc_private)

fn trans_stats(&self) -> bool

Unstable (rustc_private)

fn meta_stats(&self) -> bool

Unstable (rustc_private)

fn asm_comments(&self) -> bool

Unstable (rustc_private)

fn no_verify(&self) -> bool

Unstable (rustc_private)

fn borrowck_stats(&self) -> bool

Unstable (rustc_private)

fn print_llvm_passes(&self) -> bool

Unstable (rustc_private)

fn lto(&self) -> bool

Unstable (rustc_private)

fn no_landing_pads(&self) -> bool

Unstable (rustc_private)

fn unstable_options(&self) -> bool

Unstable (rustc_private)

fn print_enum_sizes(&self) -> bool

Unstable (rustc_private)

fn nonzeroing_move_hints(&self) -> bool

Unstable (rustc_private)

fn sysroot<'a>(&'a self) -> &'a Path

Unstable (rustc_private)

fn target_filesearch(&self, kind: PathKind) -> FileSearch

Unstable (rustc_private)

fn host_filesearch(&self, kind: PathKind) -> FileSearch

Unstable (rustc_private)