Enum rustc::traits::ObligationCauseCode

pub enum ObligationCauseCode<'tcx> {
    MiscObligation,
    SliceOrArrayElem,
    TupleElem,
    ProjectionWf(ProjectionTy<'tcx>),
    ItemObligation(DefId),
    ReferenceOutlivesReferent(Ty<'tcx>),
    ObjectTypeBound(Ty<'tcx>, &'tcx Region),
    ObjectCastObligation(Ty<'tcx>),
    AssignmentLhsSized,
    StructInitializerSized,
    VariableType(NodeId),
    ReturnType,
    RepeatVec,
    FieldSized,
    ConstSized,
    SharedStatic,
    BuiltinDerivedObligation(DerivedObligationCause<'tcx>),
    ImplDerivedObligation(DerivedObligationCause<'tcx>),
    CompareImplMethodObligation {
        item_name: Name,
        impl_item_def_id: DefId,
        trait_item_def_id: DefId,
        lint_id: Option<NodeId>,
    },
    ExprAssignable,
    MatchExpressionArm {
        arm_span: Span,
        source: MatchSource,
    },
    IfExpression,
    IfExpressionWithNoElse,
    EquatePredicate,
    MainFunctionType,
    StartFunctionType,
    IntrinsicType,
    MethodReceiver,
}

🔬 This is a nightly-only experimental API. (rustc_private)

Variants
MiscObligation

🔬 This is a nightly-only experimental API. (rustc_private)

Not well classified or should be obvious from span.

SliceOrArrayElem

🔬 This is a nightly-only experimental API. (rustc_private)

A slice or array is WF only if T: Sized

TupleElem

🔬 This is a nightly-only experimental API. (rustc_private)

A tuple is WF only if its middle elements are Sized

ProjectionWf(ProjectionTy<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

This is the trait reference from the given projection

ItemObligation(DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

In an impl of trait X for type Y, type Y must also implement all supertraits of X.

ReferenceOutlivesReferent(Ty<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

A type like &'a T is WF only if T: 'a.

ObjectTypeBound(Ty<'tcx>, &'tcx Region)

🔬 This is a nightly-only experimental API. (rustc_private)

A type like Box<Foo<'a> + 'b> is WF only if 'b: 'a.

ObjectCastObligation(Ty<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

Obligation incurred due to an object cast.

AssignmentLhsSized

🔬 This is a nightly-only experimental API. (rustc_private)

Various cases where expressions must be sized/copy/etc:

StructInitializerSized

🔬 This is a nightly-only experimental API. (rustc_private)

VariableType(NodeId)

🔬 This is a nightly-only experimental API. (rustc_private)

ReturnType

🔬 This is a nightly-only experimental API. (rustc_private)

RepeatVec

🔬 This is a nightly-only experimental API. (rustc_private)

FieldSized

🔬 This is a nightly-only experimental API. (rustc_private)

ConstSized

🔬 This is a nightly-only experimental API. (rustc_private)

SharedStatic

🔬 This is a nightly-only experimental API. (rustc_private)

BuiltinDerivedObligation(DerivedObligationCause<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

ImplDerivedObligation(DerivedObligationCause<'tcx>)

🔬 This is a nightly-only experimental API. (rustc_private)

CompareImplMethodObligation

🔬 This is a nightly-only experimental API. (rustc_private)

Fields of CompareImplMethodObligation

item_name: Name

🔬 This is a nightly-only experimental API. (rustc_private)

impl_item_def_id: DefId

🔬 This is a nightly-only experimental API. (rustc_private)

trait_item_def_id: DefId

🔬 This is a nightly-only experimental API. (rustc_private)

lint_id: Option<NodeId>

🔬 This is a nightly-only experimental API. (rustc_private)

ExprAssignable

🔬 This is a nightly-only experimental API. (rustc_private)

MatchExpressionArm

🔬 This is a nightly-only experimental API. (rustc_private)

Fields of MatchExpressionArm

arm_span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

source: MatchSource

🔬 This is a nightly-only experimental API. (rustc_private)

IfExpression

🔬 This is a nightly-only experimental API. (rustc_private)

IfExpressionWithNoElse

🔬 This is a nightly-only experimental API. (rustc_private)

EquatePredicate

🔬 This is a nightly-only experimental API. (rustc_private)

MainFunctionType

🔬 This is a nightly-only experimental API. (rustc_private)

StartFunctionType

🔬 This is a nightly-only experimental API. (rustc_private)

IntrinsicType

🔬 This is a nightly-only experimental API. (rustc_private)

MethodReceiver

🔬 This is a nightly-only experimental API. (rustc_private)