Struct rustc::ty::layout::Struct

pub struct Struct {
    pub align: Align,
    pub primitive_align: Align,
    pub packed: bool,
    pub sized: bool,
    pub offsets: Vec<Size>,
    pub memory_index: Vec<u32>,
    pub min_size: Size,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A structure, a product type in ADT terms.

Fields
align: Align

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Maximum alignment of fields and repr alignment.

primitive_align: Align

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Primitive alignment of fields without repr alignment.

packed: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If true, no alignment padding is used.

sized: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If true, the size is exact, otherwise it's only a lower bound.

offsets: Vec<Size>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Offsets for the first byte of each field, ordered to match the source definition order. This vector does not go in increasing order. FIXME(eddyb) use small vector optimization for the common case.

memory_index: Vec<u32>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Maps source order field indices to memory order indices, depending how fields were permuted. FIXME (camlorn) also consider small vector optimization here.

min_size: Size

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx> Struct

pub fn stride(&self) -> Size

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Get the size with trailing alignment padding.

pub fn field_index_by_increasing_offset<'b>(
&'b self
) -> impl Iterator<Item = usize> + 'b

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Get indices of the tys that made this struct by increasing offset.

pub fn over_align(&self) -> Option<u32>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?