Struct rustc::ty::layout::TargetDataLayout

pub struct TargetDataLayout {
    pub endian: Endian,
    pub i1_align: Align,
    pub i8_align: Align,
    pub i16_align: Align,
    pub i32_align: Align,
    pub i64_align: Align,
    pub f32_align: Align,
    pub f64_align: Align,
    pub pointer_size: Size,
    pub pointer_align: Align,
    pub aggregate_align: Align,
    pub vector_align: Vec<(Size, Align)>,
}

Unstable (rustc_private)

Parsed Data layout for a target, which contains everything needed to compute layouts.

Fields
endian: Endian

Unstable (rustc_private)

i1_align: Align

Unstable (rustc_private)

i8_align: Align

Unstable (rustc_private)

i16_align: Align

Unstable (rustc_private)

i32_align: Align

Unstable (rustc_private)

i64_align: Align

Unstable (rustc_private)

f32_align: Align

Unstable (rustc_private)

f64_align: Align

Unstable (rustc_private)

pointer_size: Size

Unstable (rustc_private)

pointer_align: Align

Unstable (rustc_private)

aggregate_align: Align

Unstable (rustc_private)

vector_align: Vec<(Size, Align)>

Unstable (rustc_private)

Alignments for vector types.

Methods

impl TargetDataLayout

fn parse(sess: &Session) -> TargetDataLayout

Unstable (rustc_private)

fn obj_size_bound(&self) -> u64

Unstable (rustc_private)

Return exclusive upper bound on object size.

The theoretical maximum object size is defined as the maximum positive isize value. This ensures that the offset semantics remain well-defined by allowing it to correctly index every address within an object along with one byte past the end, along with allowing isize to store the difference between any two pointers into an object.

The upper bound on 64-bit currently needs to be lower because LLVM uses a 64-bit integer to represent object size in bits. It would need to be 1 << 61 to account for this, but is currently conservatively bounded to 1 << 47 as that is enough to cover the current usable address space on 64-bit ARMv8 and x86_64.

fn ptr_sized_integer(&self) -> Integer

Unstable (rustc_private)