Module rustc::ty::maps::queries

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

adt_def 	 [
Experimental
]

adt_destructor 	 [
Experimental
]

adt_dtorck_constraint 	 [
Experimental
]

adt_sized_constraint 	 [
Experimental
]

all_crate_nums 	 [
Experimental
]

all_trait_implementations 	 [
Experimental
]

associated_item 	 [
Experimental
]

associated_item_def_ids 	 [
Experimental
]

borrowck 	 [
Experimental
]

closure_kind 	 [
Experimental
]

codegen_unit 	 [
Experimental
]

coerce_unsized_info 	 [
Experimental
]

coherent_trait 	 [
Experimental
]

collect_and_partition_translation_items 	 [
Experimental
]

compile_codegen_unit 	 [
Experimental
]

const_eval 	 [
Experimental
]

const_is_rvalue_promotable_to_static 	 [
Experimental
]

contains_extern_indicator 	 [
Experimental
]

crate_disambiguator 	 [
Experimental
]

crate_hash 	 [
Experimental
]

crate_inherent_impls 	 [
Experimental
]

crate_inherent_impls_overlap_check 	 [
Experimental
]

crate_name 	 [
Experimental
]

crate_variances 	 [
Experimental
]

def_span 	 [
Experimental
]

def_symbol_name 	 [
Experimental
]

defined_lang_items 	 [
Experimental
]

dep_kind 	 [
Experimental
]

derive_registrar_fn 	 [
Experimental
]

describe_def 	 [
Experimental
]

dylib_dependency_formats 	 [
Experimental
]

erase_regions_ty 	 [
Experimental
]

export_name 	 [
Experimental
]

exported_symbol_ids 	 [
Experimental
]

exported_symbols 	 [
Experimental
]

extern_const_body 	 [
Experimental
]

extern_crate 	 [
Experimental
]

extern_mod_stmt_cnum 	 [
Experimental
]

fn_arg_names 	 [
Experimental
]

fn_sig 	 [
Experimental
]

freevars 	 [
Experimental
]

generator_sig 	 [
Experimental
]

generics_of 	 [
Experimental
]

get_lang_items 	 [
Experimental
]

has_clone_closures 	 [
Experimental
]

has_copy_closures 	 [
Experimental
]

has_global_allocator 	 [
Experimental
]

has_typeck_tables 	 [
Experimental
]

impl_defaultness 	 [
Experimental
]

impl_parent 	 [
Experimental
]

impl_polarity 	 [
Experimental
]

impl_trait_ref 	 [
Experimental
]

implementations_of_trait 	 [
Experimental
]

in_scope_traits_map 	 [
Experimental
]

inferred_outlives_of 	 [
Experimental
]

inherent_impls 	 [
Experimental
]

is_compiler_builtins 	 [
Experimental
]

is_const_fn 	 [
Experimental
]

is_copy_raw 	 [
Experimental
]

is_default_impl 	 [
Experimental
]

is_dllimport_foreign_item 	 [
Experimental
]

is_exported_symbol 	 [
Experimental
]

is_foreign_item 	 [
Experimental
]

is_freeze_raw 	 [
Experimental
]

is_late_bound_map 	 [
Experimental
]

is_mir_available 	 [
Experimental
]

is_no_builtins 	 [
Experimental
]

is_object_safe 	 [
Experimental
]

is_panic_runtime 	 [
Experimental
]

is_profiler_runtime 	 [
Experimental
]

is_sanitizer_runtime 	 [
Experimental
]

is_sized_raw 	 [
Experimental
]

is_statically_included_foreign_item 	 [
Experimental
]

is_translated_function 	 [
Experimental
]

item_attrs 	 [
Experimental
]

item_body_nested_bodies 	 [
Experimental
]

item_children 	 [
Experimental
]

layout_raw 	 [
Experimental
]

link_args 	 [
Experimental
]

lint_levels 	 [
Experimental
]

lookup_deprecation_entry 	 [
Experimental
]

lookup_stability 	 [
Experimental
]

maybe_unused_extern_crates 	 [
Experimental
]

maybe_unused_trait_import 	 [
Experimental
]

mir_borrowck 	 [
Experimental
]

mir_const 	 [
Experimental
]

mir_const_qualif 	 [
Experimental
]

mir_keys 	 [
Experimental
]

mir_shims 	 [
Experimental
]

mir_validated 	 [
Experimental
]

missing_extern_crate_item 	 [
Experimental
]

missing_lang_items 	 [
Experimental
]

module_exports 	 [
Experimental
]

named_region_map 	 [
Experimental
]

native_libraries 	 [
Experimental
]

native_library_kind 	 [
Experimental
]

needs_drop_raw 	 [
Experimental
]

object_lifetime_defaults_map 	 [
Experimental
]

optimized_mir 	 [
Experimental
]

original_crate_name 	 [
Experimental
]

output_filenames 	 [
Experimental
]

panic_strategy 	 [
Experimental
]

param_env 	 [
Experimental
]

plugin_registrar_fn 	 [
Experimental
]

postorder_cnums 	 [
Experimental
]

predicates_of 	 [
Experimental
]

privacy_access_levels 	 [
Experimental
]

reachable_set 	 [
Experimental
]

region_scope_tree 	 [
Experimental
]

rvalue_promotable_map 	 [
Experimental
]

specialization_graph_of 	 [
Experimental
]

specializes 	 [
Experimental
]

stability_index 	 [
Experimental
]

super_predicates_of 	 [
Experimental
]

symbol_name 	 [
Experimental
]

trait_def 	 [
Experimental
]

trait_impls_of 	 [
Experimental
]

trait_of_item 	 [
Experimental
]

trans_fulfill_obligation 	 [
Experimental
]

type_of 	 [
Experimental
]

type_param_predicates 	 [
Experimental
]

typeck_item_bodies 	 [
Experimental
]

typeck_tables_of 	 [
Experimental
]

unsafety_violations 	 [
Experimental
]

used_crate_source 	 [
Experimental
]

variances_of 	 [
Experimental
]

visibility 	 [
Experimental
]

visible_parent_map 	 [
Experimental
]

vtable_methods 	 [
Experimental
]