Struct rustc::ty::maps::Providers

pub struct Providers<'tcx> {
    pub type_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Ty<'tcx>,
    pub generics_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Generics,
    pub predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>,
    pub super_predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>,
    pub type_param_predicates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> GenericPredicates<'tcx>,
    pub trait_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TraitDef,
    pub adt_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx AdtDef,
    pub adt_destructor: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Destructor>,
    pub adt_sized_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx [Ty<'tcx>],
    pub adt_dtorck_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> DtorckConstraint<'tcx>,
    pub is_const_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_default_impl: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub crate_variances: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateVariancesMap>,
    pub variances_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Variance>>,
    pub associated_item_def_ids: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>,
    pub associated_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> AssociatedItem,
    pub impl_trait_ref: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<TraitRef<'tcx>>,
    pub impl_polarity: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ImplPolarity,
    pub inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>,
    pub mir_keys: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdSet>,
    pub mir_const_qualif: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> (u8, Rc<IdxSetBuf<Local>>),
    pub mir_const: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>,
    pub mir_validated: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>,
    pub optimized_mir: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Mir<'tcx>,
    pub closure_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ClosureKind,
    pub fn_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> PolyFnSig<'tcx>,
    pub generator_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<PolyGenSig<'tcx>>,
    pub coerce_unsized_info: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> CoerceUnsizedInfo,
    pub typeck_item_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CompileResult,
    pub typeck_tables_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TypeckTables<'tcx>,
    pub has_typeck_tables: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub coherent_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId)),
    pub borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId),
    pub mir_borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId),
    pub crate_inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CrateInherentImpls,
    pub crate_inherent_impls_overlap_check: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum),
    pub const_eval: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, (DefId, &'tcx Substs<'tcx>)>) -> EvalResult<'tcx>,
    pub privacy_access_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<AccessLevels>,
    pub reachable_set: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<NodeSet>,
    pub region_scope_tree: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<ScopeTree>,
    pub mir_shims: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: InstanceDef<'tcx>) -> &'tcx Mir<'tcx>,
    pub def_symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> SymbolName,
    pub symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: Instance<'tcx>) -> SymbolName,
    pub describe_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Def>,
    pub def_span: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Span,
    pub lookup_stability: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<&'tcx Stability>,
    pub lookup_deprecation_entry: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DeprecationEntry>,
    pub item_attrs: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<[Attribute]>,
    pub fn_arg_names: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Vec<Name>,
    pub impl_parent: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>,
    pub trait_of_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>,
    pub is_exported_symbol: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub item_body_nested_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<BTreeMap<BodyId, Body>>,
    pub const_is_rvalue_promotable_to_static: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_mir_available: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub trait_impls_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<TraitImpls>,
    pub specialization_graph_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Graph>,
    pub is_object_safe: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub param_env: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ParamEnv<'tcx>,
    pub is_copy_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub is_sized_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub is_freeze_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub needs_drop_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool,
    pub layout_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> Result<&'tcx Layout, LayoutError<'tcx>>,
    pub dylib_dependency_formats: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(CrateNum, LinkagePreference)>>,
    pub is_panic_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub is_compiler_builtins: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub has_global_allocator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub is_sanitizer_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub is_profiler_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub panic_strategy: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> PanicStrategy,
    pub is_no_builtins: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub extern_crate: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Option<ExternCrate>>,
    pub specializes: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> bool,
    pub in_scope_traits_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Rc<Vec<TraitCandidate>>>>>,
    pub module_exports: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Rc<Vec<Export>>>,
    pub lint_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LintLevelMap>,
    pub impl_defaultness: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Defaultness,
    pub exported_symbols: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<DefId>>,
    pub native_libraries: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<NativeLibrary>>,
    pub plugin_registrar_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Option<DefId>,
    pub derive_registrar_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Option<DefId>,
    pub crate_disambiguator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol,
    pub crate_hash: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Svh,
    pub original_crate_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol,
    pub implementations_of_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId)) -> Rc<Vec<DefId>>,
    pub all_trait_implementations: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<DefId>>,
    pub is_dllimport_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub is_statically_included_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub native_library_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<NativeLibraryKind>,
    pub link_args: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<String>>,
    pub named_region_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Region>>>,
    pub is_late_bound_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashSet<ItemLocalId>>>,
    pub object_lifetime_defaults_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Rc<Vec<ObjectLifetimeDefault>>>>>,
    pub visibility: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Visibility,
    pub dep_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> DepKind,
    pub crate_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol,
    pub item_children: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Export>>,
    pub extern_mod_stmt_cnum: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<CrateNum>,
    pub get_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LanguageItems>,
    pub defined_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(DefIndex, usize)>>,
    pub missing_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<LangItem>>,
    pub extern_const_body: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Body,
    pub visible_parent_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdMap<DefId>>,
    pub missing_extern_crate_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool,
    pub used_crate_source: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateSource>,
    pub postorder_cnums: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<CrateNum>>,
    pub freevars: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Rc<Vec<Freevar>>>,
    pub maybe_unused_trait_import: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool,
    pub maybe_unused_extern_crates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(DefId, Span)>>,
    pub stability_index: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Index<'tcx>>,
    pub all_crate_nums: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<CrateNum>>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
type_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

generics_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Generics

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

super_predicates_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type_param_predicates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TraitDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx AdtDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_destructor: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Destructor>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_sized_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx [Ty<'tcx>]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

adt_dtorck_constraint: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> DtorckConstraint<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_const_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_default_impl: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_variances: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateVariancesMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

variances_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Variance>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

associated_item_def_ids: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

associated_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> AssociatedItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_trait_ref: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<TraitRef<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_polarity: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ImplPolarity

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_keys: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdSet>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_const_qualif: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> (u8, Rc<IdxSetBuf<Local>>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_const: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_validated: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Steal<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

optimized_mir: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Mir<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

closure_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ClosureKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> PolyFnSig<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

generator_sig: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<PolyGenSig<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

coerce_unsized_info: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> CoerceUnsizedInfo

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

typeck_item_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CompileResult

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

typeck_tables_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx TypeckTables<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

has_typeck_tables: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

coherent_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId))

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_borrowck: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_inherent_impls: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> CrateInherentImpls

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_inherent_impls_overlap_check: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

const_eval: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, (DefId, &'tcx Substs<'tcx>)>) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

privacy_access_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<AccessLevels>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

reachable_set: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<NodeSet>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

region_scope_tree: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<ScopeTree>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

mir_shims: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: InstanceDef<'tcx>) -> &'tcx Mir<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> SymbolName

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

symbol_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: Instance<'tcx>) -> SymbolName

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

describe_def: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Def>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_span: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

lookup_stability: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<&'tcx Stability>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

lookup_deprecation_entry: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DeprecationEntry>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

item_attrs: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<[Attribute]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn_arg_names: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Vec<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_parent: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_of_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_exported_symbol: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

item_body_nested_bodies: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<BTreeMap<BodyId, Body>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

const_is_rvalue_promotable_to_static: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_mir_available: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_impls_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<TraitImpls>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

specialization_graph_of: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Graph>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_object_safe: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

param_env: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> ParamEnv<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_copy_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_sized_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_freeze_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

needs_drop_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

layout_raw: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: ParamEnvAnd<'tcx, Ty<'tcx>>) -> Result<&'tcx Layout, LayoutError<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dylib_dependency_formats: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(CrateNum, LinkagePreference)>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_panic_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_compiler_builtins: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

has_global_allocator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_sanitizer_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_profiler_runtime: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

panic_strategy: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> PanicStrategy

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_no_builtins: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

extern_crate: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Option<ExternCrate>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

specializes: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (DefId, DefId)) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

in_scope_traits_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Rc<Vec<TraitCandidate>>>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

module_exports: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Rc<Vec<Export>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

lint_levels: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LintLevelMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl_defaultness: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Defaultness

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

exported_symbols: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

native_libraries: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<NativeLibrary>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

plugin_registrar_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

derive_registrar_fn: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_disambiguator: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_hash: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Svh

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

original_crate_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

implementations_of_trait: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: (CrateNum, DefId)) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

all_trait_implementations: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_dllimport_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_statically_included_foreign_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

native_library_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<NativeLibraryKind>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

link_args: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<String>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

named_region_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Region>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

is_late_bound_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashSet<ItemLocalId>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

object_lifetime_defaults_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefIndex) -> Option<Rc<FxHashMap<ItemLocalId, Rc<Vec<ObjectLifetimeDefault>>>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

visibility: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Visibility

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dep_kind: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> DepKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

crate_name: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

item_children: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Rc<Vec<Export>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

extern_mod_stmt_cnum: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<CrateNum>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

get_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<LanguageItems>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

defined_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(DefIndex, usize)>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

missing_lang_items: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<LangItem>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

extern_const_body: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> &'tcx Body

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

visible_parent_map: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<DefIdMap<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

missing_extern_crate_item: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

used_crate_source: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<CrateSource>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

postorder_cnums: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<CrateNum>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

freevars: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> Option<Rc<Vec<Freevar>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

maybe_unused_trait_import: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

maybe_unused_extern_crates: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<(DefId, Span)>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

stability_index: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Index<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

all_crate_nums: fn<'a>(_: TyCtxt<'a, 'tcx, 'tcx>, _: CrateNum) -> Rc<Vec<CrateNum>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?