Struct rustc::ty::AdtDef

pub struct AdtDef {
    pub did: DefId,
    pub variants: Vec<VariantDef>,
    pub repr: ReprOptions,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

The definition of an abstract data type - a struct or enum.

These are all interned (by intern_adt_def) into the adt_defs table.

Fields
did: DefId

🔬 This is a nightly-only experimental API. (rustc_private)

variants: Vec<VariantDef>

🔬 This is a nightly-only experimental API. (rustc_private)

repr: ReprOptions

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'a, 'gcx, 'tcx> AdtDef

fn uninhabited_from(&self,
visited: &mut FxHashMap<DefId, FxHashSet<&'tcx Substs<'tcx>>>,
tcx: TyCtxt<'a, 'gcx, 'tcx>,
substs: &'tcx Substs<'tcx>)
-> DefIdForest

🔬 This is a nightly-only experimental API. (rustc_private)

Calculate the forest of DefIds from which this adt is visibly uninhabited.

impl<'a, 'gcx, 'tcx> AdtDef

fn is_struct(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_union(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_enum(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn adt_kind(&self) -> AdtKind

🔬 This is a nightly-only experimental API. (rustc_private)

Returns the kind of the ADT - Struct or Enum.

fn descr(&self) -> &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

fn variant_descr(&self) -> &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_dtorck(&'gcx self, tcx: TyCtxt) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns whether this is a dtorck type. If this returns true, this type being safe for destruction requires it to be alive; Otherwise, only the contents are required to be.

fn is_fundamental(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns whether this type is #[fundamental] for the purposes of coherence checking.

fn is_phantom_data(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if this is PhantomData.

fn is_box(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns true if this is Box.

fn has_dtor(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Returns whether this type has a destructor.

fn struct_variant(&self) -> &VariantDef

🔬 This is a nightly-only experimental API. (rustc_private)

Asserts this is a struct and returns the struct's unique variant.

fn predicates(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> GenericPredicates<'gcx>

🔬 This is a nightly-only experimental API. (rustc_private)

fn all_fields<'s>(&'s self) -> impl Iterator<Item=&'s FieldDef>

🔬 This is a nightly-only experimental API. (rustc_private)

Returns an iterator over all fields contained by this ADT.

fn is_univariant(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_payloadfree(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn variant_with_id(&self, vid: DefId) -> &VariantDef

🔬 This is a nightly-only experimental API. (rustc_private)

fn variant_index_with_id(&self, vid: DefId) -> usize

🔬 This is a nightly-only experimental API. (rustc_private)

fn variant_of_def(&self, def: Def) -> &VariantDef

🔬 This is a nightly-only experimental API. (rustc_private)

fn discriminants(&'a self,
tcx: TyCtxt<'a, 'gcx, 'tcx>)
-> impl Iterator<Item=ConstInt> + 'a

🔬 This is a nightly-only experimental API. (rustc_private)

fn destructor(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Option<Destructor>

🔬 This is a nightly-only experimental API. (rustc_private)

fn sized_constraint(&self, tcx: TyCtxt<'a, 'gcx, 'tcx>) -> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

Returns a simpler type such that Self: Sized if and only if that type is Sized, or TyErr if this type is recursive.

HACK: instead of returning a list of types, this function can return a tuple. In that case, the result is Sized only if all elements of the tuple are Sized.

This is generally the struct_tail if this is a struct, or a tuple of them if this is an enum.

Oddly enough, checking that the sized-constraint is Sized is actually more expressive than checking all members: the Sized trait is inductive, so an associated type that references Self would prevent its containing ADT from being Sized.

Due to normalization being eager, this applies even if the associated type is behind a pointer, e.g. issue #31299.