Struct rustc::ty::ParameterEnvironment

pub struct ParameterEnvironment<'a, 'tcx> {
    pub tcx: &'a TyCtxt<'tcx>,
    pub free_substs: Substs<'tcx>,
    pub implicit_region_bound: Region,
    pub caller_bounds: Vec<Predicate<'tcx>>,
    pub selection_cache: SelectionCache<'tcx>,
    pub evaluation_cache: EvaluationCache<'tcx>,
    pub free_id_outlive: CodeExtent,
}

Unstable (rustc_private)

When type checking, we use the ParameterEnvironment to track details about the type/lifetime parameters that are in scope. It primarily stores the bounds information.

Note: This information might seem to be redundant with the data in tcx.ty_param_defs, but it is not. That table contains the parameter definitions from an "outside" perspective, but this struct will contain the bounds for a parameter as seen from inside the function body. Currently the only real distinction is that bound lifetime parameters are replaced with free ones, but in the future I hope to refine the representation of types so as to make more distinctions clearer.

Fields

tcx
Unstable (rustc_private)

free_substs
Unstable (rustc_private)

See construct_free_substs for details.

implicit_region_bound
Unstable (rustc_private)

Each type parameter has an implicit region bound that indicates it must outlive at least the function body (the user may specify stronger requirements). This field indicates the region of the callee.

caller_bounds
Unstable (rustc_private)

Obligations that the caller must satisfy. This is basically the set of bounds on the in-scope type parameters, translated into Obligations, and elaborated and normalized.

selection_cache
Unstable (rustc_private)

Caches the results of trait selection. This cache is used for things that have to do with the parameters in scope.

evaluation_cache
Unstable (rustc_private)

Caches the results of trait evaluation.

free_id_outlive
Unstable (rustc_private)

Scope that is attached to free regions for this scope. This is usually the id of the fn body, but for more abstract scopes like structs we often use the node-id of the struct.

FIXME(#3696). It would be nice to refactor so that free regions don't have this implicit scope and instead introduce relationships in the environment.

Methods

impl<'a, 'tcx> ParameterEnvironment<'a, 'tcx>

fn can_type_implement_copy(&self, self_type: Ty<'tcx>, span: Span) -> Result<(), CopyImplementationError>

Unstable (rustc_private)

impl<'a, 'tcx> ParameterEnvironment<'a, 'tcx>

fn with_caller_bounds(&self, caller_bounds: Vec<Predicate<'tcx>>) -> ParameterEnvironment<'a, 'tcx>

Unstable (rustc_private)

fn for_item(cx: &'a TyCtxt<'tcx>, id: NodeId) -> ParameterEnvironment<'a, 'tcx>

Unstable (rustc_private)

Construct a parameter environment given an item, impl item, or trait item