Struct rustc::ty::Tables

pub struct Tables<'tcx> {
    pub node_types: NodeMap<Ty<'tcx>>,
    pub item_substs: NodeMap<ItemSubsts<'tcx>>,
    pub adjustments: NodeMap<AutoAdjustment<'tcx>>,
    pub method_map: MethodMap<'tcx>,
    pub upvar_capture_map: UpvarCaptureMap<'tcx>,
    pub closure_tys: DefIdMap<ClosureTy<'tcx>>,
    pub closure_kinds: DefIdMap<ClosureKind>,
    pub liberated_fn_sigs: NodeMap<FnSig<'tcx>>,
    pub fru_field_types: NodeMap<Vec<Ty<'tcx>>>,
}

Unstable (rustc_private)

Fields
node_types: NodeMap<Ty<'tcx>>
Unstable (rustc_private)

Stores the types for various nodes in the AST. Note that this table is not guaranteed to be populated until after typeck. See typeck::check::fn_ctxt for details.

item_substs: NodeMap<ItemSubsts<'tcx>>
Unstable (rustc_private)

Stores the type parameters which were substituted to obtain the type of this node. This only applies to nodes that refer to entities parameterized by type parameters, such as generic fns, types, or other items.

adjustments: NodeMap<AutoAdjustment<'tcx>>
Unstable (rustc_private)
method_map: MethodMap<'tcx>
Unstable (rustc_private)
upvar_capture_map: UpvarCaptureMap<'tcx>
Unstable (rustc_private)

Borrows

closure_tys: DefIdMap<ClosureTy<'tcx>>
Unstable (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

closure_kinds: DefIdMap<ClosureKind>
Unstable (rustc_private)

Records the type of each closure. The def ID is the ID of the expression defining the closure.

liberated_fn_sigs: NodeMap<FnSig<'tcx>>
Unstable (rustc_private)

For each fn, records the "liberated" types of its arguments and return type. Liberated means that all bound regions (including late-bound regions) are replaced with free equivalents. This table is not used in trans (since regions are erased there) and hence is not serialized to metadata.

fru_field_types: NodeMap<Vec<Ty<'tcx>>>
Unstable (rustc_private)

For each FRU expression, record the normalized types of the fields of the struct - this is needed because it is non-trivial to normalize while preserving regions. This table is used only in MIR construction and hence is not serialized to metadata.

Methods

impl<'a, 'gcx, 'tcx> Tables<'tcx>

fn empty() -> Tables<'tcx>

Unstable (rustc_private)